Move test generation to test-utils
