+++ documentation/ArtisticDesign.json
@@ -3,116 +3,116 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "y90feEMUuKCddZb",
-    "ownerAffiliation" : "https://www.example.org/nonea"
+    "owner" : "1IMjjQMB1fBlk",
+    "ownerAffiliation" : "https://www.example.org/utconsequatur"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/laboriosamet",
+    "id" : "https://www.example.org/eteum",
     "labels" : {
-      "es" : "Xd3CQCGclfOoFsX"
+      "se" : "KfXf6o2cEA45a92"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/minussunt",
-  "doi" : "https://doi.org/10.1234/provident",
-  "link" : "https://www.example.org/odionemo",
+  "handle" : "https://www.example.org/consecteturea",
+  "doi" : "https://doi.org/10.1234/quo",
+  "link" : "https://www.example.org/estex",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "Jwp1w4OuT7jJF95b21",
+    "mainTitle" : "UkhhIlHLmt2l13G9LbX",
     "alternativeTitles" : {
-      "es" : "SGvoXezYYEwdxk"
+      "ca" : "Mm4eyYmNaCKheX"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "g7jHoqiVE0X39HIjz",
-      "month" : "gonxcj8yd5qcNwT",
-      "day" : "hCXxir5sJUt8o4WMt"
+      "year" : "FcxulqIuz1eBbS",
+      "month" : "hMPnCKibbpF",
+      "day" : "ItzNBjmqTC"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/kFoK5hAxjoHN",
-        "name" : "Q4iV3xaOfXpw",
+        "id" : "https://www.example.com/HgLXiHE9W7WjwOMr",
+        "name" : "No2owGD4vK",
         "nameType" : "Organizational",
-        "orcId" : "ojGJYVmuGKleXOw4"
+        "orcId" : "otVwMhPES5LRk"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/nxTD7f8Ut4",
+        "id" : "https://www.example.com/U5vVceTZwy8",
         "labels" : {
-          "es" : "X9CxXVBCZ4qPefbB0"
+          "zh" : "nKgisPXftmQE"
         }
       } ],
-      "role" : "Advisor",
-      "sequence" : 6,
+      "role" : "InterviewSubject",
+      "sequence" : 7,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/VOXwQd8w6jscoFnH",
-        "name" : "iQTplcQOUitqq3yO",
-        "nameType" : "Personal",
-        "orcId" : "LXcVzC9GXCio"
+        "id" : "https://www.example.com/d8ThVgWjHLxtLheKM",
+        "name" : "3K10gMUqv2OWr",
+        "nameType" : "Organizational",
+        "orcId" : "XXYiPdbNHVpFPF0qv2"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/r87VyuNzsP6h",
+        "id" : "https://www.example.com/6LkoPaH0xiaNhiBUkJ4",
         "labels" : {
-          "da" : "FiytUCmLFz"
+          "af" : "t77SEcx6KIg1PibqL"
         }
       } ],
-      "role" : "TranslatorAdapter",
-      "sequence" : 2,
+      "role" : "Researcher",
+      "sequence" : 8,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "wwyruzYClggR9khT6C",
-    "tags" : [ "oSsn6pE4Uqeq4" ],
-    "description" : "SPMMPUarS5R5crGJls",
+    "npiSubjectHeading" : "LpbFZwILNEP",
+    "tags" : [ "GauAJrtYPxvGWm" ],
+    "description" : "LaBd5YrRmcU96zru",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Artistic"
       },
-      "doi" : "https://www.example.com/5y7uNKPkgJX7iD6ASUL",
+      "doi" : "https://www.example.com/PmQvPwb1Abb0YFx",
       "publicationInstance" : {
         "type" : "ArtisticDesign",
         "subtype" : {
-          "type" : "LightDesign"
+          "type" : "InteriorDesign"
         },
-        "description" : "Th8urbmUKi1tI2o",
+        "description" : "HCoZYtzhcCk6",
         "venues" : [ {
           "type" : "Venue",
           "place" : {
             "type" : "UnconfirmedPlace",
-            "label" : "aHXsCRL4YqSDnohU",
-            "country" : "RPS7JQcfsNlT"
+            "label" : "glFDVfQkKo5vZtq9F4",
+            "country" : "T1MA54J52gev"
           },
           "date" : {
             "type" : "Period",
             "from" : "2020-09-23T09:51:23.044996Z",
             "to" : "2020-09-23T09:51:23.044996Z"
           },
-          "sequence" : 735982515
+          "sequence" : 679308884
         }, {
           "type" : "Venue",
           "place" : {
             "type" : "UnconfirmedPlace",
-            "label" : "2MJ0Swt6ZT7DEKbo",
-            "country" : "alFvdVlpP1dEtg"
+            "label" : "D5IrlHnh2o",
+            "country" : "uYbdoiEBrPwKnkTG"
           },
           "date" : {
             "type" : "Period",
             "from" : "2020-09-23T09:51:23.044996Z",
             "to" : "2020-09-23T09:51:23.044996Z"
           },
-          "sequence" : 1337754737
+          "sequence" : 1191028800
         } ],
         "peerReviewed" : false,
         "pages" : {
@@ -120,24 +120,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/K55zm1e7cx",
-    "abstract" : "uvAHzb03DyR6MOWB"
+    "metadataSource" : "https://www.example.com/NhTJOwExMP9",
+    "abstract" : "Wgq0LRX8pr9hlFo"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/possimusab",
-    "name" : "nui41h6nkpeZTg",
+    "id" : "https://www.example.org/essevero",
+    "name" : "LQA7DCAbBbyBqkcW",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "TomsN6vLPe3loGjAQIP",
-      "id" : "6JEJTqRJEiLGlQh7"
+      "source" : "yhpru1ZTPFWOEQF",
+      "id" : "ShUs75ydYRvVk3UamE"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "NS8Rx64MSP3AP2"
+      "approvedBy" : "REK",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "r5XgaL5h0yRWqasPCU"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -145,30 +145,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/quivoluptatem" ],
+  "subjects" : [ "https://www.example.org/nihileos" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "d583e77e-221b-4201-a54a-557ea04d016c",
-    "name" : "hYI1gYi1q8Md",
-    "mimeType" : "wb48eQAnWz0l6IuNNR0",
-    "size" : 447940809,
+    "identifier" : "dc8a4ebe-226c-4bfe-9730-e5c431942e76",
+    "name" : "NiZzi4bd8o",
+    "mimeType" : "2CKqCi9hhyHEsH",
+    "size" : 2129448255,
     "license" : {
       "type" : "License",
-      "identifier" : "DQMs68NpzZR",
+      "identifier" : "PEaJmCXqEXpm9Ka",
       "labels" : {
-        "fi" : "JiCB64xW0TZWhIJq"
+        "fi" : "6mlkDlDPPhTT7j8Mk"
       },
-      "link" : "https://www.example.com/05HqTdYDIUhtvr"
+      "link" : "https://www.example.com/WCvQmmlLewCgaTp"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/4zfBJR9Aqb6gscKAp",
-    "name" : "f60SKqL1K4eEvXh",
-    "description" : "oB56qJPvktUiF"
+    "id" : "https://www.example.com/WNDIifzysVQzuMoKw0F",
+    "name" : "cst3YPVdw9V",
+    "description" : "ESby8gkeQAhAMDthKm3"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
