+++ documentation/BookAbstracts.json
@@ -3,127 +3,127 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT",
   "resourceOwner" : {
-    "owner" : "QOFPPoKE0FD",
-    "ownerAffiliation" : "https://www.example.org/debitisveritatis"
+    "owner" : "uIYV74T3vDVWCH2nNf3",
+    "ownerAffiliation" : "https://www.example.org/occaecativelit"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/doloremdolorem",
+    "id" : "https://www.example.org/consequatursunt",
     "labels" : {
-      "sv" : "uIcvFMGMmN54e"
+      "da" : "ZkjCIlioRF"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quisnulla",
-  "doi" : "https://doi.org/10.1234/dolor",
-  "link" : "https://www.example.org/voluptatemut",
+  "handle" : "https://www.example.org/veniamquia",
+  "doi" : "https://doi.org/10.1234/mollitia",
+  "link" : "https://www.example.org/veniamea",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "bFldI1c3H02lO",
+    "mainTitle" : "koHJdF8J3nbBs1QS5DK",
     "alternativeTitles" : {
-      "nn" : "eoyD0VjtzdUqH"
+      "pt" : "EnUYgGiuq13z"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "ILRTdefcm0fJ1",
-      "month" : "904cyOzS3hPk7pmB",
-      "day" : "psBD3oZFIHDcp1Lqw9"
+      "year" : "iumDssVV9lq30YzemB",
+      "month" : "2jhkqUgCjl6",
+      "day" : "xfBXLOfkMQgy"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/b665W9WUy7cXpyvd",
-        "name" : "tD87JvkkbhofK",
-        "nameType" : "Personal",
-        "orcId" : "J2ildDjR6CMaYtJAx"
+        "id" : "https://www.example.com/kSnWylp5qB2s",
+        "name" : "fmlT6z03dJR",
+        "nameType" : "Organizational",
+        "orcId" : "XyyBHGlDQeCZ9gmKH"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/4UZRojUAgUzY5xIDHi",
+        "id" : "https://www.example.com/lKzWiNVW5n",
         "labels" : {
-          "nb" : "FSSb2acnwYylaJ2y"
+          "cs" : "RCd02V0KIaBZESO"
         }
       } ],
-      "role" : "Distributor",
+      "role" : "Architect",
       "sequence" : 9,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/BwibLXpgBJ4ER",
-        "name" : "BfrkMYDtTGs8Qws",
-        "nameType" : "Organizational",
-        "orcId" : "qbsJZwygoa8osX6YBO"
+        "id" : "https://www.example.com/B5xu1t306Jzsutk9Oy",
+        "name" : "S4xFXwM69WKvU6m9LjC",
+        "nameType" : "Personal",
+        "orcId" : "VByi4jDkDdFV3HNje"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/TqlKiWlijM3qRnvU7V8",
+        "id" : "https://www.example.com/c8VpGJbbEe",
         "labels" : {
-          "es" : "Davyi3YirF8c47NNtiu"
+          "af" : "7YMkNnHTsrq9gixYp"
         }
       } ],
-      "role" : "ProgrammeLeader",
-      "sequence" : 8,
+      "role" : "Screenwriter",
+      "sequence" : 5,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "0calfuOlxash9LRG2Z",
-    "tags" : [ "XK72JGutamc7G2l6ZOs" ],
-    "description" : "HFmhB7VXUEs3yCGMYL",
+    "npiSubjectHeading" : "hmZnsTN9qsX",
+    "tags" : [ "cJr7HXj5V4S" ],
+    "description" : "LyRGsyxuHoUJERQ",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Book",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/hT3268lAr1xM8nrDhq"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/wjVoPoX4ZEsmEbQFZxY"
         },
-        "seriesNumber" : "C53KANsFw4gnU1VCF",
+        "seriesNumber" : "t2Lq6UbvfX",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/pnB3NhTwrHO"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/nll5vJAlYA5LawYQiYV"
         },
-        "isbnList" : [ "9790015695353" ]
+        "isbnList" : [ "9780886836917" ]
       },
-      "doi" : "https://www.example.com/aQEjOWOhvihlvJcpgL",
+      "doi" : "https://www.example.com/82tqmH5yOQJSnBw3",
       "publicationInstance" : {
         "type" : "BookAbstracts",
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "sdaXZWpIrWeyXBBq8SD",
-            "end" : "4nDxpAMQ8j2B0ZZ"
+            "begin" : "cy0CBCEH0wPMsrAvL9",
+            "end" : "BXpd3nGynp3FNjeoy"
           },
-          "pages" : "WvPgPYDfHg",
+          "pages" : "CZ5uRS8PzB",
           "illustrated" : false
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/TnamKtNlW3IH",
-    "abstract" : "uEtTDomAgxt4mhfhrX"
+    "metadataSource" : "https://www.example.com/5aIUFr4G7VuoVry8UeN",
+    "abstract" : "txBdpvQV5EGSZ6v3"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/quisblanditiis",
-    "name" : "5b3uH9yAUKcxfT1DiJj",
+    "id" : "https://www.example.org/doloremminus",
+    "name" : "8AA5SJAwR4hd",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "mVfB8tSoeCFH3m7Cx",
-      "id" : "Q6SoLkrqcoDlMMU"
+      "source" : "UaT2aB3jj4v",
+      "id" : "dO3fs18G8XUKoCO"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "XPgLJio284AeGKTjr6"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "OOGuy6v4AlqDx28"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -131,30 +131,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/voluptatemipsum" ],
+  "subjects" : [ "https://www.example.org/esseiure" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "89e1f38e-3e26-4049-ae78-9fdfef6f17cd",
-    "name" : "YsCGotJPPYg",
-    "mimeType" : "hRsbsMGWwQ4h",
-    "size" : 2102965018,
+    "identifier" : "d860e4f7-bde0-46ba-a966-fbc3c55e3e14",
+    "name" : "RTWZ0ymx2m",
+    "mimeType" : "fKQ0OoFu7E6mYNaT71",
+    "size" : 595872248,
     "license" : {
       "type" : "License",
-      "identifier" : "EgfDZoPIKnMrkPVvkQh",
+      "identifier" : "0D9JBRd5fFVmSwXzq",
       "labels" : {
-        "en" : "AdMTmPla9n5TY8"
+        "pt" : "bye5SqXC73ijw94w5a"
       },
-      "link" : "https://www.example.com/CG549mRTi1fStebi27"
+      "link" : "https://www.example.com/erPmBtEibQ"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/CiTRcdIAvgCgQy",
-    "name" : "0RxTRZQmIMVHAhlu10",
-    "description" : "iM9hfz7Hf5lhzj"
+    "id" : "https://www.example.com/d1oTaU9XhKIXN4o",
+    "name" : "BWp1Y4pd1k",
+    "description" : "ZbqrLdV1hl4Ua"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
