+++ documentation/BookAnthology.json
@@ -3,127 +3,127 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT",
   "resourceOwner" : {
-    "owner" : "E68Be6iM0Xtc9Pr",
-    "ownerAffiliation" : "https://www.example.org/voluptatestotam"
+    "owner" : "L0WN66xUVth",
+    "ownerAffiliation" : "https://www.example.org/accusamusvoluptas"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/autveritatis",
+    "id" : "https://www.example.org/placeatet",
     "labels" : {
-      "fr" : "tbu9mX54XCF3"
+      "bg" : "DW8MI6uKe7Gym"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/repellendusvoluptatum",
-  "doi" : "https://doi.org/10.1234/ea",
-  "link" : "https://www.example.org/iustofugiat",
+  "handle" : "https://www.example.org/nihilnatus",
+  "doi" : "https://doi.org/10.1234/dolor",
+  "link" : "https://www.example.org/eumdoloremque",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "vPviHYkuKjmt",
+    "mainTitle" : "ZwIDHGkkMiVZIIDV",
     "alternativeTitles" : {
-      "is" : "8lTbMXX6pNQh3PzBhPC"
+      "fi" : "pbNvm9e42qaHul4"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "Y3neaiTQebbKdEpnv",
-      "month" : "AhXs9DccIZSqnQM",
-      "day" : "KCDiyh7LG3vV5"
+      "year" : "kTNs7WL3Ba5az33Q",
+      "month" : "frvxJqUBrFkrEZNWyP",
+      "day" : "7SF5E5Dp08imnBv"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Wh7pM2fXGT",
-        "name" : "7utSLT4fOzhZRBh5D",
-        "nameType" : "Organizational",
-        "orcId" : "5msUQPysNZa2B"
+        "id" : "https://www.example.com/lV4NTCq9uPWV16cV",
+        "name" : "m2gmYxSURkrVX",
+        "nameType" : "Personal",
+        "orcId" : "DOAMHrZwVxA"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/eXN8jVsaf8Rb",
+        "id" : "https://www.example.com/Q4NmGp0nlHI8TQp",
         "labels" : {
-          "hu" : "SqmzHZ9FbM"
+          "is" : "chpizHwbz3A9B5Cz"
         }
       } ],
-      "role" : "Journalist",
+      "role" : "Consultant",
       "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Ss4ZayzkC1S5",
-        "name" : "hlJ84Ije72k",
+        "id" : "https://www.example.com/4hjF12g4mmaCjIDWUBT",
+        "name" : "3LCDuFFrwWvZtTX",
         "nameType" : "Personal",
-        "orcId" : "YzwGB9HThyGzW"
+        "orcId" : "3dscVo15n87g7rPQ8h0"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/zq5rmrcv8rN",
+        "id" : "https://www.example.com/JoUY1RYIcmxjCMRJ",
         "labels" : {
-          "fi" : "01e9yOF1gtW"
+          "nl" : "jxB1MtCsnMUkphA"
         }
       } ],
-      "role" : "VideoEditor",
-      "sequence" : 0,
+      "role" : "InterviewSubject",
+      "sequence" : 5,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "vHGf4yIxmIOB",
-    "tags" : [ "MAtYS3PMR5gpkgHq" ],
-    "description" : "ltgpxc6JEvxivABQ3",
+    "npiSubjectHeading" : "G3NUM6uLogYVlcOtQDy",
+    "tags" : [ "9YT4CsFX6n" ],
+    "description" : "7WHVXON6oC2n4qi",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Book",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/stCZq6yOAShd"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/soS91hG6Y3"
         },
-        "seriesNumber" : "tmpW2vJjS4I",
+        "seriesNumber" : "ltrBkKOXiMr",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/FMMoRU4lfVpo03"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/TbaLS6dLlpq"
         },
-        "isbnList" : [ "9790976973538" ]
+        "isbnList" : [ "9780049499133" ]
       },
-      "doi" : "https://www.example.com/cr9JTx2vXdS",
+      "doi" : "https://www.example.com/3VLKBcvL88JX2",
       "publicationInstance" : {
         "type" : "BookAnthology",
-        "peerReviewed" : false,
+        "peerReviewed" : true,
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "YlLqgi3zUxyoie6gK7",
-            "end" : "LNyxKHxCwNEv"
+            "begin" : "K4j7lCwF62avan",
+            "end" : "75lVghC820"
           },
-          "pages" : "EgXfeyhQpKcMgWY",
-          "illustrated" : false
+          "pages" : "baY4sun8m6OTaSz0FPs",
+          "illustrated" : true
         }
       }
     },
-    "metadataSource" : "https://www.example.com/vseSOjqVdhyYMr",
-    "abstract" : "STTf3Uwhcol2"
+    "metadataSource" : "https://www.example.com/MHkSjzBiqeVQ0Gb",
+    "abstract" : "7htlRFJkpUxz"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/doloremminima",
-    "name" : "EXiAfYroMH8mZIR1g5",
+    "id" : "https://www.example.org/laborumadipisci",
+    "name" : "aIpy12hDh9630w",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "OijlcFz1I5Mlv",
-      "id" : "wElFHXDQ00ivwx"
+      "source" : "EbuDxGyzEv45s3",
+      "id" : "z63UuWYHbyBLHdTm1nE"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
-      "approvalStatus" : "DECLINED",
-      "applicationCode" : "Eo2f4RkQAdBtFO3"
+      "approvedBy" : "DIRHEALTH",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "UaJZCYItWgu4"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -131,30 +131,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/numquamdolorem" ],
+  "subjects" : [ "https://www.example.org/sintiusto" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "1e0d6d39-1341-491d-9581-11018034ebb9",
-    "name" : "kZHDShK1pFH0IVR6fH",
-    "mimeType" : "jw9nfOxGGIv4fUumt",
-    "size" : 486387060,
+    "identifier" : "92d10074-2077-4d67-902b-2bf4c58993f8",
+    "name" : "JsWZEWamRH",
+    "mimeType" : "Z2gFrFPbviLlFfq",
+    "size" : 8260167,
     "license" : {
       "type" : "License",
-      "identifier" : "VkytoNfaOrAJDcLc4LJ",
+      "identifier" : "qQL6uPEe8TtYDVZJtrM",
       "labels" : {
-        "fr" : "iGZZEnXTzzoPMnQNv"
+        "el" : "i7ko6MHMQYhDuLU18R"
       },
-      "link" : "https://www.example.com/bEtTCDuy54glvIjwjCW"
+      "link" : "https://www.example.com/V8o68kl6GIXu"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/hbFAST639GJm",
-    "name" : "pX5EG3pvTF",
-    "description" : "0humgE5L5IRE"
+    "id" : "https://www.example.com/5f1XTg14JqV3BzBaStD",
+    "name" : "mnAEHJxMKP",
+    "description" : "C3Ln8Q2oXqdN"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
