+++ documentation/BookMonograph.json
@@ -1,131 +1,131 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "qDNCy9BsxA",
-    "ownerAffiliation" : "https://www.example.org/fugiatasperiores"
+    "owner" : "1mdNnqLQ5h34kE",
+    "ownerAffiliation" : "https://www.example.org/velassumenda"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/rationeexpedita",
+    "id" : "https://www.example.org/nostrumnihil",
     "labels" : {
-      "hu" : "yoePYtbC9YMXvvI"
+      "pl" : "adE5kIYITqZjda"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/doloribusculpa",
-  "doi" : "https://doi.org/10.1234/voluptate",
-  "link" : "https://www.example.org/sapientesint",
+  "handle" : "https://www.example.org/inciduntquidem",
+  "doi" : "https://doi.org/10.1234/neque",
+  "link" : "https://www.example.org/utest",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "Z7UkyXtTQ2lIe",
+    "mainTitle" : "wQXHFKS7BJ",
     "alternativeTitles" : {
-      "ca" : "bEJYzdkezRFpc"
+      "hu" : "uMsWtGSGCIWxC8Lt"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "xLMg2QdgMpsah",
-      "month" : "bxRsgnv5L4LQW2",
-      "day" : "gi76YBLaLqhIi9"
+      "year" : "mCbbUCoPVHHT",
+      "month" : "LBISjDyg3LP",
+      "day" : "OYkNYFP07nO"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/BpAR7oTCcU5OGXGe",
-        "name" : "8tMM4QzyJy6",
-        "nameType" : "Personal",
-        "orcId" : "tBWngQ9OwiR5379OquQ"
+        "id" : "https://www.example.com/zXkq5kGONvr",
+        "name" : "baEMlyYDeWe9fhv99n",
+        "nameType" : "Organizational",
+        "orcId" : "L5g0xTyz5hiTqWp"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/Lp5VRyIaLfhFtYEqQ",
+        "id" : "https://www.example.com/Qa3g7r0ZZHSEv7",
         "labels" : {
-          "ru" : "VaGOu8BjxRrSaCW6oy"
+          "nl" : "tTVvQ2Mzmmd"
         }
       } ],
-      "role" : "RegistrationAgency",
-      "sequence" : 9,
+      "role" : "Scenographer",
+      "sequence" : 4,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/CkP3ycMhyNTKbwYXeN",
-        "name" : "mPukf8I0JQRP1Q",
-        "nameType" : "Organizational",
-        "orcId" : "StBmDK29zmPLYu"
+        "id" : "https://www.example.com/LjSL78tvzecg",
+        "name" : "T5T8mRjyBV2kfoVpb",
+        "nameType" : "Personal",
+        "orcId" : "jr6bI2PKsga6ZP"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/x9Q0BU4kyhiLQP0k",
+        "id" : "https://www.example.com/WgExkVBv0nod5crWsX",
         "labels" : {
-          "bg" : "YpOkwEYgfXpEk1"
+          "el" : "Sntjvy34AG9jMVcDMCq"
         }
       } ],
-      "role" : "Funder",
-      "sequence" : 1,
+      "role" : "ProjectLeader",
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "YPtVCLAayM",
-    "tags" : [ "UJZV5HNoZBzU6nNkSZ4" ],
-    "description" : "NUIoYJHCet2d",
+    "npiSubjectHeading" : "N71tccWaNHT",
+    "tags" : [ "NH5kh5gRdEU" ],
+    "description" : "BSKb5oRztkmFX0DNYF",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Book",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/tzT1GT1rQoMYhP"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/YFIT9nn8upjTxnxD"
         },
-        "seriesNumber" : "Fv7Gjxshgt",
+        "seriesNumber" : "Lh626TkdsP2oEN",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/o4vNv2KzNFjzUAZPnXM"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/0x6Cvm1xOTdqT"
         },
-        "isbnList" : [ "9791815310088" ]
+        "isbnList" : [ "9781956986532" ]
       },
-      "doi" : "https://www.example.com/22UweMdkQnGIWrPN",
+      "doi" : "https://www.example.com/Pef39Hp2huZM",
       "publicationInstance" : {
         "type" : "BookMonograph",
         "contentType" : "NonFictionMonograph",
         "originalResearch" : false,
-        "peerReviewed" : true,
+        "peerReviewed" : false,
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "2BEELCMYwhLHAM",
-            "end" : "bYkXWB9MUZNEFHQ"
+            "begin" : "MYs1mjfkSse9EZbNU",
+            "end" : "yhAQHWX4dVrERvAYLO"
           },
-          "pages" : "ou7hIZsIbO1dZ0rjG",
+          "pages" : "k5rst3i8AZ7Tpc14nwd",
           "illustrated" : true
         }
       }
     },
-    "metadataSource" : "https://www.example.com/kLtXSnTlxZJYxXtEt",
-    "abstract" : "OioemCIyMrBkKzg"
+    "metadataSource" : "https://www.example.com/tVsiP69g3YLy",
+    "abstract" : "AHqsUyj45zQvzuX0"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/auteveniet",
-    "name" : "edQijVhFNRtdczu",
+    "id" : "https://www.example.org/possimuseligendi",
+    "name" : "CrzNbKGhioZC",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "Pn3QXOKjqobS6SkyafJ",
-      "id" : "el24yxazvfwSN"
+      "source" : "6OtMI9t7oo5N",
+      "id" : "sctH7MJTHvYmMhovz5"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "itu9KFRhhmMQs"
+      "approvedBy" : "NMA",
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "dyS8w2z5AhDbb6WRww"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -133,30 +133,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/eumpraesentium" ],
+  "subjects" : [ "https://www.example.org/rerumut" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "c1fcbd5f-34e6-4c85-ad41-eb9b7296db68",
-    "name" : "VAHVw0mGBlr",
-    "mimeType" : "MFPhD6NxY1d",
-    "size" : 1163238932,
+    "identifier" : "742d9c29-3dc5-4b2e-8cc2-e675525a566a",
+    "name" : "R8mMSpb2daCyHgg6VQ1",
+    "mimeType" : "raZ7OvllqDUG",
+    "size" : 710466842,
     "license" : {
       "type" : "License",
-      "identifier" : "9qyIrTOXdRf",
+      "identifier" : "s2uMgAllsmku8mIkYD",
       "labels" : {
-        "nn" : "PjEbHljW3gZhjHGfi"
+        "sv" : "YWY2yMRgDzUV2"
       },
-      "link" : "https://www.example.com/CuaxDoXWAwWM"
+      "link" : "https://www.example.com/RDEoD8j8HXl8J"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/idI9cDSfxbknTWoRJY",
-    "name" : "6tgExtly6TO",
-    "description" : "7Gg6OjA6hBuZgoxY"
+    "id" : "https://www.example.com/qABQxzPKij",
+    "name" : "tPfsY6KL68sPUlOm38",
+    "description" : "hyHSJbct56VuJ"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
