+++ documentation/ChapterArticle.json
@@ -3,115 +3,115 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "0F72Al5WAd",
-    "ownerAffiliation" : "https://www.example.org/commodiarchitecto"
+    "owner" : "BfDbO8PqBi7qTw",
+    "ownerAffiliation" : "https://www.example.org/sitquidem"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/quasquisquam",
+    "id" : "https://www.example.org/eaatque",
     "labels" : {
-      "da" : "Fj2RcOPteuzw2MgD"
+      "en" : "oWg40yGamjeL"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/uteos",
-  "doi" : "https://doi.org/10.1234/omnis",
-  "link" : "https://www.example.org/sitillum",
+  "handle" : "https://www.example.org/cumsaepe",
+  "doi" : "https://doi.org/10.1234/deserunt",
+  "link" : "https://www.example.org/minusomnis",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "XH4tRJr1jjR2j",
+    "mainTitle" : "LZf5rlIlT0wEkiZL",
     "alternativeTitles" : {
-      "it" : "5W2qYKdfPN8kVHaZ1n"
+      "hu" : "1aWUDWut3lxF5Ag"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "AO2rifpJXi",
-      "month" : "NlPpu0tfAEoxQ",
-      "day" : "Z4wdnoQeJJ"
+      "year" : "AzOsSkfVdbNDioB6v",
+      "month" : "Lk9HhbKXDkx5oG",
+      "day" : "zOGajaNGHi"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/PDWWVnIcdCoW",
-        "name" : "2sZjU3yW8dv",
+        "id" : "https://www.example.com/8ZYJDNvQPBJds",
+        "name" : "GryYQkiBjZe1LcQob",
         "nameType" : "Personal",
-        "orcId" : "Lp0htIiNWEiRtc8VAq"
+        "orcId" : "4f6QXNRUgyxbp4ZH"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/lCk2zGlEnDZPAag",
+        "id" : "https://www.example.com/ox8Yg6Mty5AYyDFUNYh",
         "labels" : {
-          "cs" : "TYo434EVjfNrlY"
+          "af" : "rTBWhOKfYhNpmOJ8y"
         }
       } ],
-      "role" : "Dramatist",
-      "sequence" : 6,
+      "role" : "SoundDesigner",
+      "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/5LqLyMdUtf",
-        "name" : "npYDXmJ3DAMk",
+        "id" : "https://www.example.com/op3eF8yFCEIT",
+        "name" : "KzdN3xwIigoGj9Si",
         "nameType" : "Organizational",
-        "orcId" : "3yjZ6DbbagQ6yFUp"
+        "orcId" : "LgKkXsJu2KY"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/mi10nd1CYm",
+        "id" : "https://www.example.com/k7UCy8XGbSlC",
         "labels" : {
-          "is" : "OyNZfWINMFlJFt479e"
+          "en" : "vk07VRnbctTVFbh"
         }
       } ],
-      "role" : "Illustrator",
-      "sequence" : 8,
+      "role" : "Screenwriter",
+      "sequence" : 1,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "sSrttfp2TasXOTeSUzC",
-    "tags" : [ "cE5OHEyBZU" ],
-    "description" : "wBi3unyHcliJCs",
+    "npiSubjectHeading" : "6A8gZmLSuTQj",
+    "tags" : [ "7BdAv4QuMoyxhTM0k9m" ],
+    "description" : "Ufd2lqfARa",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Chapter",
-        "partOf" : "https://www.example.com/6NwOFn5pEgVRPqY20"
+        "partOf" : "https://www.example.com/JXuxnEIekE7jU"
       },
-      "doi" : "https://www.example.com/tAOvRCO0hObsetv",
+      "doi" : "https://www.example.com/lhh8RakPQJ5hRhtT6",
       "publicationInstance" : {
         "type" : "ChapterArticle",
         "contentType" : "EncyclopediaChapter",
         "peerReviewed" : true,
-        "originalResearch" : true,
+        "originalResearch" : false,
         "pages" : {
           "type" : "Range",
-          "begin" : "IokoaZHSeQ",
-          "end" : "hdg8fcVyDHTe"
+          "begin" : "4B71sU6KET",
+          "end" : "T7Ofl4U1auscZZ8Wi"
         }
       }
     },
-    "metadataSource" : "https://www.example.com/0FtoG2sYQC",
-    "abstract" : "Zra2oGq1AM"
+    "metadataSource" : "https://www.example.com/vPHxzicIePUfuKq9S",
+    "abstract" : "hajBnkv4CYwP"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/omnisminima",
-    "name" : "G2znXczI21hjmsRo",
+    "id" : "https://www.example.org/pariaturunde",
+    "name" : "5n3tnI4I5E0A56yxe",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "Fjq2soXw91iiZ2fw",
-      "id" : "cjbYULLzs7f0Skh"
+      "source" : "QydZKOS0RR2cIO2",
+      "id" : "nfYEBtaTZM6NoHMtLPS"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
+      "approvedBy" : "REK",
       "approvalStatus" : "APPLIED",
-      "applicationCode" : "G82d2DIDlxE"
+      "applicationCode" : "e2dIduCk57aWKXl08M"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -119,30 +119,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/expeditaat" ],
+  "subjects" : [ "https://www.example.org/autat" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "13686ecd-e7b3-4157-9efd-60763bfc8866",
-    "name" : "V4AkNSac7LF",
-    "mimeType" : "Z4jgMlMnX5NG0rHHDB",
-    "size" : 1616788360,
+    "identifier" : "12265989-a905-4b6c-ba59-ef3f16bf2478",
+    "name" : "hx2afcwbgbI8KM",
+    "mimeType" : "TT2e1pphWsHNw7z6",
+    "size" : 597188567,
     "license" : {
       "type" : "License",
-      "identifier" : "Me6aqxUflx",
+      "identifier" : "TX41THft8jPb6",
       "labels" : {
-        "zh" : "mbMJWfGVj9U0Vt"
+        "af" : "dqgwYcr3tos7jgbit"
       },
-      "link" : "https://www.example.com/nnqxJC8k4d"
+      "link" : "https://www.example.com/aPMp54Xq40"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/Uy64CZd4PHX0",
-    "name" : "awC2KMlCA8dBiJv074u",
-    "description" : "QezbMxNdXQQfc"
+    "id" : "https://www.example.com/m4ULXceYOuwLTVEasS",
+    "name" : "y2aWAmPPwVraDhAQK",
+    "description" : "gOWVqgILcYanX"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
