+++ documentation/ChapterConferenceAbstract.json
@@ -3,113 +3,113 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT",
   "resourceOwner" : {
-    "owner" : "6canEnGtkzEFv",
-    "ownerAffiliation" : "https://www.example.org/necessitatibustotam"
+    "owner" : "SJXPkcbOJsty3fFQTxx",
+    "ownerAffiliation" : "https://www.example.org/voluptatemut"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/fugitalias",
+    "id" : "https://www.example.org/quoin",
     "labels" : {
-      "hu" : "SzGUfzCzGpX6g"
+      "sv" : "7h4QTJNTeJ"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/sinteum",
-  "doi" : "https://doi.org/10.1234/qui",
-  "link" : "https://www.example.org/estveritatis",
+  "handle" : "https://www.example.org/nequenihil",
+  "doi" : "https://doi.org/10.1234/est",
+  "link" : "https://www.example.org/omnismollitia",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "NJkem0PxUzwd82GnWn",
+    "mainTitle" : "hQHVpLRsUgk2gmY",
     "alternativeTitles" : {
-      "sv" : "8P9nQ0MPvIsW"
+      "af" : "9fSsM1qXmCwK"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "OhCDMERtQ5wiV",
-      "month" : "UdhmW8w7EGdzUxvupe",
-      "day" : "AqebyvFyCvwl2UrnBB"
+      "year" : "oDpEczU296",
+      "month" : "qlUulMtNCzVnh9RYk",
+      "day" : "oGkKh1EA8MfbuyEu3s"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/TmXWpucqEx",
-        "name" : "ukqgzdgyCIiQx0K",
-        "nameType" : "Organizational",
-        "orcId" : "IULlW1B38WuSD"
+        "id" : "https://www.example.com/fpA045YOaOCEfKu",
+        "name" : "OhAkeGlZZqPD",
+        "nameType" : "Personal",
+        "orcId" : "F5WAfyfwmhHQ1OI4j5"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/j4pNdnFUwKvTUTVpDD",
+        "id" : "https://www.example.com/zQVGV7L1pfm",
         "labels" : {
-          "fi" : "FXT7Yi5LV1oWwQKwDN3"
+          "el" : "OxKwyMglZdULSo"
         }
       } ],
-      "role" : "Writer",
+      "role" : "VideoEditor",
       "sequence" : 9,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/yKpZXacmUVO6M3JYM",
-        "name" : "f9yVhaTMiqKI0YT",
+        "id" : "https://www.example.com/BQYoYCpNmbQ",
+        "name" : "FuDQb2Qe0mM4lEOQo83",
         "nameType" : "Personal",
-        "orcId" : "gtR2gR5Z7njxhK"
+        "orcId" : "JEDgJhVum6zH"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/eWQE7u5K9GTk47sc8z",
+        "id" : "https://www.example.com/kCGZFrRtHKe8RNj",
         "labels" : {
-          "de" : "409vT95CfFOcn"
+          "it" : "lLwuCty8BJdnc"
         }
       } ],
-      "role" : "Scenographer",
+      "role" : "ProjectLeader",
       "sequence" : 4,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "pXTSYbH7HCp09",
-    "tags" : [ "0INDLXDPtu" ],
-    "description" : "BdcAlIbGLaT849EMcc0",
+    "npiSubjectHeading" : "GHYRFOwyUInOTo5rWoY",
+    "tags" : [ "uZZifwmr5nvKi50f7xa" ],
+    "description" : "cXuZNRnjg41EJOzL",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Chapter",
-        "partOf" : "https://www.example.com/B6gb2QEKXtNBjPkJ"
+        "partOf" : "https://www.example.com/M25PIAsWRcbrEA"
       },
-      "doi" : "https://www.example.com/3UzhPSa4I3be",
+      "doi" : "https://www.example.com/rfJkIfsNmRSXJoLov",
       "publicationInstance" : {
         "type" : "ChapterConferenceAbstract",
         "peerReviewed" : false,
         "pages" : {
           "type" : "Range",
-          "begin" : "7OWd9EY4FkM8",
-          "end" : "bivArkTbp2k1HQTE1jf"
+          "begin" : "NpYinbAB4ZVrjH",
+          "end" : "jhAtQ8xgBI4vrhOo"
         }
       }
     },
-    "metadataSource" : "https://www.example.com/XqMCU5Yo3AXU",
-    "abstract" : "oX32LjwlylhYdz5lB"
+    "metadataSource" : "https://www.example.com/a9eEbKejF9",
+    "abstract" : "vXuGTwhpGybuZodMibV"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/porrodoloremque",
-    "name" : "mDAw7JMSMv",
+    "id" : "https://www.example.org/temporesed",
+    "name" : "cQRA4NBPleAAL",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "OcVcrUa0laNPKUz",
-      "id" : "EftbTFOIu258z"
+      "source" : "eg5jc0qS5nD7ya0uP",
+      "id" : "kTuxHJ3T1TqCxIKbw"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "Z3W2JhxMk2xej8sXn"
+      "approvedBy" : "NMA",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "6b79nhGwGAbVGblQs"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -117,30 +117,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/quibusdamqui" ],
+  "subjects" : [ "https://www.example.org/voluptatumaspernatur" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "c0d197dd-ce0f-4721-b296-aa0b0ce9194c",
-    "name" : "8FOUBho133jKi",
-    "mimeType" : "cYrShCHXGUEeUC3TBgJ",
-    "size" : 458050093,
+    "identifier" : "d36d208f-53fe-4f17-bfc0-ef89b313f96f",
+    "name" : "VX4927usx3mEOA6zUTX",
+    "mimeType" : "KlHHxGTbSzWp",
+    "size" : 259620948,
     "license" : {
       "type" : "License",
-      "identifier" : "JEPziBc2R6A",
+      "identifier" : "yrpJB2r5f0sYpcifZ",
       "labels" : {
-        "sv" : "13rfj7c5IZ0"
+        "sv" : "zY7dRBV8xGsinN9pfGA"
       },
-      "link" : "https://www.example.com/F1gUHMBQGTXsG"
+      "link" : "https://www.example.com/JH7Qxe8tJuldBUhZkkW"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/90MWVmW3KH",
-    "name" : "GX09vPcPTkTOhc",
-    "description" : "NZrwS6wbfbLEJSCKx"
+    "id" : "https://www.example.com/JOp78SJhdejMoDwj",
+    "name" : "D0DxHuD8oZgRg8nr",
+    "description" : "HVBFJJ9VwMt"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
