+++ documentation/ChapterInReport.json
@@ -1,115 +1,115 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
+  "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "rmEUlL9NzBJJLeRN",
-    "ownerAffiliation" : "https://www.example.org/modiatque"
+    "owner" : "TNXrTy6MqL",
+    "ownerAffiliation" : "https://www.example.org/ametvoluptatibus"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/autid",
+    "id" : "https://www.example.org/consequunturquo",
     "labels" : {
-      "es" : "2WN7TudlOi3f0"
+      "en" : "MCEu0JLPyKP"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quidolores",
-  "doi" : "https://doi.org/10.1234/est",
-  "link" : "https://www.example.org/dolorescum",
+  "handle" : "https://www.example.org/autdebitis",
+  "doi" : "https://doi.org/10.1234/dolor",
+  "link" : "https://www.example.org/eostempora",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "3UbpFywvFgulpt",
+    "mainTitle" : "7s2fn7aug1ppOOr",
     "alternativeTitles" : {
-      "sv" : "6p7x5uCZNxvVsSYvB"
+      "es" : "lnlRd9kZ7ILR"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "cnqbSyoexFe72mGGrlS",
-      "month" : "7EYrZrLTBHrdnt43f",
-      "day" : "LyvvMiOru1YtrrxguQp"
+      "year" : "8G7SYe446u1Zy",
+      "month" : "6sLwRdACdBP",
+      "day" : "iimdO95gIn4e"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/h8k9cEKvFlhpne",
-        "name" : "7Iwqh8mzqNX",
-        "nameType" : "Personal",
-        "orcId" : "nGr72qwmW0bSG8vh2"
+        "id" : "https://www.example.com/W3VQ0CrqS3K7",
+        "name" : "OWqwaFLhtOk",
+        "nameType" : "Organizational",
+        "orcId" : "BX7QeZeFZZO2UCQ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/0DtwZQaah2r2nOD3I",
+        "id" : "https://www.example.com/QhUFHoS852NaSlJKxHp",
         "labels" : {
-          "bg" : "sFnHFqVMGRMguNyoq"
+          "ca" : "NpHKFXvPe2tWdOqTTSS"
         }
       } ],
-      "role" : "Dancer",
-      "sequence" : 7,
+      "role" : "ProgrammeLeader",
+      "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/1eL4rUP5WG",
-        "name" : "YLbA0ijhIo6dV",
-        "nameType" : "Organizational",
-        "orcId" : "klyF0V5c4hP6z"
+        "id" : "https://www.example.com/kOqhM6JH8RKeOciyB8e",
+        "name" : "AuybH49GRx",
+        "nameType" : "Personal",
+        "orcId" : "jcuIJSrIp3"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/gV8rmi7i1FIHWCX2X2",
+        "id" : "https://www.example.com/8ibJ5As0ek0ucqx0x",
         "labels" : {
-          "es" : "soElzWUbsetqMbC"
+          "hu" : "DmEt25fkJ8CywZhoy"
         }
       } ],
-      "role" : "Journalist",
-      "sequence" : 6,
+      "role" : "DataCollector",
+      "sequence" : 2,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "YqvQ4YOkeRUI6zew",
-    "tags" : [ "8u64s9HrxU" ],
-    "description" : "umkzPK3ImaS",
+    "npiSubjectHeading" : "KGk8pnze9D5iQpkiX2K",
+    "tags" : [ "Y7HwExD3RlVYnIR" ],
+    "description" : "ILSmylrAX0T",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Chapter",
-        "partOf" : "https://www.example.com/SpzENkUfrx69CpxfkF"
+        "partOf" : "https://www.example.com/7IcWuqWJLiYfXstD3t"
       },
-      "doi" : "https://www.example.com/twlNm7Rv4g00kjgU4",
+      "doi" : "https://www.example.com/H6N0UUSC2xC",
       "publicationInstance" : {
         "type" : "ChapterInReport",
         "peerReviewed" : false,
         "pages" : {
           "type" : "Range",
-          "begin" : "7aHGJVnIgJim",
-          "end" : "0GOuHZuv17ttxQ"
+          "begin" : "8CruyFYiwmSY",
+          "end" : "pr41oxwfBCqcOR"
         }
       }
     },
-    "metadataSource" : "https://www.example.com/ZgaPJgBtD3iaAec6WKX",
-    "abstract" : "BSgzGOTMHwT1W9"
+    "metadataSource" : "https://www.example.com/ITcU9l2uhASidJ",
+    "abstract" : "SPblHcNLs5m3THl"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/blanditiiset",
-    "name" : "xl15IXad4I9LYBqxZ",
+    "id" : "https://www.example.org/idmolestiae",
+    "name" : "l3fHpA0fbLt3GfZnp",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "CeI00frLYPTmDA",
-      "id" : "venZ0gmbka"
+      "source" : "lNzeieoOxR30WCPqeL",
+      "id" : "6NPdg1Y5pW"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "NARA",
-      "approvalStatus" : "DECLINED",
-      "applicationCode" : "qdRPvfDZGW5T"
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "ZEnP5eLqMr8"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -117,30 +117,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/optiout" ],
+  "subjects" : [ "https://www.example.org/repellenduseaque" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "dbdc8cb9-8a8a-44e8-a07e-d43ab7e2846a",
-    "name" : "EbXMg9f5DNJ1TWWhWWD",
-    "mimeType" : "RtJJwyPUcfaK",
-    "size" : 1567904804,
+    "identifier" : "89c1849d-ade9-4506-bc63-674abe93260a",
+    "name" : "wgjlAoObfie1YeZE",
+    "mimeType" : "zYwA8BHQR9ACki",
+    "size" : 1538532054,
     "license" : {
       "type" : "License",
-      "identifier" : "AbysYQAWLa",
+      "identifier" : "nKQUMHnZdw7E4Qm73l",
       "labels" : {
-        "el" : "zx8ma46iZXX"
+        "en" : "4909LE1p5lkPT"
       },
-      "link" : "https://www.example.com/72tfANPDEpGfXdquy"
+      "link" : "https://www.example.com/7SywF3C6JBT"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/vtoBHCOR3xQjot",
-    "name" : "gngFNvpRw6R1T",
-    "description" : "O9SYp4DBd2eGTKBq"
+    "id" : "https://www.example.com/kjK31CqhXArV32ABuk",
+    "name" : "tY2ZQJkaBOiky5w",
+    "description" : "7kuDt9MnMfdDQ"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
