+++ documentation/ConferenceAbstract.json
@@ -1,118 +1,118 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "NEW",
+  "status" : "DRAFT",
   "resourceOwner" : {
-    "owner" : "smHKDMXbA2BO",
-    "ownerAffiliation" : "https://www.example.org/providentdoloribus"
+    "owner" : "ltTVEPINn0WTVAfz",
+    "ownerAffiliation" : "https://www.example.org/expeditavel"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/ducimusdolore",
+    "id" : "https://www.example.org/sedet",
     "labels" : {
-      "pl" : "7avLeLEuSMydGXzb"
+      "es" : "ZCXUzI6W2RkDcsAS2T"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/rerumquidem",
-  "doi" : "https://doi.org/10.1234/nemo",
-  "link" : "https://www.example.org/eossit",
+  "handle" : "https://www.example.org/voluptatemsunt",
+  "doi" : "https://doi.org/10.1234/quis",
+  "link" : "https://www.example.org/suntesse",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "sO8PsGNCnq5",
+    "mainTitle" : "cxHrkdtzRXyzi8",
     "alternativeTitles" : {
-      "bg" : "KH6XYwHL8dW7g4"
+      "zh" : "IJYVgsDHxD"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "hX21aahLsBR82Ypu",
-      "month" : "dwksRSICddJL5",
-      "day" : "OSETAzvPDUju6YfyRlM"
+      "year" : "wK2M86TMJNoIqvGEI0",
+      "month" : "KLJCFl0ieR",
+      "day" : "nSMkKoISjP0"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/KAPjVjj9sndJNTWcw",
-        "name" : "xFmd6m8otBYcONAgF",
-        "nameType" : "Personal",
-        "orcId" : "ajf3ZqyXEoA9h"
+        "id" : "https://www.example.com/xw9cIOiHlu8gD",
+        "name" : "BfB2fM0dsdf9Pn4jX",
+        "nameType" : "Organizational",
+        "orcId" : "F0G7yxxpQxMeh4l6Wz0"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/7CLA482Hq7RWiOGlsV",
+        "id" : "https://www.example.com/tBFrgnMj85v",
         "labels" : {
-          "fr" : "KX4k3xDHgl3SEl1F"
+          "it" : "ozNKzZI31CrDp"
         }
       } ],
-      "role" : "SoundDesigner",
-      "sequence" : 9,
+      "role" : "ArchitecturalPlanner",
+      "sequence" : 6,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/k462GY2xPWZx0",
-        "name" : "yPeyBq9std24",
+        "id" : "https://www.example.com/ljMsRdMxSaNd6yW",
+        "name" : "Vz9GXTKXxvt",
         "nameType" : "Personal",
-        "orcId" : "YjqDzA3iefeU4nh6"
+        "orcId" : "KRfctmC5MTfrwf53Es"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/R1NDVGJoQibWEW",
+        "id" : "https://www.example.com/vbU6ZWRBdiFjPW",
         "labels" : {
-          "it" : "rWQkRDbevbdAGDO"
+          "de" : "yyKczZrCU8Y5YcAwqP"
         }
       } ],
-      "role" : "LandscapeArchitect",
-      "sequence" : 3,
+      "role" : "Researcher",
+      "sequence" : 9,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "mkOAy5Gb97PL870C",
-    "tags" : [ "lzCLdSDNrK6YUyWaq" ],
-    "description" : "QyWMz34cUonS9p",
+    "npiSubjectHeading" : "24wamgfhVmV",
+    "tags" : [ "SpsvQF00px" ],
+    "description" : "6GQDl9c3YB",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/5mxAihHr80v"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/yowC3FmElNCl7R4ZfJ"
       },
-      "doi" : "https://www.example.com/xkK1LUf7VV8Ijg",
+      "doi" : "https://www.example.com/lBrpQTpZEYFA9m",
       "publicationInstance" : {
         "type" : "ConferenceAbstract",
-        "volume" : "G9enMu7JkEsfLT7Zyo",
-        "issue" : "nMBwSYhqeXIM7h",
-        "articleNumber" : "LpE4J2NUjrZ6jvhYz",
+        "volume" : "Lt2WQmgnrllM",
+        "issue" : "rIOIUX1B0KvqD",
+        "articleNumber" : "MqS7W92sFmLRe4N5w",
         "pages" : {
           "type" : "Range",
-          "begin" : "3S4nfcPYTNL4lnoN6uO",
-          "end" : "vtYXfVOQYs"
+          "begin" : "7tgypnk7Zn",
+          "end" : "L01RtGAehj"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/1AOqZiHHYli5XMGc",
-    "abstract" : "XVuyE3wO7oDjnHVty"
+    "metadataSource" : "https://www.example.com/A9HHgoxuGfb5DTt2Yu3",
+    "abstract" : "HYMNAhG1Ms"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/aliquamvoluptatem",
-    "name" : "tE3yZNiJ0xu6Gv05GCG",
+    "id" : "https://www.example.org/errortenetur",
+    "name" : "hVxreJc7t2Uoq2QzJx7",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "ZWG1v2ZYJf7OoDgv",
-      "id" : "BpbajcnhAo1aT"
+      "source" : "AsOV1fLLaZXZ",
+      "id" : "dscD2cUNWTBNsqJea9"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "os4vKm6O5WeEg4t"
+      "approvedBy" : "DIRHEALTH",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "mohClq9zTIB"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -120,30 +120,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/utex" ],
+  "subjects" : [ "https://www.example.org/omnisnihil" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "7fa5dd0f-4ae7-4528-abba-5d048bed0bcb",
-    "name" : "BBYnoqUuVQllP20Uuu",
-    "mimeType" : "8JtPPuiqh2wtBsrOaQp",
-    "size" : 1193001113,
+    "identifier" : "7072e26e-b74f-460b-9711-77d9f477fad2",
+    "name" : "IHL4P3s7TUGl9e9",
+    "mimeType" : "kIPgqwixbUi",
+    "size" : 630628522,
     "license" : {
       "type" : "License",
-      "identifier" : "eTLCi4B8csqTAQ",
+      "identifier" : "Y7BGvwfXVlNE8",
       "labels" : {
-        "da" : "cjfgGFO1Xc"
+        "zh" : "fw0uUjF6qa"
       },
-      "link" : "https://www.example.com/qwqnJr6Bqj7g4p"
+      "link" : "https://www.example.com/l3hfClOe52QCuG1sjy"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/vHJJh1Ts13gc0KNmt",
-    "name" : "yPJcXyt68E8jRqZR6QJ",
-    "description" : "Iua0XGXyGEUu3xjm"
+    "id" : "https://www.example.com/5GgfmWbnAR3AsQEd5mX",
+    "name" : "svHhewqWY99Rl2FGTVa",
+    "description" : "nALV9hLVpcK2JJ"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
