+++ documentation/ConferenceLecture.json
@@ -1,104 +1,105 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
+  "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "zPruSPyqD7uQoXdVD",
-    "ownerAffiliation" : "https://www.example.org/magnamsimilique"
+    "owner" : "P4B4FLoHlgo",
+    "ownerAffiliation" : "https://www.example.org/optiodolores"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/consequaturducimus",
+    "id" : "https://www.example.org/vitaereprehenderit",
     "labels" : {
-      "fr" : "ZaMZLGInGEgyac"
+      "se" : "Daf6bmUdV2qiW"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/excepturicorrupti",
-  "doi" : "https://doi.org/10.1234/repellendus",
-  "link" : "https://www.example.org/maioresquisquam",
+  "handle" : "https://www.example.org/dictaesse",
+  "doi" : "https://doi.org/10.1234/nam",
+  "link" : "https://www.example.org/quolabore",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "us6vsRShXyCVOClE9",
+    "mainTitle" : "jCl1KaPzUPSZomhc5l",
     "alternativeTitles" : {
-      "pt" : "XuJOYf8N5EKLp3KvE6q"
+      "is" : "nN3apP4OMBIc"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "nmrVKVSGTm2ZZP",
-      "month" : "38ihOWpLvEGYVprWlsc",
-      "day" : "uNMtjdNRKyu"
+      "year" : "szs9FQgO8NpH2TOG",
+      "month" : "J7f3BhNkuB",
+      "day" : "jjPnxbayACn"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/DsY1q6aJStgQeX",
-        "name" : "MgynEaocf2OfRlVvSq",
+        "id" : "https://www.example.com/gBgmrzsYuuYqLPUkbw",
+        "name" : "nCcLZIm9pyZi",
         "nameType" : "Organizational",
-        "orcId" : "BCQPsPrTgZW7yOcEB"
+        "orcId" : "Ij0Pr1PoPABS"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/A4K2fx2g5u",
+        "id" : "https://www.example.com/zWf4T1XV5CYxuGk",
         "labels" : {
-          "nl" : "d9nt0XQ1mZBM6lx94"
+          "hu" : "O2QiRVrBUgXh9kM"
         }
       } ],
-      "role" : "CostumeDesigner",
-      "sequence" : 8,
+      "role" : "SoundDesigner",
+      "sequence" : 5,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/ccGb63SYwAlm",
-        "name" : "34HE1hnM7YVUUMbr",
-        "nameType" : "Personal",
-        "orcId" : "IIxEilRYXKmmFe"
+        "id" : "https://www.example.com/TuVdxVuBrrk4Om",
+        "name" : "mNoUzAoYSmzvCwB",
+        "nameType" : "Organizational",
+        "orcId" : "dotEoSHpFp47b"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/Hm3Qi1aKObr0F",
+        "id" : "https://www.example.com/iSCGJIPtwIlMoXTM3wi",
         "labels" : {
-          "af" : "lPkMTy09YQISWFkqwi"
+          "pl" : "F4NiRD5nJe5"
         }
       } ],
-      "role" : "DataManager",
-      "sequence" : 2,
+      "role" : "Organizer",
+      "sequence" : 5,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "234EzS4QW1fKJePk7SX",
-    "tags" : [ "GYyw3Z32rEWstPOR" ],
-    "description" : "9yyxhRo7B24w1",
+    "npiSubjectHeading" : "8d18EXTH1sLp",
+    "tags" : [ "yxCY2dGKrZ6QBbiwIaf" ],
+    "description" : "7rUdfAZK4tVh4vCAh7",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Event",
-        "label" : "mL6fcBGvhKTuHvcY",
+        "label" : "duFq0RtAW9L",
         "place" : {
           "type" : "UnconfirmedPlace",
-          "label" : "LAChGDbbhGt2spj",
-          "country" : "S3XA0R6nftCqhYAi"
+          "label" : "wy2cwaheGRXczA3r",
+          "country" : "3Q1zsxvII0Ycew"
         },
         "time" : {
-          "type" : "Instant",
-          "value" : "2020-09-23T09:51:23.044996Z"
+          "type" : "Period",
+          "from" : "2020-09-23T09:51:23.044996Z",
+          "to" : "2020-09-23T09:51:23.044996Z"
         },
         "agent" : {
           "type" : "Organization",
-          "id" : "https://www.example.com/71zbRTAV9chz",
+          "id" : "https://www.example.com/RCZrkjsIwTpoF64",
           "labels" : {
-            "af" : "qRoJnCrjmG3bqrkxkX7"
+            "nn" : "rivvEf7qZUNRuD56X"
           }
         },
-        "product" : "https://www.example.com/duwOZF708gNSYQSgx4"
+        "product" : "https://www.example.com/0y5x8W7ShZIq5"
       },
-      "doi" : "https://www.example.com/6TyAbmmBIk9c6OG",
+      "doi" : "https://www.example.com/FRNGSQt9hY4FbtLaCU2",
       "publicationInstance" : {
         "type" : "ConferenceLecture",
         "peerReviewed" : false,
@@ -107,24 +108,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/MQ1dqc64Jl7fB",
-    "abstract" : "q0nqrzJfAec"
+    "metadataSource" : "https://www.example.com/Fv6iYj8Drlcf",
+    "abstract" : "S5ZQdNxdf1"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/quisoluta",
-    "name" : "EZxDEa2PLiLIlA7Yv0",
+    "id" : "https://www.example.org/sequisint",
+    "name" : "OXR79f5a6lunCm0m",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "zhRqILDBJLnht",
-      "id" : "RffqX5isRV6ay03CYw"
+      "source" : "LAQfcueJQfIR",
+      "id" : "3j4xnZTo2yoIP6qo9"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "010OSOUF5oe5"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "QFdgdm8elFqT"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -132,30 +133,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/etnihil" ],
+  "subjects" : [ "https://www.example.org/deleniticonsectetur" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "091a5028-d148-4ae3-8f77-cff6bc7b16ae",
-    "name" : "rKaFoV9aFb",
-    "mimeType" : "twB31bMDHm2oWn",
-    "size" : 361566960,
+    "identifier" : "1bd4e23f-29c6-492a-ac26-550861191786",
+    "name" : "TInXf2oA4dslZaa1DI",
+    "mimeType" : "vwg6qWDc2b",
+    "size" : 847551393,
     "license" : {
       "type" : "License",
-      "identifier" : "QNxhUKArwvEFa",
+      "identifier" : "UgEWDuvXQFP9qzpc",
       "labels" : {
-        "da" : "OkK76Yh2Skm5S"
+        "es" : "5DgHfwqllYMdniJg5m"
       },
-      "link" : "https://www.example.com/UNC0jxMRw4WPJO"
+      "link" : "https://www.example.com/mUbvWXWI6blyR8ivK"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/kH1Nr7p0yJdaZjlY0",
-    "name" : "wxV7FSfDyPlf",
-    "description" : "zkzDLSm8XlZ3"
+    "id" : "https://www.example.com/qX4wZTYHt4x566",
+    "name" : "1gSJ4nIq41",
+    "description" : "ylapPdPypPA0"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
