+++ documentation/ConferencePoster.json
@@ -1,89 +1,89 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
+  "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "1E5p8KNJgiQY1K",
-    "ownerAffiliation" : "https://www.example.org/ipsumqui"
+    "owner" : "1DxjUIZelJIkSUH6MH",
+    "ownerAffiliation" : "https://www.example.org/voluptatibusqui"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/nemonisi",
+    "id" : "https://www.example.org/exercitationemconsequatur",
     "labels" : {
-      "bg" : "E7I8Ig2VVEsdFyMU"
+      "de" : "rwu7n77ylJae9ga"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/veroest",
-  "doi" : "https://doi.org/10.1234/facilis",
-  "link" : "https://www.example.org/quispariatur",
+  "handle" : "https://www.example.org/omniset",
+  "doi" : "https://doi.org/10.1234/sit",
+  "link" : "https://www.example.org/maximeunde",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "6WqDSjegTd",
+    "mainTitle" : "Tcg3T91PdK0Pt3MLk9",
     "alternativeTitles" : {
-      "es" : "4XjBvadTUUDxNePdX"
+      "fi" : "bpHRbGpr1mSfuLTraj"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "gb3FB5FEy12vrVA",
-      "month" : "ye2IBogTfPuMk8MdMpI",
-      "day" : "CUFNTs38uleBYk0CoDg"
+      "year" : "WhGv3WCXTTl0",
+      "month" : "9JscdEnZQy7DHeujHk",
+      "day" : "Xdqzegbs5wlln"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Bax5KJHGPfIV2",
-        "name" : "JxoptxUSUf",
-        "nameType" : "Organizational",
-        "orcId" : "AfUwkRS8zYVYM"
+        "id" : "https://www.example.com/APM6S3yIwU",
+        "name" : "smmSwbkr4uTnXAU",
+        "nameType" : "Personal",
+        "orcId" : "adSv6XTvmoyJSlCv"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/myXbMoqRXKrMDVf",
+        "id" : "https://www.example.com/IPGlw0QXWYa3uua",
         "labels" : {
-          "fi" : "s80GEzosuYXm1"
+          "nn" : "2klXHuMVjd"
         }
       } ],
-      "role" : "EditorialBoardMember",
-      "sequence" : 0,
+      "role" : "Researcher",
+      "sequence" : 6,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/lFgUZW2ILK4dMNuM",
-        "name" : "rP94mgYSm8Tl",
-        "nameType" : "Personal",
-        "orcId" : "6BGdqsmcPnlcVy6qWW4"
+        "id" : "https://www.example.com/fy1dswcU2d2Cdhw",
+        "name" : "k068qt2TmOWW9",
+        "nameType" : "Organizational",
+        "orcId" : "AwcWjfz13lrt"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/3cOWkrM6U7N",
+        "id" : "https://www.example.com/6Gg4NxIwarw6ik",
         "labels" : {
-          "hu" : "Qxl9AyeGq4x"
+          "af" : "1xvpRnShNVPU"
         }
       } ],
-      "role" : "ProductionDesigner",
-      "sequence" : 6,
+      "role" : "InterviewSubject",
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "WjyIWlQ4usUYhc78Lh",
-    "tags" : [ "PDY71WkoN2USWUZ" ],
-    "description" : "gGEt0ljPuBAymu8Ux",
+    "npiSubjectHeading" : "btjMm4gO3srkL3",
+    "tags" : [ "8hfLB05VWIcN4" ],
+    "description" : "fgCAgu94prFryq",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Event",
-        "label" : "S8io8xw5RuJKIiN",
+        "label" : "GHwS8E5lcSVq5V24915",
         "place" : {
           "type" : "UnconfirmedPlace",
-          "label" : "yCDnO3DUaU8",
-          "country" : "XBOBrVfyDGqpQ"
+          "label" : "gKedCSb6ho1expME",
+          "country" : "9inIXS8S29aj05"
         },
         "time" : {
           "type" : "Period",
@@ -92,14 +92,14 @@
         },
         "agent" : {
           "type" : "Organization",
-          "id" : "https://www.example.com/Iw8Mli4bJAO1BoHa",
+          "id" : "https://www.example.com/xnZbLUn2f1gdkpo6",
           "labels" : {
-            "da" : "VhVmxLdBgReN"
+            "is" : "Z0rh6lgFx9pp"
           }
         },
-        "product" : "https://www.example.com/CQ2MCGMfhjZ"
+        "product" : "https://www.example.com/5IwRKJlwtb9nc"
       },
-      "doi" : "https://www.example.com/wlFKuy1hr1FfUJaAL",
+      "doi" : "https://www.example.com/fQ5wqnWK8G5XR0dV",
       "publicationInstance" : {
         "type" : "ConferencePoster",
         "peerReviewed" : false,
@@ -108,24 +108,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/ETs0FQwG41A",
-    "abstract" : "MimmplB7LxzzthqSN"
+    "metadataSource" : "https://www.example.com/iTUFE9PYXoY3",
+    "abstract" : "dec33AqMHpbPNq"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/aperiamsed",
-    "name" : "MkEdGAMccaxfDU",
+    "id" : "https://www.example.org/quoet",
+    "name" : "dYiRRdrmA48hFiw",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "EnRYABZ4KWPv4CZVGhK",
-      "id" : "y1oGzWR3qa79pVisM"
+      "source" : "DKem5pN7k4QF",
+      "id" : "RvLUFHu3hi8ZjqWNe"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
-      "approvalStatus" : "DECLINED",
-      "applicationCode" : "lhKJOe32lcn"
+      "approvedBy" : "REK",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "FsoDW5XZi8gS"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -133,30 +133,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/etpossimus" ],
+  "subjects" : [ "https://www.example.org/doloresdistinctio" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "8f58b9a7-8f14-4926-ba17-59bec05d6268",
-    "name" : "Z0zso9gZoDAsAQhLmFP",
-    "mimeType" : "JtcsRn0Afpxy6NLF",
-    "size" : 683144697,
+    "identifier" : "a8cf8daf-202e-420f-8039-56c783e8affc",
+    "name" : "O3z9FTEWfOV",
+    "mimeType" : "bkEbodD8ellnGqmi",
+    "size" : 1197301598,
     "license" : {
       "type" : "License",
-      "identifier" : "87uxcx8Zp2llC",
+      "identifier" : "ZlZfpypDMNFIssYqRJ",
       "labels" : {
-        "de" : "peq3ZpuOpZHuSwqTe"
+        "cs" : "ZPa0W5vLNCntSkHOSH"
       },
-      "link" : "https://www.example.com/lCqdykX6eNaq"
+      "link" : "https://www.example.com/y7Lhcz3qorBT"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/MfjfvewuFgNklaQ6M3j",
-    "name" : "h03JBeIczaK",
-    "description" : "tXZG4dw7DxfYvU"
+    "id" : "https://www.example.com/S7ESiMv9xy5PWYVoV",
+    "name" : "9CkkeEZhIxTkK",
+    "description" : "e6vQiuPNUo5N"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
