+++ documentation/DataManagementPlan.json
@@ -1,124 +1,124 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
+  "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "BSnE0UgOTJXzPLq5",
-    "ownerAffiliation" : "https://www.example.org/aliastempora"
+    "owner" : "ejSldZzdFCBGCy9",
+    "ownerAffiliation" : "https://www.example.org/placeataperiam"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/uteveniet",
+    "id" : "https://www.example.org/sitvoluptatum",
     "labels" : {
-      "fr" : "zqVvM8VQU66YyvZqd"
+      "zh" : "JPI4FEY2vGWr2MO"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/utest",
-  "doi" : "https://doi.org/10.1234/iste",
-  "link" : "https://www.example.org/itaqueaut",
+  "handle" : "https://www.example.org/laborumdolores",
+  "doi" : "https://doi.org/10.1234/asperiores",
+  "link" : "https://www.example.org/quaeratsint",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "6G2qMIcShi1bNKhCXjM",
+    "mainTitle" : "Mq8MSNucVBXd218i",
     "alternativeTitles" : {
-      "se" : "jXXBIQX9Px16g"
+      "nl" : "MG0GmSpzz4"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "5pSvsExHsypBNlk",
-      "month" : "NC6AclJhkh1OZjU",
-      "day" : "70eMkrzx9ih1316yqrb"
+      "year" : "JOpDulFH8vVH1V7E3DW",
+      "month" : "BpsiU6x5Z5gf8m4MzTt",
+      "day" : "cyNnFlrrHJ"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/1G3NDoQnuP5I4pho",
-        "name" : "OzILpzjHNkaSzkch",
-        "nameType" : "Personal",
-        "orcId" : "tcLHLSz1fb5"
+        "id" : "https://www.example.com/UzBn59i00xr9e",
+        "name" : "RgHTgRcYnalaAf",
+        "nameType" : "Organizational",
+        "orcId" : "j2BUbjn0TVm"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/n144zc4WrBI7nCQ9",
+        "id" : "https://www.example.com/3NKyychs0CgYV9Rfa",
         "labels" : {
-          "en" : "S2befXfl52GxW"
+          "es" : "u88tmDBXFJq88ZueUE"
         }
       } ],
-      "role" : "CostumeDesigner",
-      "sequence" : 1,
+      "role" : "Photographer",
+      "sequence" : 7,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/b98Q5M1ntN94W2EWy",
-        "name" : "nWRlwJtBad",
-        "nameType" : "Personal",
-        "orcId" : "QoUDEE2rS3FMs"
+        "id" : "https://www.example.com/jbU0S0b7ge7aP",
+        "name" : "UeLstU9opilc",
+        "nameType" : "Organizational",
+        "orcId" : "8IKXyB4nxTVw4GZygbG"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/GuPY7oDydIiF4g",
+        "id" : "https://www.example.com/dZVzc1lu05Z",
         "labels" : {
-          "ru" : "gFXUy8e6qSyOFCfzl"
+          "cs" : "HgDzPFOW3F"
         }
       } ],
-      "role" : "Supervisor",
-      "sequence" : 9,
+      "role" : "Soloist",
+      "sequence" : 4,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "6prklUDa0KfSF",
-    "tags" : [ "uDZufQhROZir1VCdo7" ],
-    "description" : "hPyja6RuwLlCft3SQh9",
+    "npiSubjectHeading" : "dTNU9j1i6xDl",
+    "tags" : [ "ywD0n58z74" ],
+    "description" : "nKFody5QzYqaeoX",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "ResearchData",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/pHU0Yr3paOX"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/0HuCghy5LdTxhGokw3v"
         }
       },
-      "doi" : "https://www.example.com/2il5HYg9tPelh",
+      "doi" : "https://www.example.com/Y2A3jSL9iia4vSVXWX6",
       "publicationInstance" : {
         "type" : "DataManagementPlan",
-        "related" : [ "https://www.example.com/kB2pzuC5CYv7m" ],
+        "related" : [ "https://www.example.com/MLVRK2ujZh20MpYi" ],
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "Lm4DkbyXjKR",
-            "end" : "nWoJyixfXCaq7Nyuye"
+            "begin" : "vMFNZJaXuvQeeA3x",
+            "end" : "2P8jYsBwd1EErc"
           },
-          "pages" : "3HVlvudLFZie",
+          "pages" : "9oz4ciRhTYgbEB",
           "illustrated" : false
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/eTxUYRsZRBbHZG",
-    "abstract" : "EZ7crHYrovF1"
+    "metadataSource" : "https://www.example.com/TPMIzb292MkYiEzPl",
+    "abstract" : "Xec4gXpvCd"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/consequunturnon",
-    "name" : "q2XEdwm0apzJ",
+    "id" : "https://www.example.org/voluptatesmagnam",
+    "name" : "jhzCF32qglw7h9gg",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "VGeQzWxJLMWHcdG",
-      "id" : "kK6ZMHqvHm"
+      "source" : "YQGwNKDl9XclpwmVA",
+      "id" : "SvJO6dMuxE5uM9v"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "MZkXqEeTCr"
+      "approvedBy" : "NMA",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "UtBrIpVnfvbSzCR2p"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -126,30 +126,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/accusamusamet" ],
+  "subjects" : [ "https://www.example.org/illodolor" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "735f5ca8-3850-4e33-99f3-7a4e2f393276",
-    "name" : "EtRzkTgXnzEtSy",
-    "mimeType" : "6TOh2WuB5voA",
-    "size" : 1490791096,
+    "identifier" : "493bddeb-c0df-4d84-a886-bf193321f16e",
+    "name" : "aHPghBl78B7CgL7qbm",
+    "mimeType" : "9EaiPbsETiXqYYeMrgz",
+    "size" : 710880315,
     "license" : {
       "type" : "License",
-      "identifier" : "1i1aKkqy9O",
+      "identifier" : "Gmw2UtzJEcai",
       "labels" : {
-        "el" : "MaJSkwo9iH0BAPtQIlB"
+        "pt" : "7I7j94EF1pjIQBdO"
       },
-      "link" : "https://www.example.com/wyi5Zv3RDEX06"
+      "link" : "https://www.example.com/gWRgFjMVzcxX9BFDHX"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/ZPnXOjs7zI8AoMJN",
-    "name" : "icXJUJLeWItkcuqljYh",
-    "description" : "r9N1ch2dDkgHpO2Qzb4"
+    "id" : "https://www.example.com/XeteQSa7mfp3W",
+    "name" : "9XkKbUWoFtLhYWNixi",
+    "description" : "hM5KfIXWtfvMIcqKs"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
