+++ documentation/DataSet.json
@@ -3,122 +3,122 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "wOOLGGcm1l9Q2TxbLxY",
-    "ownerAffiliation" : "https://www.example.org/aperiamaperiam"
+    "owner" : "J2A488zMngjakxTc5T",
+    "ownerAffiliation" : "https://www.example.org/voluptatumillo"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/culpased",
+    "id" : "https://www.example.org/reiciendisperspiciatis",
     "labels" : {
-      "nn" : "IAjX3UocCwtJP4"
+      "zh" : "jxdmXS94lxXb"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/magnamipsa",
-  "doi" : "https://doi.org/10.1234/pariatur",
-  "link" : "https://www.example.org/ipsamipsam",
+  "handle" : "https://www.example.org/aliasiste",
+  "doi" : "https://doi.org/10.1234/maxime",
+  "link" : "https://www.example.org/liberoquisquam",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "UDqIfH9Oy9biAsBPq5J",
+    "mainTitle" : "LtYr67UWfNWXzVYi",
     "alternativeTitles" : {
-      "it" : "sPb4zRnlcfCIw"
+      "bg" : "KDnFJpoIwXBd4R5ON"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "vBcRIw4mRuK8BoHpB3S",
-      "month" : "R82BRQeSNhJ89f4D",
-      "day" : "Ia3scUivxZSL"
+      "year" : "MYloKUm9unxZps",
+      "month" : "c3eZ4N9vfwYjjRWcq",
+      "day" : "rqUnOnZ7MpYZnl"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/ElULPiQUeOv8pInH",
-        "name" : "g6yWQz5o0o7TfyNq4",
+        "id" : "https://www.example.com/cs7CnT7gu36q",
+        "name" : "UvQOA5s2Sn1oMn1g9j5",
         "nameType" : "Organizational",
-        "orcId" : "EMwdyTfO9S04hQfjQ3"
+        "orcId" : "raswzlcOqou0vMG"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/b7JWjlP3s9w",
+        "id" : "https://www.example.com/Zker4KcjGcYefw",
         "labels" : {
-          "fi" : "WyiEwUnyWqgyIIL"
+          "zh" : "4tDqFjJg3uON5jG"
         }
       } ],
-      "role" : "VideoEditor",
-      "sequence" : 3,
+      "role" : "ContactPerson",
+      "sequence" : 9,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/JOIts1Kxh0s8fusk",
-        "name" : "MksttUGKvcu",
+        "id" : "https://www.example.com/XhIcep4KQYQs",
+        "name" : "x0Zf0uNVARRgDH45tQ",
         "nameType" : "Organizational",
-        "orcId" : "vUJTStr55XTEwKrGEs"
+        "orcId" : "yrPGasAzVrvrSCRqYj"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/SkaSL6Voz9SNCPhC",
+        "id" : "https://www.example.com/LZO5d8BHhcrnHre4O",
         "labels" : {
-          "is" : "gJHe92EmJ1pQfyy"
+          "sv" : "M3tYeKoHCgJOz4"
         }
       } ],
-      "role" : "ProjectManager",
-      "sequence" : 1,
+      "role" : "HostingInstitution",
+      "sequence" : 4,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "Bl5yBezwswm",
-    "tags" : [ "aatJTHhns4fm1NV" ],
-    "description" : "fEyw3bpDG1x8oznOxw",
+    "npiSubjectHeading" : "hEDe1jCobC",
+    "tags" : [ "VPTVpK7ffxz6ck" ],
+    "description" : "4g4yidxiOT",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "ResearchData",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/aBQNc2TxaRG"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/BAmtd1XX2LiI"
         }
       },
-      "doi" : "https://www.example.com/7Nm48nsqtQf1FXgftt",
+      "doi" : "https://www.example.com/Q6d7pbTUAx",
       "publicationInstance" : {
         "type" : "DataSet",
-        "userAgreesToTermsAndConditions" : true,
+        "userAgreesToTermsAndConditions" : false,
         "geographicalCoverage" : {
           "type" : "GeographicalDescription",
-          "description" : "HpCMgF37JwMqFQMh"
+          "description" : "61iU15XWE3kZzdgx9"
         },
-        "referencedBy" : [ "https://www.example.com/TQboXeafUu" ],
-        "related" : [ "https://www.example.com/6zKLIaTiHL" ],
-        "compliesWith" : [ "https://www.example.com/53ounzPh4IiYkDd" ],
+        "referencedBy" : [ "https://www.example.com/X5xSjbC94bNevVeYai" ],
+        "related" : [ "https://www.example.com/IXfrwRNLOLxXd" ],
+        "compliesWith" : [ "https://www.example.com/c3SD73NNQR56IeLMLi" ],
         "peerReviewed" : false,
         "pages" : {
           "type" : "NullPages"
         }
       }
     },
-    "metadataSource" : "https://www.example.com/9PVfatl7xrO271AG",
-    "abstract" : "8aAoZctyQr4"
+    "metadataSource" : "https://www.example.com/i6mk797FlfB7XNlcbtr",
+    "abstract" : "ij9T8Ki1XFBvYox"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/inet",
-    "name" : "qMH8jycuG3tNo7K",
+    "id" : "https://www.example.org/consequaturnon",
+    "name" : "jPWiW5Oqezh1qWpT",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "WFr0ZPJxktuGG",
-      "id" : "TwILccl3h8o"
+      "source" : "wrakmKm7zMtb",
+      "id" : "SPLGVRjlWL9p4Dk"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
+      "approvedBy" : "REK",
       "approvalStatus" : "APPLIED",
-      "applicationCode" : "LckqN4vMiBXlkVV"
+      "applicationCode" : "ERQYH2UNLaG09d6"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -126,30 +126,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/quiaharum" ],
+  "subjects" : [ "https://www.example.org/utipsam" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "5a7a1e1b-9310-4475-9aa2-6496d9aca933",
-    "name" : "zXWt0tYjsJW6F",
-    "mimeType" : "hpQmVXmprW7VOjns",
-    "size" : 1736886792,
+    "identifier" : "d18c4b04-d2d8-4919-adc6-cbc9b997ab51",
+    "name" : "dbv8luikvZSp0ltVdX",
+    "mimeType" : "llkvW9s0TkjVs3r1it",
+    "size" : 2088007349,
     "license" : {
       "type" : "License",
-      "identifier" : "1X5NW104eL6VLt6MhG0",
+      "identifier" : "WBm57mG8G8",
       "labels" : {
-        "de" : "besLDsbRXBrzdSnBC"
+        "nl" : "BhcDneJiESVCHtSI7"
       },
-      "link" : "https://www.example.com/vw1P1DAXtS"
+      "link" : "https://www.example.com/fYH5idTQ1v"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/h8ntmsF1XsU",
-    "name" : "gO91j9BS1ZSXN",
-    "description" : "3WLt5M5L1HP"
+    "id" : "https://www.example.com/buJgkZ5WAIe6nRW",
+    "name" : "hTHIzVXd3ZtJjZS9",
+    "description" : "rvHg5E0Pqp8R"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
