+++ documentation/DegreeBachelor.json
@@ -1,135 +1,135 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
+  "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "glDYLFGWrAb",
-    "ownerAffiliation" : "https://www.example.org/dolorumrecusandae"
+    "owner" : "5hUelGWBtLuiLYggCx",
+    "ownerAffiliation" : "https://www.example.org/remquia"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/laborumnulla",
+    "id" : "https://www.example.org/vitaeeos",
     "labels" : {
-      "se" : "f63mGDsvxdCyFKRMYY6"
+      "de" : "LbJMAFEkkMkPB"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/molestiasadipisci",
-  "doi" : "https://doi.org/10.1234/quos",
-  "link" : "https://www.example.org/rerumad",
+  "handle" : "https://www.example.org/rationefacere",
+  "doi" : "https://doi.org/10.1234/dignissimos",
+  "link" : "https://www.example.org/namqui",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "3fsBJbWQMKVEmkeW7M",
+    "mainTitle" : "JpoHDiTaJLC8WEAOmH1",
     "alternativeTitles" : {
-      "sv" : "za4xLW5aKA4rzA"
+      "cs" : "msaAlEpovRQijjR"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "DqSyd7kzqED",
-      "month" : "W39tmgPk55rM",
-      "day" : "Gf9oJAsQmv9sT0yHbRz"
+      "year" : "xQZ9cpij5Xlz8paAYpE",
+      "month" : "uJLRprVGb0l",
+      "day" : "uijkzgCTFkDY9Yc"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/sWwaScaGDjJwMYSh",
-        "name" : "8ctnl6Y4RnUih197fO8",
+        "id" : "https://www.example.com/F4PpQff4Ihe",
+        "name" : "DKcvJk9R982",
         "nameType" : "Organizational",
-        "orcId" : "CBSqFvE131CNw"
+        "orcId" : "lvd5CYbUjp"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/deqaOwOrJoJv",
+        "id" : "https://www.example.com/orN46yxofb9pPD8ssxp",
         "labels" : {
-          "af" : "izTJrCp1fT8QAm3Ok"
+          "nn" : "eZFzCuvqIjv6BL2"
         }
       } ],
-      "role" : "Supervisor",
-      "sequence" : 6,
+      "role" : "ArchitecturalPlanner",
+      "sequence" : 7,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/CuQpVZZ6KHWRImRX694",
-        "name" : "V8XEEXcxR9hkDYyZ",
+        "id" : "https://www.example.com/J7tzN7rjHE7m7T19S",
+        "name" : "tQYy0cp1jvOZa",
         "nameType" : "Personal",
-        "orcId" : "KrvLBJwTMiqiNgvTk"
+        "orcId" : "f5Ge2YXGlZdszAhq5"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/fAs5ha58gjmPu",
+        "id" : "https://www.example.com/5TxFvC2OTMNzBbxzDZ2",
         "labels" : {
-          "el" : "l4xtGvhxD5bd26"
+          "da" : "duPheSWovGsNWzCO2D"
         }
       } ],
-      "role" : "Choreographer",
-      "sequence" : 0,
+      "role" : "Supervisor",
+      "sequence" : 9,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "LuSdnEy6GqDn",
-    "tags" : [ "FOtHDrQOMTDd" ],
-    "description" : "KKhVOhbTNk",
+    "npiSubjectHeading" : "KieSmLdxYx3o",
+    "tags" : [ "u1u3R7is2tK0T" ],
+    "description" : "VPg0Ycyjw3",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Degree",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/b7926tcLFIckEJMqoj"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/yaGjeJn8qK"
         },
-        "seriesNumber" : "7UfnJ4xGq2gOsLW",
+        "seriesNumber" : "9ClZgDhWtNpL91",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/PNbIiyHtKS0hApy"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/bZlErAoZsZrTQG"
         },
-        "isbnList" : [ "9780065266313" ]
+        "isbnList" : [ "9781062229592" ]
       },
-      "doi" : "https://www.example.com/1pNrUmQSjnSI5jQdpW",
+      "doi" : "https://www.example.com/U8o8U9t4UQ1HDds1i",
       "publicationInstance" : {
         "type" : "DegreeBachelor",
         "submittedDate" : {
           "type" : "PublicationDate",
-          "year" : "NLIrEj77Ck",
-          "month" : "HBbcayQbWesR6",
-          "day" : "YuzvTG23vR3713whOgM"
+          "year" : "PSsEaYMg5IckY",
+          "month" : "U5LcumakII5s",
+          "day" : "PB2BdpaywOc0z18TC5"
         },
         "peerReviewed" : false,
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "fsVPUJ5kxMBstZ",
-            "end" : "3onO9bSLY3Q"
+            "begin" : "SbdF3fCNUAN1WZwf",
+            "end" : "leXmdzoZvAf9vt2v7jG"
           },
-          "pages" : "Z8KaZdb0LuYIpxcw",
+          "pages" : "UJj8Bo3yA9sDQbU",
           "illustrated" : true
         }
       }
     },
-    "metadataSource" : "https://www.example.com/Njc9ToE9KNmEdTeN",
-    "abstract" : "AqYJ9eOUuq00ob"
+    "metadataSource" : "https://www.example.com/eyYgiLY5YKZGRP9",
+    "abstract" : "h6H9FHA9o7a4q6tHBn"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/iustovelit",
-    "name" : "82aec3cr1SOHgrt0",
+    "id" : "https://www.example.org/quirerum",
+    "name" : "bUIhG67jxsWNEX2AUz",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "3pk8F328UZ",
-      "id" : "00GQ6yURPDZ97HO4m"
+      "source" : "ctZBHJTGDj2V9CvDjw",
+      "id" : "MOJ1BP0QcDbf"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "REK",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "ChNltg2snvdozrQ"
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "yDbnR4lD4FjFxD"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -137,30 +137,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/etvoluptatibus" ],
+  "subjects" : [ "https://www.example.org/quaeculpa" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "15d557ac-b561-455d-9288-813513806a2f",
-    "name" : "3CnELBTE1Ck1u",
-    "mimeType" : "XABHYpok4KPKfsC4F",
-    "size" : 270891000,
+    "identifier" : "4fc83081-69bc-4789-a435-7d01a3f3c598",
+    "name" : "jwB7PNU4TR",
+    "mimeType" : "Qrx74iEM1fplqxV",
+    "size" : 1670931921,
     "license" : {
       "type" : "License",
-      "identifier" : "PHVUV6zvRrY7jV",
+      "identifier" : "7YwVsb1zs2ptJew",
       "labels" : {
-        "it" : "PytGRLzqleEecpKI"
+        "nn" : "YswWC3QZ1w13yOqhIGK"
       },
-      "link" : "https://www.example.com/lg6ZDPBRqxwhhgekaLr"
+      "link" : "https://www.example.com/ZxINkLwvI7c6T"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/Dd49wwCiDxs4Dhp6zQY",
-    "name" : "h8Mrdk5ftpkfreIpoLg",
-    "description" : "GPZ6WgXdsSBCs"
+    "id" : "https://www.example.com/ZcfoestGQ2mN1",
+    "name" : "i5mL7supoxuvLHWMEBb",
+    "description" : "l2w8T8UHMWd7J"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
