+++ documentation/DegreeLicentiate.json
@@ -1,135 +1,135 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "NEW",
+  "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "n31lISgQbI3OEmv4",
-    "ownerAffiliation" : "https://www.example.org/accusamusexcepturi"
+    "owner" : "pzHRAtNznm",
+    "ownerAffiliation" : "https://www.example.org/officiishic"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/eligendinesciunt",
+    "id" : "https://www.example.org/similiquequidem",
     "labels" : {
-      "de" : "kJoVgeVSjkQl6H2Gu"
+      "zh" : "k3l3jUXvQiY"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/similiquetempora",
-  "doi" : "https://doi.org/10.1234/voluptatibus",
-  "link" : "https://www.example.org/utvel",
+  "handle" : "https://www.example.org/magniest",
+  "doi" : "https://doi.org/10.1234/eius",
+  "link" : "https://www.example.org/inventorequod",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "dl4WkBZXPEttv5MqvoC",
+    "mainTitle" : "LfJLGtZLdwA58ZFRJQN",
     "alternativeTitles" : {
-      "nb" : "XwZipLWK3e2c"
+      "es" : "1CtFvhhjY5h8AImpu"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "whaz06i4jI4kkmXSrv",
-      "month" : "Z7AVgiVYjKJQy",
-      "day" : "FzGD2ky8E1X6BiWoa1a"
+      "year" : "VEKIr6ERAAjv2w",
+      "month" : "yIU2AgjbtHNO",
+      "day" : "3o1wEqxvOBSxNM"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/n9pd7yDMZPtYvF",
-        "name" : "zxKI6b7yljEi",
-        "nameType" : "Personal",
-        "orcId" : "42TL49NyKtVwdB"
+        "id" : "https://www.example.com/b9VFAOsTxxgTOJqzr",
+        "name" : "pfgAcGKFl7Ri",
+        "nameType" : "Organizational",
+        "orcId" : "UYWEM5RSZImUiwq01"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/odVpkf4szET",
+        "id" : "https://www.example.com/DXfV3bSxRQKVekvMO",
         "labels" : {
-          "da" : "WvgP7gtI8XdAY"
+          "el" : "D3z5c0cv31"
         }
       } ],
-      "role" : "ProductionDesigner",
-      "sequence" : 3,
+      "role" : "Screenwriter",
+      "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/1mnQ2kccqlMIORvR",
-        "name" : "yTeeO6V3CHToE",
-        "nameType" : "Personal",
-        "orcId" : "AvBoLnrRpVLcCPL3RqK"
+        "id" : "https://www.example.com/hljf5UhKVaC7tD1",
+        "name" : "Tg8lgHAtFeoER4",
+        "nameType" : "Organizational",
+        "orcId" : "FK3IM4vD02po28"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/JBvzUcXBkla9",
+        "id" : "https://www.example.com/AJpUAuAmlrZNoR",
         "labels" : {
-          "it" : "i8Pksf8VtelV8y6A5"
+          "cs" : "20KDmU69R2m6Sy1YPs"
         }
       } ],
-      "role" : "Editor",
-      "sequence" : 5,
+      "role" : "Designer",
+      "sequence" : 3,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "nUg6CqUuXgf9oLdOmJV",
-    "tags" : [ "54BZKJ3NnkESb" ],
-    "description" : "ay1sywhprU6W",
+    "npiSubjectHeading" : "sCH6KsfASFb3heUoxDn",
+    "tags" : [ "KMC2Fl4lL5tgyEstK76" ],
+    "description" : "XuRn0zGjORLQ",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Degree",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/9ErIsqsS3coE7"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/jt8VGQFcIFI4mA2"
         },
-        "seriesNumber" : "h1q8DqP0xs6XGtD",
+        "seriesNumber" : "7HLgvj0ZVI",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/mmrBfSYgWIcqq"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/cIyvOwvDGauH"
         },
-        "isbnList" : [ "9781905842681" ]
+        "isbnList" : [ "9790993684004" ]
       },
-      "doi" : "https://www.example.com/2BH7dxVg4jp",
+      "doi" : "https://www.example.com/FcUnyjMF3Z4iL0L",
       "publicationInstance" : {
         "type" : "DegreeLicentiate",
         "submittedDate" : {
           "type" : "PublicationDate",
-          "year" : "1CS8PyJaEzh4HIFo",
-          "month" : "2lqsNCdlRH",
-          "day" : "1EvE5XaPebK"
+          "year" : "RYiGQ0VzChdCi",
+          "month" : "Pg7Xjc8AX4XiUy",
+          "day" : "RkzcJQQ8YQXcYSBY"
         },
         "peerReviewed" : false,
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "ZHKzUmPEB58fa3CU",
-            "end" : "wzGTmciHqMuYU"
+            "begin" : "iwiYeriTeAOd1z",
+            "end" : "X2FTnN9zLT"
           },
-          "pages" : "ftFf1p0tHnp",
-          "illustrated" : true
+          "pages" : "t8kcKGSzFzyr20C",
+          "illustrated" : false
         }
       }
     },
-    "metadataSource" : "https://www.example.com/eitdN3Cfmj5eWBC",
-    "abstract" : "OCAiB4GhEL"
+    "metadataSource" : "https://www.example.com/CukNusiN5YY",
+    "abstract" : "ACAdesriciC"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/estadipisci",
-    "name" : "czR5QDuVKlTOfsMkRc",
+    "id" : "https://www.example.org/adipiscienim",
+    "name" : "rtGjzeoDpjkVQFw",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "7dB3fptNs64OaMZ",
-      "id" : "mlnlMtXgzE8XRRVoz"
+      "source" : "HP2W4CbqW0tIii",
+      "id" : "XR2tT6rCeyBgRgDrraO"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "NMA",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "JskY4vNC2HGT6H"
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "SxoCcM19kPJCrn07vAC"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -137,30 +137,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/impeditsuscipit" ],
+  "subjects" : [ "https://www.example.org/quideleniti" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "c3de74e6-dbba-4dd8-8179-770f75bccda8",
-    "name" : "yCXYcpdhSIJUr9",
-    "mimeType" : "62ZcS8VTXwbSJp52",
-    "size" : 1943318868,
+    "identifier" : "da8708c5-694c-4c6a-a33e-36729c93dde7",
+    "name" : "SPkKDftQrWtHfxyZCpZ",
+    "mimeType" : "WICNIows8v5SEoYEchF",
+    "size" : 657933688,
     "license" : {
       "type" : "License",
-      "identifier" : "nmIihXq3TG4okc",
+      "identifier" : "63yLigDj87fhdjNri98",
       "labels" : {
-        "it" : "zPXYzx2SKVkWdv"
+        "hu" : "Z2lLgJ2g7ergaz"
       },
-      "link" : "https://www.example.com/Vp0r4GrsAU83"
+      "link" : "https://www.example.com/EDapwTdvHFRU30xgo"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/CgwOVtLcYvNaNb7telw",
-    "name" : "SFjmLnKP1pkE",
-    "description" : "vx82yFrSiJyi"
+    "id" : "https://www.example.com/Ie3VsYm51K8PzeFz6",
+    "name" : "OfV4WvJNl6",
+    "description" : "1211JlgT4zaast1JBY"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
