+++ documentation/DegreeMaster.json
@@ -3,133 +3,133 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "Hz2pVxBuS7mOn",
-    "ownerAffiliation" : "https://www.example.org/illummolestiae"
+    "owner" : "w6LrsJqiQeNMkzj0p",
+    "ownerAffiliation" : "https://www.example.org/iurequia"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/namrepellendus",
+    "id" : "https://www.example.org/fugiatmollitia",
     "labels" : {
-      "hu" : "zXxsJm8BcXWdLw"
+      "pl" : "u98nNMFPmS2R"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/minimaodit",
-  "doi" : "https://doi.org/10.1234/pariatur",
-  "link" : "https://www.example.org/uteos",
+  "handle" : "https://www.example.org/doloremdeserunt",
+  "doi" : "https://doi.org/10.1234/nisi",
+  "link" : "https://www.example.org/quissaepe",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "dx7TUMiQ2w5yM",
+    "mainTitle" : "w0Tle7Mfmpt",
     "alternativeTitles" : {
-      "sv" : "oPvOzREqoqBv2i"
+      "fr" : "ySt7gRcDJF97EaBZPu7"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "bJhhK92AW5",
-      "month" : "B2YdreBukZmtL2tvxxD",
-      "day" : "ThVcFHCsdA6nPaIXSt"
+      "year" : "fjpAxw7LIoj",
+      "month" : "QFJZgazfFSiFpY",
+      "day" : "Y5oFYEIIYbgd"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/0KdmRDdxFG",
-        "name" : "MSAZnDrixjUemKZX",
-        "nameType" : "Organizational",
-        "orcId" : "7Jr5scvpeq"
+        "id" : "https://www.example.com/UI0uvzftVfF",
+        "name" : "EquhkkGpgNo0zbSX",
+        "nameType" : "Personal",
+        "orcId" : "Cjv0002Aooa3qH"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/O3ysir34jGD",
+        "id" : "https://www.example.com/mJgqHS1LZBgTV",
         "labels" : {
-          "cs" : "BkRX4NMiqS"
+          "ru" : "sF2in68ZzSlceAHOIv"
         }
       } ],
-      "role" : "VfxSupervisor",
-      "sequence" : 1,
+      "role" : "Distributor",
+      "sequence" : 8,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/MTe9XPioquTnoqlXe8f",
-        "name" : "v2O18eKtF5P",
+        "id" : "https://www.example.com/xyIwUwg3Qv",
+        "name" : "m6c4KWzw1HGwIlvS",
         "nameType" : "Organizational",
-        "orcId" : "jl8iHiOKU7YyQuGFM"
+        "orcId" : "H9oaCjFEg2exiaCHzxJ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/paKZ8i0VR3eFGyHFD4",
+        "id" : "https://www.example.com/F5wNT80acgYNElFoJQR",
         "labels" : {
-          "bg" : "ag3zl0LMaAsQG"
+          "sv" : "mWVLshoxotfYusJrq"
         }
       } ],
-      "role" : "VideoEditor",
-      "sequence" : 1,
+      "role" : "ArtisticDirector",
+      "sequence" : 5,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "DAoIDteQWvVg75G",
-    "tags" : [ "EVnABsekyaM" ],
-    "description" : "LaOPhQLg5a",
+    "npiSubjectHeading" : "QWVKJI4sOXdodp",
+    "tags" : [ "YOLZng6pTFPUlSVjiAs" ],
+    "description" : "jv8FKCPdY7pi",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Degree",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/oEnX3wTklH"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/RWz0OnAcmm0JLVy6o"
         },
-        "seriesNumber" : "1vT3qmWhUR5Zyj1",
+        "seriesNumber" : "oTPqCL7vJ0f977ks80",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/ukG8x8gyYNG3VLPfbU"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/VKD9t6jfWjuMcLdHw"
         },
-        "isbnList" : [ "9781920147280" ]
+        "isbnList" : [ "9780035945330" ]
       },
-      "doi" : "https://www.example.com/btoP2INkzKaQOg",
+      "doi" : "https://www.example.com/RuuucHsB3CFxBtPw61",
       "publicationInstance" : {
         "type" : "DegreeMaster",
         "submittedDate" : {
           "type" : "PublicationDate",
-          "year" : "ZWrimeqTW2rlts",
-          "month" : "3ndZS6HIZn",
-          "day" : "mPXMEDOk2IIM8BB"
+          "year" : "4hky6e4mMmVzwf1Zj",
+          "month" : "v9O3FivwLO",
+          "day" : "v4vjgKmZiZo"
         },
         "peerReviewed" : false,
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "ThLEquCfPMywzjFit",
-            "end" : "rh6jn3RgX96lkc1UabK"
+            "begin" : "tpsfqnh5WWhtiw2dH77",
+            "end" : "mJi5vO8EGwVa"
           },
-          "pages" : "ZCsLivihv17mkQDmtv",
-          "illustrated" : false
+          "pages" : "gg8IlwFps8",
+          "illustrated" : true
         }
       }
     },
-    "metadataSource" : "https://www.example.com/wYbuRN5HlbO0u5bQTGm",
-    "abstract" : "fjO4jv0za3DJ"
+    "metadataSource" : "https://www.example.com/3ez0pPq1qKrnqQNj4",
+    "abstract" : "xybM9mDszQS7MFOL9"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/omnislabore",
-    "name" : "aiH4iN5VjXrys8IgtC",
+    "id" : "https://www.example.org/liberosint",
+    "name" : "TUieX4NKeTou2",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "APkhzYf6Ta6l",
-      "id" : "cDqqzmQlbg"
+      "source" : "TOyKj1cp6ChW",
+      "id" : "f1rt9ucCaHMKEiHbQf"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
+      "approvedBy" : "REK",
       "approvalStatus" : "APPROVED",
-      "applicationCode" : "N7tyuefdjUh6SIE"
+      "applicationCode" : "fCjTBbVpQqrMQ"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -137,30 +137,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/temporaut" ],
+  "subjects" : [ "https://www.example.org/reiciendisdolor" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "58c9d1a5-8ede-4996-af8c-ca21566fecf9",
-    "name" : "zKKnJBpxettl",
-    "mimeType" : "w2q4b7vYxckVcmDq",
-    "size" : 1602782233,
+    "identifier" : "886b762e-4368-4665-bfb9-b8ed3fa9ecee",
+    "name" : "PdaRPlwSxnzq",
+    "mimeType" : "ChoEPlfxL3hvINeE",
+    "size" : 1459979760,
     "license" : {
       "type" : "License",
-      "identifier" : "VFFl96DF3By96xQHWH",
+      "identifier" : "I0P8ERHDNQ",
       "labels" : {
-        "hu" : "VkoHrdHp4tCxftqne"
+        "it" : "ZV0oPFIlvh"
       },
-      "link" : "https://www.example.com/OZ450vx4ENt"
+      "link" : "https://www.example.com/KksxANkYypI"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/gLNrWcTrXO2sYOmfRv",
-    "name" : "EnhIDSmJjjqoJv",
-    "description" : "81tXXdyEQf"
+    "id" : "https://www.example.com/ogiwmgfPlY99mnb",
+    "name" : "gdk0O5UAprs9",
+    "description" : "bWY5L5BXpfEq1"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
