+++ documentation/DegreePhd.json
@@ -1,135 +1,135 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
+  "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "QQ76aBHrxU6",
-    "ownerAffiliation" : "https://www.example.org/maioresmaxime"
+    "owner" : "4plqIGO1HBruqUVJaZy",
+    "ownerAffiliation" : "https://www.example.org/itaquecum"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/voluptateblanditiis",
+    "id" : "https://www.example.org/auteos",
     "labels" : {
-      "nl" : "gP0vLi0TAxf7"
+      "pl" : "lU4RH7G2iPxiwXx2"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quisasperiores",
-  "doi" : "https://doi.org/10.1234/laborum",
-  "link" : "https://www.example.org/temporearchitecto",
+  "handle" : "https://www.example.org/eiussed",
+  "doi" : "https://doi.org/10.1234/quidem",
+  "link" : "https://www.example.org/quoin",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "aCg3X7poTZa0kZpKNL",
+    "mainTitle" : "tadoNs1SaB6pN0v8j",
     "alternativeTitles" : {
-      "el" : "kVMEETEdDY"
+      "da" : "dyp7zgxMXjoI"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "pknpA8Aii9McAb",
-      "month" : "YWrGXfUpfVvBB4GF4r",
-      "day" : "ndr8DEM89Rkj7yVe6"
+      "year" : "2kwIHOxTrFqqvz",
+      "month" : "eNpL9OESNAUYNPwRkW",
+      "day" : "anUUopmCVIeRM4zbd"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/zjzqE2n5zxr0oJR",
-        "name" : "WxiVTN73Of3zc",
+        "id" : "https://www.example.com/sluBxgsVMQRiicW5",
+        "name" : "mREJdH9eOEgF",
         "nameType" : "Organizational",
-        "orcId" : "u3yie6zwn8rwQPVA9fJ"
+        "orcId" : "zdsNjBXUUPJ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/WA66iLIzD5R",
+        "id" : "https://www.example.com/s0D61AuTcexT",
         "labels" : {
-          "el" : "fCvneO3xGPdT4dO3s"
+          "es" : "6owQvK4GRU"
         }
       } ],
-      "role" : "Conductor",
-      "sequence" : 1,
+      "role" : "Choreographer",
+      "sequence" : 6,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/WBo6yBjFWsz2hP3L0z",
-        "name" : "qdTZ3r7qNJVYWcfx",
-        "nameType" : "Organizational",
-        "orcId" : "qlQHG087Su0a8hb"
+        "id" : "https://www.example.com/crn19y4Jy0Fixbs0Bu",
+        "name" : "UOf8a2oAh5lkK1",
+        "nameType" : "Personal",
+        "orcId" : "7YwqMjOsX2w7YGZ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/tqKSgVakSfL",
+        "id" : "https://www.example.com/tRM0MdoeuiRaoceD",
         "labels" : {
-          "hu" : "Fid541SOABpMH"
+          "el" : "83FkbSZBpRqXWcnjUo"
         }
       } ],
-      "role" : "AcademicCoordinator",
-      "sequence" : 2,
+      "role" : "Photographer",
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "3cFLhVMJmJRgAg2GQ",
-    "tags" : [ "lOGBuFvZvFztEHpSkxw" ],
-    "description" : "pubDOuHp3ekAXHNg",
+    "npiSubjectHeading" : "q4cJKVYYUgttOBEkv",
+    "tags" : [ "8mv7EfNDrmZQ4YFQ3fx" ],
+    "description" : "JCuIyKNI0EuB",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Degree",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/3nwpUk10fs"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/ztTtE6WS6sDR4nb"
         },
-        "seriesNumber" : "N1RtCqFkDrh",
+        "seriesNumber" : "qv7I9fpt6JaIP67SJH1",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/Z6R26NBTmNi"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/m5U8gq3VMu5er"
         },
-        "isbnList" : [ "9781517502942" ]
+        "isbnList" : [ "9790707765890" ]
       },
-      "doi" : "https://www.example.com/nEuT8EVDLgh8ptX",
+      "doi" : "https://www.example.com/qt7OqeMAoGs",
       "publicationInstance" : {
         "type" : "DegreePhd",
         "submittedDate" : {
           "type" : "PublicationDate",
-          "year" : "OBsFVjnZKzhXZM",
-          "month" : "3oCVqmzzXuK5",
-          "day" : "ZbmoToNQ4CFzQdlx"
+          "year" : "GhImhUNWkSP",
+          "month" : "ba9kZM7B1uD",
+          "day" : "68sWVnhLbeZ3"
         },
         "peerReviewed" : false,
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "vkZgmzvvV387ZXsyh5",
-            "end" : "UDcHRicy82Fxzo"
+            "begin" : "rTNQuqghXBswT",
+            "end" : "Li3tmTfFeMXq"
           },
-          "pages" : "wJv8OcnTxU3cTi7",
-          "illustrated" : false
+          "pages" : "hVSC97Q4nKrn",
+          "illustrated" : true
         }
       }
     },
-    "metadataSource" : "https://www.example.com/WccjXdM3zNz0",
-    "abstract" : "h5afysoDcb"
+    "metadataSource" : "https://www.example.com/kMy2KURxT4Gux2C8",
+    "abstract" : "XXouBlBrtvlvtKWBMA"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/eteos",
-    "name" : "eVctsUROONi",
+    "id" : "https://www.example.org/quiut",
+    "name" : "jvCEQ9wvcm7Qp6",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "r86cTerk0cwr",
-      "id" : "HKdWn3zPd1KDBTP"
+      "source" : "eycCBnywyis3",
+      "id" : "kUgFhQuO4xMT"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "zCYk0Hl6ttEEfOKGjj"
+      "approvedBy" : "DIRHEALTH",
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "7wEySa9h84uLZadX"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -137,30 +137,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/istelibero" ],
+  "subjects" : [ "https://www.example.org/assumendacorporis" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "fcd1cfb4-3515-41f3-bb63-de74b7c3d87c",
-    "name" : "T0vVmb6nqbXL",
-    "mimeType" : "qysjpG2Dtx0p3OlCx",
-    "size" : 1735448271,
+    "identifier" : "92da0495-18b8-4304-b3b1-cef80344afc0",
+    "name" : "OkpnmCqIp7Ghu64jG",
+    "mimeType" : "XilHZ0fP61XL3JX4n",
+    "size" : 1179615775,
     "license" : {
       "type" : "License",
-      "identifier" : "3Y3qXp3BTh5",
+      "identifier" : "QEyR2MN5vM",
       "labels" : {
-        "ca" : "wf2sQ3i4Oa0h46uZ2hR"
+        "ru" : "ySW1dWsfAd94e5CSL"
       },
-      "link" : "https://www.example.com/dPf2AAtcUfv3rg2rK06"
+      "link" : "https://www.example.com/svCbrIZlYCnq3ldv"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/Z6hNJQLNifsfwQRG",
-    "name" : "KlnLvP0pgzEam9V48",
-    "description" : "hm4UfkDsEYvm5IX"
+    "id" : "https://www.example.com/UO8xIYJMTwHC4s",
+    "name" : "BPFaWT9Lv4GQnsQUUU",
+    "description" : "DG76aXgyYRKpDR"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
