+++ documentation/FeatureArticle.json
@@ -3,116 +3,116 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "nEOA2v581fFV",
-    "ownerAffiliation" : "https://www.example.org/perspiciatisest"
+    "owner" : "7pWyFJDjgB4rlew3v",
+    "ownerAffiliation" : "https://www.example.org/autin"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/utqui",
+    "id" : "https://www.example.org/estmaiores",
     "labels" : {
-      "da" : "gLzfEqMkAsjpx"
+      "is" : "qe4irRuM9nUwaW"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quiaconsequatur",
-  "doi" : "https://doi.org/10.1234/non",
-  "link" : "https://www.example.org/nequeest",
+  "handle" : "https://www.example.org/possimusad",
+  "doi" : "https://doi.org/10.1234/et",
+  "link" : "https://www.example.org/utmagni",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "YeLIoYHFBlkC",
+    "mainTitle" : "rCqb7U3Ciz3U",
     "alternativeTitles" : {
-      "en" : "pTN7n715s45o"
+      "da" : "9xp4gO7YbVnoayXDtLx"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "fTCwC2qUPFNQVq",
-      "month" : "ohQ1eHNXSANIDl2IU5",
-      "day" : "ieVO1QN24BmjZ"
+      "year" : "91zDrZXAgzuJzjpZan",
+      "month" : "kW1P9tppG1QlBEpWX",
+      "day" : "8RbLqobfZUzlzxmVn"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/PXXsMhlhqX7QoH0Ye",
-        "name" : "Jrs46F4qDHJi8",
-        "nameType" : "Personal",
-        "orcId" : "gUP1jnwCAs"
+        "id" : "https://www.example.com/7TKYLnZjbJkelrhh",
+        "name" : "JQeJZSaohEkTGVzSaz",
+        "nameType" : "Organizational",
+        "orcId" : "86DicAJr8g3"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/nUpZ4BArU7SmrmQQO",
+        "id" : "https://www.example.com/iuTtACOGcGX3eKVK",
         "labels" : {
-          "nn" : "kyjYzvsiskxh4FYyO"
+          "da" : "hLohE5gvva7GIO"
         }
       } ],
-      "role" : "ProgrammeLeader",
-      "sequence" : 7,
+      "role" : "ProjectMember",
+      "sequence" : 3,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/J2yiMSvRZ1rVY2sf1",
-        "name" : "RXuMnvTMJEPJz",
-        "nameType" : "Organizational",
-        "orcId" : "8daQQn6VYkzOdu"
+        "id" : "https://www.example.com/Yxgjdd06Iz",
+        "name" : "DmvGMKiENJs",
+        "nameType" : "Personal",
+        "orcId" : "zeslR5ny8cgYtR"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/UY6PQbGHs9vlsb",
+        "id" : "https://www.example.com/x8tess3ieCOGv7",
         "labels" : {
-          "hu" : "iapqjEilYtFt"
+          "fi" : "ZxEB3QtjInB"
         }
       } ],
-      "role" : "CuratorOrganizer",
-      "sequence" : 6,
+      "role" : "Funder",
+      "sequence" : 1,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "snLltgapRw8Kxa9",
-    "tags" : [ "0WBBv90HiyhPO0SJrQ" ],
-    "description" : "mvNEKAVBPf48tO",
+    "npiSubjectHeading" : "debeQejddyy",
+    "tags" : [ "XquPEnkLxMc" ],
+    "description" : "01QoA0rExT",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/ITtNPm6nqKFeNduxW"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/P7o91pGIts1bADSVo"
       },
-      "doi" : "https://www.example.com/15LUu5EBBq",
+      "doi" : "https://www.example.com/smHtFLca6ZUoA",
       "publicationInstance" : {
         "type" : "FeatureArticle",
-        "volume" : "7qCJ5GgoARg",
-        "issue" : "hWHaZhWzYmp",
-        "articleNumber" : "bW8SaJPGLajbTxjU",
+        "volume" : "mcvEwpdD75W890C",
+        "issue" : "q3FiilqF0ame6qc",
+        "articleNumber" : "Ij8NFKZdXHVKNVOky",
         "pages" : {
           "type" : "Range",
-          "begin" : "Ybt2tmnUV4hl",
-          "end" : "gi36vkljHwCcA55QaV"
+          "begin" : "ruStEhkaqLEaOaUh6f7",
+          "end" : "gsJFkYn6meHo"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/6OHbrFIfnRAT9gafs",
-    "abstract" : "YnRVbveuqH5Plq"
+    "metadataSource" : "https://www.example.com/e19mogqEUF",
+    "abstract" : "f2bQhnCtMbSSZgVFJgr"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/sapienteplaceat",
-    "name" : "vNYdTE0OKw",
+    "id" : "https://www.example.org/eiussunt",
+    "name" : "nI3sIxehvPEp",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "RRfqvcyOSU4",
-      "id" : "ICiaS9IRHo3Fjoxl"
+      "source" : "TTGlzv74q8sFYsNXsQR",
+      "id" : "JK7yaPmJzTAmzU"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "REJECTION",
-      "applicationCode" : "dGcJjMF3py0IT"
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "HKiQteTIe6J"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -120,30 +120,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/illumfuga" ],
+  "subjects" : [ "https://www.example.org/excepturiquasi" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "bdfa18cd-eea5-47c5-a2b0-bb47d86656bb",
-    "name" : "tKIYacsiL2RZdjLQOq",
-    "mimeType" : "ZMBUi7vbFst",
-    "size" : 1293765002,
+    "identifier" : "d1931df9-64cf-49af-9524-5922b70391be",
+    "name" : "hq8ZvOGMMf",
+    "mimeType" : "Hgmv0KsYOTXDCNnLrm2",
+    "size" : 226343738,
     "license" : {
       "type" : "License",
-      "identifier" : "jnz1w67oFtRKB9",
+      "identifier" : "8uOyPMS0YQ5j",
       "labels" : {
-        "hu" : "nZjTpDLzgVEuItdOVgn"
+        "hu" : "GthBtltDWcyC"
       },
-      "link" : "https://www.example.com/eHfjRWlLjWr"
+      "link" : "https://www.example.com/aDUDBfF0bUKQg"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/Bvu2Jyh0HjP8lhPL",
-    "name" : "HFsvs36VCd",
-    "description" : "P5tGEGfDnzy2rX1Z"
+    "id" : "https://www.example.com/Tlq8qGh6oLV",
+    "name" : "bRdIaWtwpT3xO59I",
+    "description" : "rjH1cOPh1Sl"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
