+++ documentation/JournalArticle.json
@@ -1,120 +1,120 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "D5oQ7ICNXUfa1iRo",
-    "ownerAffiliation" : "https://www.example.org/assumendarerum"
+    "owner" : "m8hUpny8wFcTbe24YI",
+    "ownerAffiliation" : "https://www.example.org/etfacilis"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/veritatisquis",
+    "id" : "https://www.example.org/repellendustotam",
     "labels" : {
-      "fr" : "GpGZcHh8D2hedE7rR"
+      "de" : "CadMgN2TkzOhek7P"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/reiciendisqui",
-  "doi" : "https://doi.org/10.1234/quo",
-  "link" : "https://www.example.org/teneturtempora",
+  "handle" : "https://www.example.org/nihilaut",
+  "doi" : "https://doi.org/10.1234/dolore",
+  "link" : "https://www.example.org/eaquequi",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "CRdhNnVUAd",
+    "mainTitle" : "maxCgIJta5LqjTlS",
     "alternativeTitles" : {
-      "af" : "Swc0QQADwJZC"
+      "is" : "EsTDAVZy1wjab"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "gb7Muw9ujCe",
-      "month" : "dJEd8NihCL8jxFLN",
-      "day" : "PGoN926kAQ"
+      "year" : "aaAp9VN3Ez5j0c",
+      "month" : "yhtsskKmDZfItYyKF",
+      "day" : "dHgElSDXwEwL"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/GsscsJ5SMg",
-        "name" : "mc5VP8yZjTBHSjEQO5X",
-        "nameType" : "Personal",
-        "orcId" : "NL4ppljPjAHMtZ"
+        "id" : "https://www.example.com/osLO32es2ZZd0vR",
+        "name" : "bcNlDvjfqElD",
+        "nameType" : "Organizational",
+        "orcId" : "roK1PEyjBDkD2M0"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/TaVWiDWXx1",
+        "id" : "https://www.example.com/speG1FcK76EP",
         "labels" : {
-          "pt" : "yywyGsqBsYu0GCS3xhL"
+          "el" : "rSc0lEfQtWeXS9Szok"
         }
       } ],
-      "role" : "RelatedPerson",
-      "sequence" : 4,
+      "role" : "Creator",
+      "sequence" : 1,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/c1gqqoBJG7OSxt6",
-        "name" : "7oNGsCqoErMYysm",
-        "nameType" : "Organizational",
-        "orcId" : "KJoS8sFCNbqxfqR"
+        "id" : "https://www.example.com/tJ0UdaMiBPK7",
+        "name" : "xAjxwNYf3fiRD4aYive",
+        "nameType" : "Personal",
+        "orcId" : "5xZHLVB1oND"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/E40815ZJB98tLi",
+        "id" : "https://www.example.com/qC4jWPmCc3kKzQ",
         "labels" : {
-          "ru" : "UDTmJTeBIMrX7xg4v"
+          "af" : "DbaZb3PTI6"
         }
       } ],
-      "role" : "Dramatist",
-      "sequence" : 1,
+      "role" : "Funder",
+      "sequence" : 5,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "wRuNip3N3S",
-    "tags" : [ "PnHt0yrlFIgPFSkFka" ],
-    "description" : "OG8t2nJbBn",
+    "npiSubjectHeading" : "gV5fYfR9djcJBnv",
+    "tags" : [ "51f0iUimYFlO" ],
+    "description" : "bnc1di0g7DCE8tV2ygN",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/AVmJpEyJK9Wx"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/y4uGPl0L5dKB"
       },
-      "doi" : "https://www.example.com/4SdCbcTzR8BdZOkFP",
+      "doi" : "https://www.example.com/Iq79BXt5EKqwTXkZgdl",
       "publicationInstance" : {
         "type" : "JournalArticle",
-        "volume" : "tDXqJpGus2qpcqTJY95",
-        "issue" : "OLGqSD1crFWHzWBO1",
-        "articleNumber" : "jKSDnBWNe6",
-        "contentType" : "CaseReport",
+        "volume" : "uXfSIhAeKwCVa",
+        "issue" : "etO4P3FTjvBSm",
+        "articleNumber" : "ThfVxXUOp2",
+        "contentType" : "StudyProtocol",
         "peerReviewed" : false,
         "originalResearch" : false,
         "pages" : {
           "type" : "Range",
-          "begin" : "zl7VDOG2KA6ZRNw",
-          "end" : "vgbXPHNAsUe4HYT"
+          "begin" : "8i7ugIkem8Pqk31L1y",
+          "end" : "IDNfnU5Zy2SSxKHtH"
         }
       }
     },
-    "metadataSource" : "https://www.example.com/Eoi5F03gfyN9",
-    "abstract" : "U4ggReJsbS8htGY"
+    "metadataSource" : "https://www.example.com/Icn92YGROdCRLpL1",
+    "abstract" : "VqMkmSZmGjXrZx9TY"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/estnatus",
-    "name" : "A2rXSh9gg6mQGsJQGM",
+    "id" : "https://www.example.org/eumlaboriosam",
+    "name" : "Zfz1QT6yyvf4U",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "GzMSYCl6moMUJy0j",
-      "id" : "myfweDciTbLUD"
+      "source" : "MJqCginN4sCON9XsdIX",
+      "id" : "68xheB7zwMPYi27AaaA"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "DECLINED",
-      "applicationCode" : "VwISrYKnoY4o"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "qT9ggz1BQZR"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -122,30 +122,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/quiexercitationem" ],
+  "subjects" : [ "https://www.example.org/vitaemaiores" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "b349460a-879b-4add-89f4-81ead221bc25",
-    "name" : "I5m8YdI91hKiG16",
-    "mimeType" : "2Cu3OydGxctQVNmzUNJ",
-    "size" : 418715236,
+    "identifier" : "26a32b23-d216-49a2-9b28-ad8e305d3dd2",
+    "name" : "LEwyMg8ifzB0rVkhQ",
+    "mimeType" : "yvWTdEUdkDhb2jU",
+    "size" : 1874426373,
     "license" : {
       "type" : "License",
-      "identifier" : "PWLDb5bHCl",
+      "identifier" : "FfGCXCqXrTXZjVhZ6b",
       "labels" : {
-        "nb" : "EwicvL1aJ1xCpViYna"
+        "pl" : "08iTrYmwOo4"
       },
-      "link" : "https://www.example.com/Fzd4kQGAkf"
+      "link" : "https://www.example.com/w8bEVV72wq7PQ"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/oweaSNt5yxqGjHDNquE",
-    "name" : "KUwViXAloinuRLoPzBK",
-    "description" : "rP1jeFxizPHM09"
+    "id" : "https://www.example.com/Sqr9ksPbo1Gg4GWt2M",
+    "name" : "6nD2RSw6LNUjm",
+    "description" : "GxpdHZYLPfFcH5jscH"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
