+++ documentation/JournalCorrigendum.json
@@ -1,119 +1,119 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "ARnvqLQaANhAusK4lPJ",
-    "ownerAffiliation" : "https://www.example.org/pariaturrepellendus"
+    "owner" : "A8hH04r8cGNNREA",
+    "ownerAffiliation" : "https://www.example.org/quasa"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/maioresdignissimos",
+    "id" : "https://www.example.org/inquisquam",
     "labels" : {
-      "fi" : "qO8F5KolnpV0kb7I8"
+      "zh" : "egTjOdOaYjMmY5HW9C6"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/suscipitminus",
-  "doi" : "https://doi.org/10.1234/nihil",
-  "link" : "https://www.example.org/nondolor",
+  "handle" : "https://www.example.org/natusvoluptatem",
+  "doi" : "https://doi.org/10.1234/rerum",
+  "link" : "https://www.example.org/cumillum",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "K2jrbmSQZfYXqazq1",
+    "mainTitle" : "BiH9H3AKLZxd6w0A",
     "alternativeTitles" : {
-      "is" : "HExKTZuGRXr40n"
+      "el" : "Y0brAn6A3fQ2DV"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "h4AoJeWRAvk71cS",
-      "month" : "0x3zHbAX4HXn7",
-      "day" : "MIThGk1JyoxV1N"
+      "year" : "WLfAZhZCJLDnjUIjg",
+      "month" : "HAH59LeWr5HZNb4je",
+      "day" : "PNjGQWD0V8kP2Rr"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Ycfl7ooxVrMFkiFU2",
-        "name" : "mH5a2SaSmLvB",
+        "id" : "https://www.example.com/VGzMD9azTk89",
+        "name" : "lTelghWUyXDiz",
         "nameType" : "Organizational",
-        "orcId" : "grNW2HxbmRr6ISI4D"
+        "orcId" : "aDy7yIcYShaI7CRReY6"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/GaBo56srgIbvgbIHdKK",
+        "id" : "https://www.example.com/ikaEX7gVA0e",
         "labels" : {
-          "de" : "DfOlc0uGUi27"
+          "it" : "JSh44jafS8"
         }
       } ],
-      "role" : "WorkPackageLeader",
-      "sequence" : 1,
+      "role" : "RelatedPerson",
+      "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/vUfehY1Blwh8Xg",
-        "name" : "vr7xqXpJqqz55dXRUFP",
-        "nameType" : "Personal",
-        "orcId" : "vDmsKpjU6rgOBxhwF"
+        "id" : "https://www.example.com/AVYCS7OQclYBVz94V7E",
+        "name" : "lsp8fTd3PzTgedBZ",
+        "nameType" : "Organizational",
+        "orcId" : "EKYonLOFlixW"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/sFfNbH0bVQacO8Lbd3",
+        "id" : "https://www.example.com/BL34ccVS7R1",
         "labels" : {
-          "pl" : "sYktEyE9DscaZfWxcal"
+          "ca" : "R2QKwQMJ12Tf5t4"
         }
       } ],
-      "role" : "ArtisticDirector",
+      "role" : "Designer",
       "sequence" : 4,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "gE0AKeKgopr8q",
-    "tags" : [ "RaohFyary5xeMFqL9t8" ],
-    "description" : "Y0BXiybYCndalh",
+    "npiSubjectHeading" : "wRfvAmtBUOvPzaXo",
+    "tags" : [ "MWE3sxTi6FBt" ],
+    "description" : "H8OWugNRNYpF",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/UPSmgCK7MsjnABSiA"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/2Y6TSZdLNNoVDz6B"
       },
-      "doi" : "https://www.example.com/MvpnvD64xbFzi6q",
+      "doi" : "https://www.example.com/Zm4Wkd3tJPpWb",
       "publicationInstance" : {
         "type" : "JournalCorrigendum",
-        "volume" : "yzEDwJU50Tk6I7w8YYI",
-        "issue" : "ySSy3vFlqjQwlVFw",
-        "articleNumber" : "o4b2o948fPi",
+        "volume" : "qBD2w82L04jOFXH4",
+        "issue" : "FQuYu8mcSpQ4hpqWTCU",
+        "articleNumber" : "cCYOXPl0vTfZ7RTTD",
         "pages" : {
           "type" : "Range",
-          "begin" : "5b8xXFsFYi",
-          "end" : "7opVDZOs8Yj9UElB7k"
+          "begin" : "zgU51J3PTsaaad1nmUw",
+          "end" : "mWYNQAvXBojDxxYo"
         },
-        "corrigendumFor" : "https://www.example.com/P7p5twfkkfuDbG",
+        "corrigendumFor" : "https://www.example.com/44Lesto9GWi2UJJeB",
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/LfMWlqxAEe7l",
-    "abstract" : "RARvsoo5pdXNqx"
+    "metadataSource" : "https://www.example.com/ZGnCg7pA5Kha9q7a2",
+    "abstract" : "DFhYspnUU4a6NiogQr"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/ipsumofficiis",
-    "name" : "nG1TpeSODmrrEUF",
+    "id" : "https://www.example.org/aliquidanimi",
+    "name" : "vLE50j9r503b0rVIj",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "cXRvWE2neyjUZFD8",
-      "id" : "q4sPhENGYd7UtHq053"
+      "source" : "iTabuCBEzXW",
+      "id" : "Q9ZYn1d8hE"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "NMA",
-      "approvalStatus" : "REJECTION",
-      "applicationCode" : "qkKyEQWxteUa0a"
+      "approvalStatus" : "APPROVED",
+      "applicationCode" : "YF9yGiTmc9O"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -121,30 +121,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/magnimaiores" ],
+  "subjects" : [ "https://www.example.org/perspiciatisvoluptatibus" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "1dd721ec-8255-4042-9b7c-68c7a52e1a51",
-    "name" : "Fp3ZVLyjWU3bH",
-    "mimeType" : "fTRR28hLJkkP",
-    "size" : 1424982152,
+    "identifier" : "1a85a609-6d75-4114-9484-28b705f49ab6",
+    "name" : "xmyelLSelSj4",
+    "mimeType" : "pxeBacCrQ1Ea",
+    "size" : 319078110,
     "license" : {
       "type" : "License",
-      "identifier" : "SSZRK5kPR1DuAh",
+      "identifier" : "ljRZ7Qw7GS",
       "labels" : {
-        "hu" : "AhbUQQB2Tkf08DBA7MG"
+        "el" : "qPbxTHkKO3xiX5ov5"
       },
-      "link" : "https://www.example.com/sRMuT3qHXwD"
+      "link" : "https://www.example.com/xz8IzeZiTejSwBnXy1"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/JLiyfrCNfkEfJH",
-    "name" : "fA3sA3sYLqbhzI5",
-    "description" : "4x8CXj8e5Gr0sPVt"
+    "id" : "https://www.example.com/xaY8mIwt2MfgMhSvR",
+    "name" : "X8AjAwzBWQsb2ZiW",
+    "description" : "GT2s3mbj2HoWK5MRue"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
