+++ documentation/JournalInterview.json
@@ -1,118 +1,118 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
+  "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "f6Q3tss3WXpH9XywxK",
-    "ownerAffiliation" : "https://www.example.org/illumculpa"
+    "owner" : "lWfopsNdIHsH",
+    "ownerAffiliation" : "https://www.example.org/quaetemporibus"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/quiaaut",
+    "id" : "https://www.example.org/quissint",
     "labels" : {
-      "zh" : "LMM2FrHWqPSZajAF"
+      "it" : "dI2dQhcmwGYJHFlO"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/etdoloribus",
-  "doi" : "https://doi.org/10.1234/unde",
-  "link" : "https://www.example.org/officiisest",
+  "handle" : "https://www.example.org/sitrem",
+  "doi" : "https://doi.org/10.1234/enim",
+  "link" : "https://www.example.org/quamut",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "ZJR56jT3vXsvIR",
+    "mainTitle" : "Ilh1itx33mVvG",
     "alternativeTitles" : {
-      "sv" : "O6HlsJdjGioc8"
+      "hu" : "jIk9p5v1JujtW"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "2nI0cUv9TgMJT",
-      "month" : "wI0gylVj47rqJlm",
-      "day" : "Q4N5MCsKtwXpcoeB6p"
+      "year" : "dBhl0KTCBP4sW",
+      "month" : "2gTmMe5lpgwOypKn1",
+      "day" : "LrED6xXjQsGe4G"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/KdzqVWbDYDSr",
-        "name" : "Xi8nlFhcNgB63O1N",
-        "nameType" : "Personal",
-        "orcId" : "LYILD9GSGoaoS"
+        "id" : "https://www.example.com/1AUxzuEWG9",
+        "name" : "oGQTP4dZ4lGWyeq",
+        "nameType" : "Organizational",
+        "orcId" : "Hp57GcxUJeZHRQf65"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/zFnyAw5sIMLwFR",
+        "id" : "https://www.example.com/wa11SeH3Po0aesp9YJU",
         "labels" : {
-          "da" : "q4YmE0K1jMfNinGXP"
+          "fr" : "cb0d059x0Nuxq82"
         }
       } ],
-      "role" : "Photographer",
-      "sequence" : 3,
+      "role" : "TranslatorAdapter",
+      "sequence" : 9,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/4OnqzHFdfD3iINEC",
-        "name" : "Z8n4msr9an7C2",
-        "nameType" : "Organizational",
-        "orcId" : "IN9ypMXqelRJw6"
+        "id" : "https://www.example.com/NxRoKMjl69",
+        "name" : "m98V3XYzMJwYH1SGBY",
+        "nameType" : "Personal",
+        "orcId" : "mNrqmGa4rE70gAe"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/0sWUc1BJ7QKgUc",
+        "id" : "https://www.example.com/wHXon0gKjU",
         "labels" : {
-          "es" : "Z0fq6Vi0SkQQJM"
+          "sv" : "5jKaElaet4mg6k"
         }
       } ],
-      "role" : "Scenographer",
+      "role" : "ResearchGroup",
       "sequence" : 3,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "RUbdOWK6m5vk",
-    "tags" : [ "o9vXDKdlf3upZ" ],
-    "description" : "Xl6u8691hkUJMd",
+    "npiSubjectHeading" : "7AbPN3PUVL4uzXGPrWU",
+    "tags" : [ "EcwTyCoTbWgG" ],
+    "description" : "6oRDY5LbTuFRLTUi",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/PIeBXz1qA4aWWyiKR1"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/6AkcpiTy4UUp3okwd"
       },
-      "doi" : "https://www.example.com/MSyHy9tAHifAqgptv7",
+      "doi" : "https://www.example.com/Gj59fTqTbphW",
       "publicationInstance" : {
         "type" : "JournalInterview",
-        "volume" : "dm65HDth5aKyplyKN4",
-        "issue" : "Dz81FRWH3rJ",
-        "articleNumber" : "Xlh3UFNRWRGPn8ee2",
+        "volume" : "S9yLGcMZyBwwn",
+        "issue" : "HoC84GPicZk",
+        "articleNumber" : "TeC9sRpFIxHhX",
         "pages" : {
           "type" : "Range",
-          "begin" : "mH9LBpYynpA0o6",
-          "end" : "WOdWUoPT1pW"
+          "begin" : "icXssUYBEmSL8Fq",
+          "end" : "YK7PUNfClJbdPM5w1U"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/wi0NSsCJxU0iS",
-    "abstract" : "ytZ5pXqiKNMu9xy9S"
+    "metadataSource" : "https://www.example.com/5XfuBqBp150lqR3ds",
+    "abstract" : "hJyzfIjir8"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/magnamid",
-    "name" : "ll0IxlIO6dQ6BzN",
+    "id" : "https://www.example.org/idcommodi",
+    "name" : "vnfpkDPqvRTOp1I1",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "m1mFEX2JGqGH",
-      "id" : "weqkLCL4ks07d"
+      "source" : "HcwOReEhfRw",
+      "id" : "5fFfOqe96Xm1Oraemuc"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
-      "approvalStatus" : "REJECTION",
-      "applicationCode" : "5nmQxOOCaik"
+      "approvedBy" : "REK",
+      "approvalStatus" : "APPLIED",
+      "applicationCode" : "myffpUPlBTQMuWMC3mP"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -120,30 +120,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/erroromnis" ],
+  "subjects" : [ "https://www.example.org/etvoluptates" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "ad4513ea-ab55-4f70-b5d1-e5336ab9fdb6",
-    "name" : "EmNPSXKTrrb5",
-    "mimeType" : "cWVS4KgCzo9P",
-    "size" : 1971038745,
+    "identifier" : "6d1e2393-6bca-40c1-b9a4-c2b264517dc9",
+    "name" : "pkvpryZoyFQ36NWhmGR",
+    "mimeType" : "3vrjM3rl9l9T",
+    "size" : 1583210287,
     "license" : {
       "type" : "License",
-      "identifier" : "R21MvT9yxGqtOHfK6F",
+      "identifier" : "Z47vFuB6B7I4DSr2L",
       "labels" : {
-        "sv" : "VO7u9vXRpPAYxQANQ"
+        "af" : "0Mz2o6siqIhbMKO"
       },
-      "link" : "https://www.example.com/HyvWYTRel0OzLu"
+      "link" : "https://www.example.com/UAgYsQx0ciRtgSvXU"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/bC4HEZdNAG8Ldx63r7K",
-    "name" : "bp5T8Os0s9",
-    "description" : "VOj8R7wH7XRU0MODmk"
+    "id" : "https://www.example.com/PULvV2X9vr3EODxt",
+    "name" : "W4U1Yx6cOkN",
+    "description" : "Mw1NJfwy1ep"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
