+++ documentation/JournalIssue.json
@@ -1,118 +1,118 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
+  "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "7M5f2s1FDZfb",
-    "ownerAffiliation" : "https://www.example.org/incum"
+    "owner" : "LPeZ95JMOKyDT4yn4X",
+    "ownerAffiliation" : "https://www.example.org/facilisipsam"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/estimpedit",
+    "id" : "https://www.example.org/temporibusofficia",
     "labels" : {
-      "pt" : "RBHeeiiHrKG4vYxKF"
+      "fi" : "BghRE3OgzE3VrROe"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/omnisrepellat",
-  "doi" : "https://doi.org/10.1234/quos",
-  "link" : "https://www.example.org/eosdoloribus",
+  "handle" : "https://www.example.org/pariaturreprehenderit",
+  "doi" : "https://doi.org/10.1234/laborum",
+  "link" : "https://www.example.org/nostrumsuscipit",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "p540mWYNlyGqUGi1oJ",
+    "mainTitle" : "vgDkW3bB6tqEE9",
     "alternativeTitles" : {
-      "nl" : "9znzOjOfibG49hdEEs0"
+      "en" : "UtqlNJmP6ZUF6zI6v16"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "dRbPxF0tu5E62",
-      "month" : "ZpTkN3Bc7QlH5I5m",
-      "day" : "ceHQnLGgX946Rdgh"
+      "year" : "AQ3tUDNWykfJ",
+      "month" : "C5DfogUOQR8",
+      "day" : "GTSIhbx12IxULLR"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/MyjeWQ3M3vUBCHctynU",
-        "name" : "KteNEE6be8K",
+        "id" : "https://www.example.com/oj9AEodXYGoMJTw",
+        "name" : "HyxtQf4Ptk",
         "nameType" : "Organizational",
-        "orcId" : "7Qn52VkW6rGoUDKdLZ"
+        "orcId" : "7Xdy5E6KZUuq"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/B5JGlbzt7p",
+        "id" : "https://www.example.com/TEjx5uZiLQxLk",
         "labels" : {
-          "it" : "V8WFKHr2YW2j2"
+          "cs" : "rnKKyHMV91We8"
         }
       } ],
-      "role" : "Actor",
-      "sequence" : 3,
+      "role" : "Conductor",
+      "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/vyOExCw6IZ",
-        "name" : "WXHnjQWoRYc",
-        "nameType" : "Personal",
-        "orcId" : "ix4nudNXLbaGL"
+        "id" : "https://www.example.com/CGskbuLuv648QRAB",
+        "name" : "zG9OklryApeec8",
+        "nameType" : "Organizational",
+        "orcId" : "6YfE0fAnu5R7FTlR"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/R1AkOmINwxl",
+        "id" : "https://www.example.com/020UYM3q5JjMc",
         "labels" : {
-          "ca" : "SRo63rRWjhztXK"
+          "bg" : "3KZvengqHz"
         }
       } ],
-      "role" : "Funder",
+      "role" : "VfxSupervisor",
       "sequence" : 1,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "HARTbSjcTSXgBqA",
-    "tags" : [ "49pURbW3VXHBfM23lyb" ],
-    "description" : "Ias89fXgYTZ7",
+    "npiSubjectHeading" : "t4ZraNRdFJ7PIA1E",
+    "tags" : [ "3Nyqbw9bAx" ],
+    "description" : "zoSaOfSKgV4wbVo",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/vvzQdF3zNEnGh"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/CR1bUlHMgx13e"
       },
-      "doi" : "https://www.example.com/VBYBidwIpJ",
+      "doi" : "https://www.example.com/O8S4GE0kDv0afWq",
       "publicationInstance" : {
         "type" : "JournalIssue",
-        "volume" : "Hd1uTJsvBBsfTvMo",
-        "issue" : "jsu3EEkTyrbQV50",
-        "articleNumber" : "HfxGBPWo887l",
+        "volume" : "fwp2AWsDXujXZTnk2",
+        "issue" : "ygqQCS9UkublF3",
+        "articleNumber" : "tuNnl0Tx0solss",
         "pages" : {
           "type" : "Range",
-          "begin" : "yQgy82Je9TAQ",
-          "end" : "zEUkybQLTky"
+          "begin" : "FV88ajJrk6",
+          "end" : "JsQwz4LRE8XThx8eX"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/cJlhs9XkEUs",
-    "abstract" : "gaEoYyFriSO3B4KdTAc"
+    "metadataSource" : "https://www.example.com/jK9tN9LhtW",
+    "abstract" : "Gud9M4eOwcKDxWk"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/enimnecessitatibus",
-    "name" : "QjRuFkZ2ui",
+    "id" : "https://www.example.org/optioest",
+    "name" : "xdZRlTNHxwZ7Pzb2rP",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "D8b6kKtq8qexq0Ypb",
-      "id" : "phVMTCHAObeIa"
+      "source" : "eqBx3kjLEJ",
+      "id" : "7KlmkgvcrQC"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
+      "approvedBy" : "NMA",
       "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "QIUKUoiQvNxK"
+      "applicationCode" : "kgN9SfxKc2i6Q"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -120,30 +120,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/nobisipsam" ],
+  "subjects" : [ "https://www.example.org/magnireiciendis" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "06aa2244-b12c-4061-9323-dca1a1088737",
-    "name" : "HcL0ay5IdffYcvC0",
-    "mimeType" : "FcWu0uWi3KuH",
-    "size" : 606355600,
+    "identifier" : "feda7b68-aa48-4e9b-a7ee-d5314241288b",
+    "name" : "j1laytrOvdYJGx",
+    "mimeType" : "ElvRBKJAquw6dGu",
+    "size" : 1895176331,
     "license" : {
       "type" : "License",
-      "identifier" : "YjM93OiFUKq",
+      "identifier" : "mvmseQ0aBaU6ioro",
       "labels" : {
-        "da" : "PKJLNFknU1"
+        "el" : "bxQLHdznKzD"
       },
-      "link" : "https://www.example.com/4eSjvNruPc6b7Ia6BI"
+      "link" : "https://www.example.com/c5EPTMIK5FbZzs"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/NtzL6wK45IcmhjcFPOv",
-    "name" : "A6vAzIzQrepgNShjXOu",
-    "description" : "ECmi67AjGxwuxiyxhU"
+    "id" : "https://www.example.com/vsyB8NhL2Qy0oOt",
+    "name" : "2U0X6ISnAfdbf1",
+    "description" : "bhtc53BzenR1q2TQMHb"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
