+++ documentation/JournalLeader.json
@@ -1,118 +1,118 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "0lzz6H0vyKvFyn",
-    "ownerAffiliation" : "https://www.example.org/etnon"
+    "owner" : "bTbB8ZdCC3kz0WnXKYQ",
+    "ownerAffiliation" : "https://www.example.org/quiquis"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/earumaut",
+    "id" : "https://www.example.org/veroin",
     "labels" : {
-      "hu" : "jzNE6f1QzPr6"
+      "nn" : "wa9cxfqCmSKf"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/estdicta",
-  "doi" : "https://doi.org/10.1234/odio",
-  "link" : "https://www.example.org/consequunturlabore",
+  "handle" : "https://www.example.org/accusamusea",
+  "doi" : "https://doi.org/10.1234/ut",
+  "link" : "https://www.example.org/quaerataut",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "NqvgK3u5HG",
+    "mainTitle" : "YBy1VXdeKqfc8OlZPXf",
     "alternativeTitles" : {
-      "zh" : "c3ec0CkRIm3RUo0Vp"
+      "nl" : "HNBp03e4zEu69"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "2MtWRPSoaugvFMnHYp",
-      "month" : "0uE3RshhLjGx9",
-      "day" : "gr1ap89Sy40ej"
+      "year" : "fJytPBydw8LIge",
+      "month" : "U393xB0RDxdw",
+      "day" : "YEwVOO96QH0V"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/ZBfnVO8a4UY8a",
-        "name" : "qfBrsnZHVVmTDSiuDO",
-        "nameType" : "Organizational",
-        "orcId" : "IEkkHCH7P0y4vP"
+        "id" : "https://www.example.com/SHpjAh88dhvZe6iZZ",
+        "name" : "gEtHGQ9TuQ67",
+        "nameType" : "Personal",
+        "orcId" : "dA9Yfkj4IZblK6mTL"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/EkMmL5PzQp",
+        "id" : "https://www.example.com/jdxygCT722smn6LId",
         "labels" : {
-          "zh" : "7JqmKLk4Z0Sv9p4"
+          "da" : "0iJvmXY7bZ"
         }
       } ],
-      "role" : "HostingInstitution",
-      "sequence" : 5,
+      "role" : "Choreographer",
+      "sequence" : 9,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Zy1aCK44ahDsT6",
-        "name" : "5XdLQfvQgD3OMi5",
-        "nameType" : "Organizational",
-        "orcId" : "GQsei9m6hKs"
+        "id" : "https://www.example.com/OEz4g3YALk7OBf",
+        "name" : "GjMa6vaApEpDAp7",
+        "nameType" : "Personal",
+        "orcId" : "1UvcrKZFRm5"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/4uTyRJ7xN7PYiWKC",
+        "id" : "https://www.example.com/P61ZI81apn1DBg7",
         "labels" : {
-          "de" : "51Gnv7lOmOPufNDG"
+          "ca" : "xE8AA8f3mm0MNnfFzDz"
         }
       } ],
-      "role" : "RegistrationAgency",
-      "sequence" : 9,
+      "role" : "TranslatorAdapter",
+      "sequence" : 5,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "yo2iyLWuLH7Zz",
-    "tags" : [ "QNlHa9ksUHOsy" ],
-    "description" : "BkIOnqdiBObq",
+    "npiSubjectHeading" : "hj5mCaQ3mO7VS0auknh",
+    "tags" : [ "NlQNZBAmsdkoAXs0vL" ],
+    "description" : "LZbIRhNQsp",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/jOErLlsteU"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/ClAwxv276HQ5th3IjbD"
       },
-      "doi" : "https://www.example.com/bjN6bNksaT",
+      "doi" : "https://www.example.com/ykceLbTZnB3x",
       "publicationInstance" : {
         "type" : "JournalLeader",
-        "volume" : "QryqRCCQL8aexv",
-        "issue" : "YSKY3Xq3s1gWPKenwGb",
-        "articleNumber" : "BXS3gjq2ARFO2G6GvQ",
+        "volume" : "tUPSdU7grdFQ1cu",
+        "issue" : "E6LIR6bMFHhq",
+        "articleNumber" : "gLCAvUC7EBPsvp",
         "pages" : {
           "type" : "Range",
-          "begin" : "UP8T8q9s79Fi4LZT5",
-          "end" : "6XUfHGZQjWh4Zk"
+          "begin" : "UdYpG6ar3DjQ5RIrmyn",
+          "end" : "faXKs0KeGmgL"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/Jbih1WYal1N",
-    "abstract" : "iepGsefgw4"
+    "metadataSource" : "https://www.example.com/mVBbAiNKeRZNDseSO4",
+    "abstract" : "W8x5Jf5jTRvWNHTWd9"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/etnihil",
-    "name" : "uUXTfu8HLJA8TIp",
+    "id" : "https://www.example.org/quiscorporis",
+    "name" : "1rd24LISpDDPZGd",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "xavzrYeGhVd1gVhVfx",
-      "id" : "4nqdyTcix5q5Je4"
+      "source" : "iX31xFbLVWMvoZ1bU6A",
+      "id" : "I3fU5TWj4nTW"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "DECLINED",
-      "applicationCode" : "fxqATg1828G2LiyH"
+      "approvalStatus" : "APPROVED",
+      "applicationCode" : "WhBinmlgYJE2pwThO7"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -120,30 +120,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/etquo" ],
+  "subjects" : [ "https://www.example.org/laudantiumnobis" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "03b77445-bd55-4fee-b5b2-8cea78d10290",
-    "name" : "o8OKsbpfzNy",
-    "mimeType" : "HtKR3PSIFfe95",
-    "size" : 380080219,
+    "identifier" : "3ee1f9e7-7c93-436c-b8f1-30e7f3316aaf",
+    "name" : "xpdRsASzsQ24",
+    "mimeType" : "XNy3eZJavEKMegK",
+    "size" : 1073279324,
     "license" : {
       "type" : "License",
-      "identifier" : "jFN69AHOVDS",
+      "identifier" : "GiwiIarCY26",
       "labels" : {
-        "sv" : "NYZanS3V0gkj27"
+        "hu" : "EFzAoWNraY08JbC"
       },
-      "link" : "https://www.example.com/YRHHSIfVlbQgF"
+      "link" : "https://www.example.com/JW5vKIthSCf10RMW"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/d3AerCIYaaS211",
-    "name" : "StzpUdWGxJJnQ1Eb",
-    "description" : "Bcn6WukcQ1D8"
+    "id" : "https://www.example.com/lV2UGSoAEOXkE1YyF",
+    "name" : "q76Zi7g4i2KgPJZy",
+    "description" : "BXn08fUDJe0v6IRMUt"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
