+++ documentation/JournalLetter.json
@@ -1,118 +1,118 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "Lkd9ZVKGOe",
-    "ownerAffiliation" : "https://www.example.org/rerummodi"
+    "owner" : "jsmN7sNBaZk",
+    "ownerAffiliation" : "https://www.example.org/illofugiat"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/exearum",
+    "id" : "https://www.example.org/minusillum",
     "labels" : {
-      "af" : "0F1xqsj0YpC7B"
+      "da" : "X7LKqgD8sY0t7FORZcQ"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/nisilaudantium",
-  "doi" : "https://doi.org/10.1234/reiciendis",
-  "link" : "https://www.example.org/nullaquas",
+  "handle" : "https://www.example.org/perspiciatisadipisci",
+  "doi" : "https://doi.org/10.1234/ab",
+  "link" : "https://www.example.org/delenitiut",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "MeTTxrvnLZ",
+    "mainTitle" : "h9aoU17TNNLM8",
     "alternativeTitles" : {
-      "ru" : "7Gx9EbAkqQwatfcw"
+      "ca" : "qgte4VIZqj8P3WYex"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "FmGMZSFUX91FhP9Qs2x",
-      "month" : "lzMLcDQGp3tGZoAR",
-      "day" : "nNJtglLtnEbd"
+      "year" : "CyrkRAnMUGzawDd5",
+      "month" : "RyaGwIRCQ9m7BHRrI",
+      "day" : "4XruRKCXpdOZmUIayvS"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/xSmL9GygamA3tQN",
-        "name" : "LcbfvKyqDNFMUBLa",
+        "id" : "https://www.example.com/e7KYVja3DfYQhbVNSqi",
+        "name" : "P0WKWfqhqTsat",
         "nameType" : "Organizational",
-        "orcId" : "SAToTVH9LvzLnnWBp"
+        "orcId" : "RTMNwpsTNVL"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/e5kUiPp5TU",
+        "id" : "https://www.example.com/7TmT9d0EU2P33VF7R",
         "labels" : {
-          "hu" : "YkcVcJotIu8CdN"
+          "is" : "6jQU333yqHX3LzL8jAo"
         }
       } ],
-      "role" : "Composer",
-      "sequence" : 0,
+      "role" : "InterviewSubject",
+      "sequence" : 5,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/OT4HY9z16oZLC",
-        "name" : "13zKUhMk9URPa",
-        "nameType" : "Organizational",
-        "orcId" : "jxFcMpKNxIeXU"
+        "id" : "https://www.example.com/sLb74UCj4a7hPWTJ28t",
+        "name" : "K0FrD0mPajgKlxGJej",
+        "nameType" : "Personal",
+        "orcId" : "rC3a9EcwdBlsNgF"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/727dG0P5sjnJEBr9",
+        "id" : "https://www.example.com/yQFOZf4P0lDB",
         "labels" : {
-          "af" : "LhZwet7NyrfOzQBDX7"
+          "nb" : "1QicNrBgxEUelEj"
         }
       } ],
-      "role" : "Musician",
-      "sequence" : 5,
+      "role" : "Producer",
+      "sequence" : 2,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "ZWXkrvdn1hTPj",
-    "tags" : [ "BJSBc4aZZ9PeSQ6qD" ],
-    "description" : "cMw67bFVBtPRFW3q",
+    "npiSubjectHeading" : "WAkUhYnRUVGUUqUh",
+    "tags" : [ "dE81eKSnLIHY" ],
+    "description" : "YsvXJPy9WPXx",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/FQGowX74jgNiyPzZRIp"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/LEZlrYRgyus"
       },
-      "doi" : "https://www.example.com/A54MxrxNHJVeTuF4",
+      "doi" : "https://www.example.com/Wswnz4ZoJaTs",
       "publicationInstance" : {
         "type" : "JournalLetter",
-        "volume" : "5JtkMUfWMJjzQP7Ja",
-        "issue" : "PPTVlg2ptOgPJ",
-        "articleNumber" : "UUF1fg1WkUcdD",
+        "volume" : "DQASZz0ccXLAcl",
+        "issue" : "3ek6t8TAGCkRUbm9MZ",
+        "articleNumber" : "j08ECXqaK9m7MZP",
         "pages" : {
           "type" : "Range",
-          "begin" : "6buCAKBu7H2f",
-          "end" : "5SzKu2U0GYFXs3"
+          "begin" : "2lkQuz12T8o01jSyL",
+          "end" : "3NTWXImwexf1psUs"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/6hdnlwgdU9",
-    "abstract" : "EitLqp6YBc"
+    "metadataSource" : "https://www.example.com/jZfC0i5l6jO6agW9",
+    "abstract" : "xyYy2XO5l3ozf4FLxhu"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/providentreiciendis",
-    "name" : "04Vq78GrRDi",
+    "id" : "https://www.example.org/odittempora",
+    "name" : "VSyc4IgeBDJVRdImP",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "jhxLwqZZ3ecMtngT",
-      "id" : "GaIYjkoacgLFvSeG"
+      "source" : "QCClkodid4tne",
+      "id" : "hAU8NTjnUrqj"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "FkoTn8t6wZJ5Q45cYol"
+      "approvedBy" : "REK",
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "JGsvLT57dosuxS"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -120,30 +120,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/nonvelit" ],
+  "subjects" : [ "https://www.example.org/sitid" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "5cc8f974-b42b-401d-ae5a-756f873d463b",
-    "name" : "OfljcrHY5jTz",
-    "mimeType" : "1JqRn7gnBQXI",
-    "size" : 732057637,
+    "identifier" : "72ab4d57-95d9-445d-8dc1-06c20dc82a11",
+    "name" : "zLvlIJ796zwPL",
+    "mimeType" : "amLfgmp1cOn",
+    "size" : 95112026,
     "license" : {
       "type" : "License",
-      "identifier" : "uhy9KVymQRes",
+      "identifier" : "8SWANp3eEC0E8zQ",
       "labels" : {
-        "de" : "2tKlSQMAvDP0PtrK6"
+        "es" : "fucCbY1aJoOaSVeYx2B"
       },
-      "link" : "https://www.example.com/MjZKlVCOz3"
+      "link" : "https://www.example.com/HFNFglI6EFBdDICujT"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/3WNHWy7FaOF",
-    "name" : "n3m1BSdTnDocWwavm4",
-    "description" : "Ftw3Mq1xu5yvcODw"
+    "id" : "https://www.example.com/9rPQCWxzpK8OL3w",
+    "name" : "VEk08vjAN5rgBYC03ip",
+    "description" : "E2vT275zAlSI9M"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
