+++ documentation/JournalReview.json
@@ -1,118 +1,118 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "NEW",
+  "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "2yvu5tpfJLUYve",
-    "ownerAffiliation" : "https://www.example.org/laborumenim"
+    "owner" : "3oXuU7fU9UucJbyty",
+    "ownerAffiliation" : "https://www.example.org/adipiscidistinctio"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/nonmollitia",
+    "id" : "https://www.example.org/isteat",
     "labels" : {
-      "da" : "b9lgeBiYLwa5APnb"
+      "fr" : "aTya3Py2SrEydiQ"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/cupiditatereprehenderit",
-  "doi" : "https://doi.org/10.1234/aliquam",
-  "link" : "https://www.example.org/quisint",
+  "handle" : "https://www.example.org/aliquidaut",
+  "doi" : "https://doi.org/10.1234/temporibus",
+  "link" : "https://www.example.org/doloremnulla",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "Eq0dMeS1tIbUfA4xyZ",
+    "mainTitle" : "2GfCWdR75X",
     "alternativeTitles" : {
-      "ru" : "oB3fUPkinYX"
+      "sv" : "EWyHPF8heiIVDJuJCm"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "bimkSzKvIDpFdKJ6Qj",
-      "month" : "G0WJ4CBgxpQzksnD13",
-      "day" : "PN7wkzIUbv"
+      "year" : "iMp5G6Ecmo5",
+      "month" : "QxwJ1kCzlLLdGd1x",
+      "day" : "22GwfyahZMtx4Hk6Ds"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/kiet5dMh6z",
-        "name" : "dybVMLNmIRF1Kf4",
-        "nameType" : "Personal",
-        "orcId" : "TORpCsFkJ1qGVg"
+        "id" : "https://www.example.com/cDiQ1OcNyjgaOz9O1",
+        "name" : "3MTpmVt1iwzng6TB",
+        "nameType" : "Organizational",
+        "orcId" : "JXlWlF9P4x6VDRR"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/59PTXY2lQkNDIAA2khp",
+        "id" : "https://www.example.com/NHsx74xS2CLG",
         "labels" : {
-          "nb" : "yVvVuwBQefBV"
+          "es" : "824Eqxojd6"
         }
       } ],
-      "role" : "ArchitecturalPlanner",
-      "sequence" : 5,
+      "role" : "TranslatorAdapter",
+      "sequence" : 6,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/5t7TiUNmIwG",
-        "name" : "N4noKDHxQMM0i0yL",
+        "id" : "https://www.example.com/UJiGUdiE9tRbakX9J",
+        "name" : "e6mxMhEaXVssicKi",
         "nameType" : "Personal",
-        "orcId" : "VWC8RffqGcrjmSoBfy"
+        "orcId" : "IX9FhdGIHq7jBQesWwR"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/hQ01oJnJZNO4Yt855iy",
+        "id" : "https://www.example.com/4xp3THsBcErHqC",
         "labels" : {
-          "ru" : "SjJPfsj8fPYpmdnv"
+          "pl" : "sEopZx87pgZj"
         }
       } ],
-      "role" : "VideoEditor",
-      "sequence" : 0,
+      "role" : "Curator",
+      "sequence" : 4,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "1lwX4wr4rMLuxIytr",
-    "tags" : [ "Nwhwxy7lyU4uOnbcgS" ],
-    "description" : "IClCju0fKf1OSa",
+    "npiSubjectHeading" : "reV5iJxjoZlXvtgHw",
+    "tags" : [ "d7fj8CUNZEgB" ],
+    "description" : "fbTjOcTAbiPqpgo9Up",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Journal",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/iTlmcl03YHxfhd"
+        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/t3PTilccvxrk4XVA1"
       },
-      "doi" : "https://www.example.com/ygUyRQ7TjVW9YPzF4",
+      "doi" : "https://www.example.com/3useDvrNQgMKRXn4",
       "publicationInstance" : {
         "type" : "JournalReview",
-        "volume" : "h1VmhXOuIqLAB7Kamy2",
-        "issue" : "Apqa4n1z3CGuMT",
-        "articleNumber" : "VE5po0FK0zxh4p62Zpm",
+        "volume" : "Cx4RURqtWg",
+        "issue" : "SzhvstaAwLiNTH",
+        "articleNumber" : "mtxy6IikBmwCtG5J",
         "pages" : {
           "type" : "Range",
-          "begin" : "9QvvTqRr1P",
-          "end" : "5UY7GO4Zk8SwiKe3l"
+          "begin" : "UdQu969sx3",
+          "end" : "OH0pGwDrKDqFPWp"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/DfG1LahlaRrz5pOWls",
-    "abstract" : "4rl78zAOQBzZq7kDU"
+    "metadataSource" : "https://www.example.com/1PS2u45xiqmFEp13cuc",
+    "abstract" : "F9wb0KQOodS"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/insint",
-    "name" : "SZ7CZFIvMdhFqN",
+    "id" : "https://www.example.org/commodiipsam",
+    "name" : "juycUz1FS8JZeQc",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "7sRWZ8okSNPTlsncbsn",
-      "id" : "E4qfdN5esiBqO65"
+      "source" : "3BCOlVRQkwomvxUuh",
+      "id" : "HC8FKHfD8QxMKmcQhdb"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "oWlt5NePma"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "sTVUEVlSbOEX"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -120,30 +120,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/ipsumdoloremque" ],
+  "subjects" : [ "https://www.example.org/enimvoluptatem" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "956d1d29-c27d-4c07-a01d-ae2b87a70fbf",
-    "name" : "PPHrqgjCHx",
-    "mimeType" : "e92gp6Y5vDF1",
-    "size" : 304369929,
+    "identifier" : "94f167fe-5bab-45c7-a295-706bfb542621",
+    "name" : "RUBFExsN8Wd85",
+    "mimeType" : "wamO4ZG6ds1F0u4m",
+    "size" : 1142994541,
     "license" : {
       "type" : "License",
-      "identifier" : "f6WWJrDvOJ",
+      "identifier" : "XabKuhkVFl4ZmLgnC",
       "labels" : {
-        "en" : "ftgMGRJW54o6JeaWB"
+        "fr" : "0k2g3k5iK6EGM6"
       },
-      "link" : "https://www.example.com/An85gZFHEsvH"
+      "link" : "https://www.example.com/7auWwaZ6NcpU3YDVbJt"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/CersKdGCUsP69",
-    "name" : "ZGVELWsDBuIsuS7",
-    "description" : "WoUFj6N0gUk3m3035H"
+    "id" : "https://www.example.com/ItRuqgQuYFfiZYFYEZ6",
+    "name" : "WXE6WPmgSb",
+    "description" : "GV0cqRZv1B"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
