+++ documentation/Lecture.json
@@ -1,104 +1,105 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
+  "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "qMtA1gmV5N0z",
-    "ownerAffiliation" : "https://www.example.org/iureet"
+    "owner" : "CfCzvyf12mCldG8CEeo",
+    "ownerAffiliation" : "https://www.example.org/veniamet"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/numquamveritatis",
+    "id" : "https://www.example.org/essequibusdam",
     "labels" : {
-      "nn" : "z9mYWGFfhpOldXAa"
+      "fr" : "kOmlqTUKQ6"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/voluptatemqui",
-  "doi" : "https://doi.org/10.1234/voluptatem",
-  "link" : "https://www.example.org/quisest",
+  "handle" : "https://www.example.org/delectussapiente",
+  "doi" : "https://doi.org/10.1234/repudiandae",
+  "link" : "https://www.example.org/autemillo",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "V1l2SgccMSyWOurXsx",
+    "mainTitle" : "QAWMSjCHv20",
     "alternativeTitles" : {
-      "da" : "uQyrF2sCdGC"
+      "ca" : "QCzuL9etCuh5NKQqHh"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "eIKQHkAInEA4RjJ",
-      "month" : "7R45tATWP3ji",
-      "day" : "9weMiIN4EJD0TSPqg3"
+      "year" : "KCHdoLK1NClF",
+      "month" : "gbHPe4R1iRL",
+      "day" : "YtGazDSJslRoH1wauu"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/CDo1v3vJVu60",
-        "name" : "BDnOjhNLXaUCSaQZG",
-        "nameType" : "Personal",
-        "orcId" : "AsOGedSPQPi3K0Gxa"
+        "id" : "https://www.example.com/nR4QOT9Dg6H",
+        "name" : "pxHCfwEh7MlYzJoOaE",
+        "nameType" : "Organizational",
+        "orcId" : "201GomjHe6NOg7R1"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/1gFDJqWejAtZApVSWj",
+        "id" : "https://www.example.com/PhySV3sMKIPY9",
         "labels" : {
-          "el" : "dx5DJgfePHAeHzn"
+          "hu" : "9kzV3ezRojAXLUGtXv"
         }
       } ],
-      "role" : "Editor",
-      "sequence" : 5,
+      "role" : "CostumeDesigner",
+      "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/33tCrb7LbHo4lNPDr",
-        "name" : "qnSnxlvhnJPxP6UXUs",
-        "nameType" : "Personal",
-        "orcId" : "igNZVhLOGMdd"
+        "id" : "https://www.example.com/3Rj3GfCjZolu",
+        "name" : "xCGobq51QIIeL",
+        "nameType" : "Organizational",
+        "orcId" : "EhUilMYJMgW"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/pqHia3u3Gf",
+        "id" : "https://www.example.com/9gwrs3SydRREjTzBiNI",
         "labels" : {
-          "is" : "EXgPduFMXyZT9vAXaK"
+          "fr" : "uO5o3E5bYDxMz1q"
         }
       } ],
-      "role" : "Consultant",
-      "sequence" : 4,
+      "role" : "RegistrationAgency",
+      "sequence" : 0,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "r27N6h3lqcf",
-    "tags" : [ "A3raifkr6AflmFtcF9" ],
-    "description" : "r4H3qcQ4RjyCnK",
+    "npiSubjectHeading" : "7o7qKhVXI8bDeYN",
+    "tags" : [ "EiTW5qZldd7h" ],
+    "description" : "WRAVLmwb0NQn",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Event",
-        "label" : "CZ2CeD3SUS7qJxwv",
+        "label" : "4IVxAsy7wWCxFeca",
         "place" : {
           "type" : "UnconfirmedPlace",
-          "label" : "ExhZh9v4ix2apqeiU",
-          "country" : "kjo0FCZyoD"
+          "label" : "qJR63Kk4xrrieDJ83Jm",
+          "country" : "INMxWgl6CKWiU"
         },
         "time" : {
-          "type" : "Instant",
-          "value" : "2020-09-23T09:51:23.044996Z"
+          "type" : "Period",
+          "from" : "2020-09-23T09:51:23.044996Z",
+          "to" : "2020-09-23T09:51:23.044996Z"
         },
         "agent" : {
           "type" : "Organization",
-          "id" : "https://www.example.com/wpleZg5tS3I",
+          "id" : "https://www.example.com/ejMceLR4CeoG",
           "labels" : {
-            "pt" : "pUaHWiCwaz"
+            "el" : "fIEc5UzUI5Jq"
           }
         },
-        "product" : "https://www.example.com/qTcUhLe8dUJWOL59n4b"
+        "product" : "https://www.example.com/3rZNhLNheYs1SEBIsE6"
       },
-      "doi" : "https://www.example.com/1UHlBKcKuS0yFDTC",
+      "doi" : "https://www.example.com/eUZOPFsSewiOZpty0rI",
       "publicationInstance" : {
         "type" : "Lecture",
         "peerReviewed" : false,
@@ -107,24 +108,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/Z85wyamVofjghI",
-    "abstract" : "Tl2wBoKD65Qn"
+    "metadataSource" : "https://www.example.com/bbpcslxXNOs5aMyWa",
+    "abstract" : "DO3g8nZ6NVj"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/atmolestiae",
-    "name" : "ziFDhqsJJBZJpvt",
+    "id" : "https://www.example.org/commodieos",
+    "name" : "LcUch4vwfQU3FX5LQsV",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "qHRJrXwgxK",
-      "id" : "UNiNuhK6ijcnBruC"
+      "source" : "K2OVRnt9sBhdy7vOtN2",
+      "id" : "JGm0HhaaOvSooVB"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "vHv53ituVmjnLjdu"
+      "approvedBy" : "REK",
+      "approvalStatus" : "REJECTION",
+      "applicationCode" : "iStBLxdoSBmEYqsEOH"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -132,30 +133,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/quidemnon" ],
+  "subjects" : [ "https://www.example.org/utperferendis" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "f6d45bd5-b4cf-419b-8df6-cdb6f8b627c4",
-    "name" : "AAqrCfiW9bisBSyWy",
-    "mimeType" : "S2s6caTq3fScU",
-    "size" : 359061314,
+    "identifier" : "356d65a0-9771-46bf-bd45-f9a30db1e6ea",
+    "name" : "gfRl8MZRYAmSNj9wm",
+    "mimeType" : "DS1sMi8xcl",
+    "size" : 8800835,
     "license" : {
       "type" : "License",
-      "identifier" : "XWdINYonaq5",
+      "identifier" : "dnk96D8pQovkR2F",
       "labels" : {
-        "nl" : "1cq5rr2jRCi6R2T7n"
+        "ru" : "KQeBXVf8VTwMsmS"
       },
-      "link" : "https://www.example.com/mNIJcOL5uV9"
+      "link" : "https://www.example.com/huocZMXlYptI5eLky"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/ZQ92jbyACiKQ33H",
-    "name" : "7zaCLlXWjv6",
-    "description" : "JbwxPdGtC9aQyWTv"
+    "id" : "https://www.example.com/oA7j8jLTwELb9en",
+    "name" : "CxKLg5Ayo2",
+    "description" : "8FBAkcPBaAunswoWjF"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
