+++ documentation/Map.json
@@ -1,124 +1,124 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT",
+  "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "z1od2vW0qXdkuLRHzRG",
-    "ownerAffiliation" : "https://www.example.org/etquis"
+    "owner" : "6bRPDSVNd0Ro1WysTF1",
+    "ownerAffiliation" : "https://www.example.org/quishic"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/etin",
+    "id" : "https://www.example.org/rerumtemporibus",
     "labels" : {
-      "pt" : "0MVKVWyPTZPCx6wzU"
+      "da" : "6ykSZJ2U53JLH"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/velsaepe",
-  "doi" : "https://doi.org/10.1234/aut",
-  "link" : "https://www.example.org/quisdolorem",
+  "handle" : "https://www.example.org/sapientenihil",
+  "doi" : "https://doi.org/10.1234/suscipit",
+  "link" : "https://www.example.org/voluptatumodit",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "x2aSkKW76B94SZOqk1a",
+    "mainTitle" : "c999O3ZPdqsx1aVgIj",
     "alternativeTitles" : {
-      "ca" : "uyQjn36ZXypdDZm"
+      "el" : "bONZZ10Gvz"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "wnYEv9YCwX",
-      "month" : "dyXGZ8jCf3T",
-      "day" : "pdFL538FNVPcG"
+      "year" : "uiQrm4hkqasuL",
+      "month" : "rdxzrE9Ahx8OSgFT",
+      "day" : "IVOtVJV7gu2XpsOl1"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/rLazIM9dHcHiXkjjBt2",
-        "name" : "5G5Wr3unMeFgL",
-        "nameType" : "Organizational",
-        "orcId" : "b9edzvDLLvr"
+        "id" : "https://www.example.com/P1XhOvt3cTasF4xTcE1",
+        "name" : "Xc2u7RDg9FiF3TBzM5A",
+        "nameType" : "Personal",
+        "orcId" : "SbU4LWBfH284yYbzKs8"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/bGj2qY6eEKE",
+        "id" : "https://www.example.com/qfG0FWitdu",
         "labels" : {
-          "en" : "lMt80r3KlLBPNKv"
+          "ca" : "KgQAmv1b7q11"
         }
       } ],
-      "role" : "Illustrator",
-      "sequence" : 7,
+      "role" : "Advisor",
+      "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/MJsMXihpXLKn",
-        "name" : "LpS1jbunpR28Bml4",
-        "nameType" : "Personal",
-        "orcId" : "hR784Qr1kZPIf9xRd1S"
+        "id" : "https://www.example.com/wK3VHvw27Nfk2svn",
+        "name" : "jXpITfbaTPct",
+        "nameType" : "Organizational",
+        "orcId" : "hpn2r2oWQu7NsJ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/FU4QSefhNbCfKGzJwT",
+        "id" : "https://www.example.com/62KMFbymhuy",
         "labels" : {
-          "ca" : "D2EjQrP9yPcSTo3gdB"
+          "ru" : "n0G4H1Eoh1upmKGn"
         }
       } ],
-      "role" : "Curator",
-      "sequence" : 0,
+      "role" : "Artist",
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "OjiYGu67xiyx",
-    "tags" : [ "VBqMWJ760lnenZ0" ],
-    "description" : "XiiDbJo7RiTfhc3",
+    "npiSubjectHeading" : "NHpqncrXMI00KcLMb7y",
+    "tags" : [ "Nm3g4mbcCxE6j" ],
+    "description" : "aD0mg3noEv8HmvWtFE",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "GeographicalContent",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/cQ7ZazAsbtzeP3y"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/mdFP35rT9lLDMdurMI"
         }
       },
-      "doi" : "https://www.example.com/GkFLHTADeticjP",
+      "doi" : "https://www.example.com/pnrveiNTk4tyrczN",
       "publicationInstance" : {
         "type" : "Map",
-        "description" : "hetqW2SkGYKJyo",
+        "description" : "ktVV6wpGMETk0H",
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "m8m5OZArsZVxTy0",
-            "end" : "fx7JtGUXkT86pI"
+            "begin" : "ZxcL9XD7ZUMxBMYV",
+            "end" : "pHbzDe5WwxJp58mk"
           },
-          "pages" : "0P2zpChs3a",
+          "pages" : "ra3bYpZFwA4",
           "illustrated" : true
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/Hu2VgKiIc1UFiN",
-    "abstract" : "mAyCJwctrY"
+    "metadataSource" : "https://www.example.com/eD9E9aOqOvmRO",
+    "abstract" : "c6q7nxo8Nk"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/explicabodebitis",
-    "name" : "EFXU2EQzJ21B1UO",
+    "id" : "https://www.example.org/accusantiumexplicabo",
+    "name" : "EauXdh5D7cDd",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "frUjrzbk7fC",
-      "id" : "c95CnK33B31r6"
+      "source" : "bNqqLI4RgE6H",
+      "id" : "XDUsG0yIpTzEOhB"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "T0ch6GbSydD69ETQ"
+      "approvedBy" : "REK",
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "LGFKYRO3x82znVFI4z6"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -126,30 +126,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/eosfacilis" ],
+  "subjects" : [ "https://www.example.org/doloremquasi" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "0d047b4a-9765-41e6-9403-d6b450c74d67",
-    "name" : "XMAHsDBsQXSQYgRF",
-    "mimeType" : "ux2BNno9lkNu8q8xvxa",
-    "size" : 2064446084,
+    "identifier" : "1f4f8aa7-a71c-41df-aea2-e40f1a5629a0",
+    "name" : "MPYC8g6Bw6D4em2I",
+    "mimeType" : "xQmF7jQMXBeM4xW2",
+    "size" : 1307297691,
     "license" : {
       "type" : "License",
-      "identifier" : "QjzQ1mvx0VfMETtvZX",
+      "identifier" : "w5vPJFmxYgO0wUa",
       "labels" : {
-        "ca" : "7tXDQQ510IO"
+        "af" : "2z7ZBkruX1Zso9vluJ"
       },
-      "link" : "https://www.example.com/SYUVi0A7wyyivzuDEyZ"
+      "link" : "https://www.example.com/wmmHHpZTose6fgDx1"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/FXsJKEHtArCmuAt",
-    "name" : "NUkfEW5Y6L",
-    "description" : "jgcqsSkk3KJzA"
+    "id" : "https://www.example.com/lONZqWD5OdM0c",
+    "name" : "KjuTAaSU04y",
+    "description" : "tPXWUqM0RsGh"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
