+++ documentation/MediaBlogPost.json
@@ -1,97 +1,96 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
+  "status" : "DRAFT",
   "resourceOwner" : {
-    "owner" : "1AJXWwNm15hCAkXO",
-    "ownerAffiliation" : "https://www.example.org/rerumminima"
+    "owner" : "dynRIFRaIk7ajzg",
+    "ownerAffiliation" : "https://www.example.org/minusa"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/laboreodio",
+    "id" : "https://www.example.org/sequiet",
     "labels" : {
-      "da" : "TvVLKyz5xKTneFqI"
+      "ru" : "cq0voTOPmWYAUn59ReB"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/nonrerum",
-  "doi" : "https://doi.org/10.1234/similique",
-  "link" : "https://www.example.org/quoenim",
+  "handle" : "https://www.example.org/eumhic",
+  "doi" : "https://doi.org/10.1234/eaque",
+  "link" : "https://www.example.org/suntamet",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "3EZBn6XeevZO3P",
+    "mainTitle" : "5kc8qV7vUyq0igde",
     "alternativeTitles" : {
-      "zh" : "fus1Y4LqrE"
+      "da" : "EuHaOIdFBXF1"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "xeMEK24Phzw1Kcd",
-      "month" : "QnL8tDZNGuadosmHgV",
-      "day" : "A7zf6MoeciM"
+      "year" : "Ly4BLqReEGyU4ZoFsu",
+      "month" : "xK3z1pXYhY4w7JHxKi",
+      "day" : "uvbM6n7ifYHzfSG"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/VFteEbObTOBBOHJ",
-        "name" : "d3sBobBuWOU6GCrJC2",
-        "nameType" : "Organizational",
-        "orcId" : "cDlZXeLtGEsFW8Fl"
+        "id" : "https://www.example.com/XHJ1HeN6HYb0Mmgg",
+        "name" : "sVOz6wA6fRajN",
+        "nameType" : "Personal",
+        "orcId" : "qkRu0jcZFuX8eq"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/V1SrXDgjeV4f",
+        "id" : "https://www.example.com/Z9X0TUkD7IJ",
         "labels" : {
-          "is" : "uhktZMLI55t3eh"
+          "sv" : "dr65L22wwHxaDBsj4nz"
         }
       } ],
-      "role" : "ProjectMember",
-      "sequence" : 6,
+      "role" : "VfxSupervisor",
+      "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/a8ZHClubH2exuFnuC",
-        "name" : "zQg7moRFZb2M",
-        "nameType" : "Personal",
-        "orcId" : "I6hYSk45dWs"
+        "id" : "https://www.example.com/rSzJpJaik8UGVW",
+        "name" : "fiH64qjGWu",
+        "nameType" : "Organizational",
+        "orcId" : "NV4fuVpkOAeS"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/z05NLS4SR45ZsEwcV",
+        "id" : "https://www.example.com/o4m4OldHJyH8",
         "labels" : {
-          "se" : "WEIStesPC4q"
+          "zh" : "IFFMj6r1E2o7ftUya"
         }
       } ],
-      "role" : "Librettist",
+      "role" : "WorkPackageLeader",
       "sequence" : 9,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "bm6Ql3SmJO7aF54mMd",
-    "tags" : [ "b5ksfvP0NoYNh0" ],
-    "description" : "jM20SqUuhnCNMxz",
+    "npiSubjectHeading" : "K6SIIvyhd6hREu",
+    "tags" : [ "Z4hpoL22etC" ],
+    "description" : "AcDzumFTiJ4",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "MediaContribution",
         "medium" : {
-          "type" : "Other",
-          "description" : "gmbuBxbMJfOcbBNNzF"
+          "type" : "Radio"
         },
-        "format" : "Sound",
-        "disseminationChannel" : "gUjYqgSExo",
+        "format" : "Text",
+        "disseminationChannel" : "xoXH6dLXfNLWN",
         "partOf" : {
           "type" : "SeriesEpisode",
-          "series" : "EVScxuJRxuQBygeT",
-          "seriesPart" : "KjXdInvIky7t2y2hIJK"
+          "series" : "38dpaqPupo",
+          "seriesPart" : "lBGCQtg1VdCOoiuZlNc"
         }
       },
-      "doi" : "https://www.example.com/cGsUv8FoCTX5MS0RTSo",
+      "doi" : "https://www.example.com/0mLcvQtQCJBsLV",
       "publicationInstance" : {
         "type" : "MediaBlogPost",
         "peerReviewed" : false,
@@ -100,24 +99,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/hQhLMuriacISukz",
-    "abstract" : "FK2ty7uJQyqq9FsjiH"
+    "metadataSource" : "https://www.example.com/RGsSX56iU7Q",
+    "abstract" : "qydOjtj6NfDy1IstB"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/illumveritatis",
-    "name" : "cesAdMYVO3mBseC2",
+    "id" : "https://www.example.org/etconsequatur",
+    "name" : "OHSqvrqNPYAPI",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "iiOvbtvgRjlX9",
-      "id" : "QY2llLbk4kckuu8p"
+      "source" : "bv0OYsEHn2Cpa48aDVD",
+      "id" : "40v0PECDHJn"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "APPROVED",
-      "applicationCode" : "tGlU7Bnk9UqbV"
+      "approvedBy" : "REK",
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "meqLCQdQClKF"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -125,30 +124,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/sequinemo" ],
+  "subjects" : [ "https://www.example.org/molestiaedeserunt" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "5cd28581-7cd5-46db-a0b9-438fd6d9d931",
-    "name" : "iFCQ6gAx51S5byEN",
-    "mimeType" : "JUyF3AsEbqPBhHF",
-    "size" : 1938008990,
+    "identifier" : "2629babd-5834-4880-a2b3-bf37984b111b",
+    "name" : "hljoyf8IRwjXeQdm",
+    "mimeType" : "wvlsHF2euU7STK2je",
+    "size" : 759154240,
     "license" : {
       "type" : "License",
-      "identifier" : "KOyeuD6o0KhFg4bPq",
+      "identifier" : "Wx6wTpoRfxJeycJLKi3",
       "labels" : {
-        "de" : "zhhkTwOB5GqKGr42"
+        "se" : "jZV3rDHM9ht5u"
       },
-      "link" : "https://www.example.com/GwifakmvsGjL2i"
+      "link" : "https://www.example.com/HeED7dcUK8zGzX"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/1I4neE2eoVOwfYv",
-    "name" : "EnryEIVAw01WSNeFfAJ",
-    "description" : "leReip1XvDck"
+    "id" : "https://www.example.com/NZEdgfNjRdE3YiF3y",
+    "name" : "Ri9UDgw1xT0OPtvv3",
+    "description" : "iwaZj2zkQpHNm"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
