+++ documentation/MediaFeatureArticle.json
@@ -1,120 +1,120 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "HB7RtOaNuPiF3mK",
-    "ownerAffiliation" : "https://www.example.org/suscipitut"
+    "owner" : "vNEzqrOmdoRSbn",
+    "ownerAffiliation" : "https://www.example.org/voluptaset"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/errordolorum",
+    "id" : "https://www.example.org/nequenumquam",
     "labels" : {
-      "pl" : "XkLOEYnDgSnz8mIN"
+      "nn" : "hF8E1GQGYYEz"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/vitaevero",
-  "doi" : "https://doi.org/10.1234/aut",
-  "link" : "https://www.example.org/voluptasquo",
+  "handle" : "https://www.example.org/eumillum",
+  "doi" : "https://doi.org/10.1234/fugiat",
+  "link" : "https://www.example.org/liberoquibusdam",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "TEwuD9kh8d3A4nC1QCm",
+    "mainTitle" : "P3S4WUQra9vSIYZ",
     "alternativeTitles" : {
-      "zh" : "Cdox0nbnEYPgmNJqt"
+      "es" : "rODYm8kINXq"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "rDbr0gjKf1CM7V",
-      "month" : "oFAWOzAHiVydlVZHjPl",
-      "day" : "fQ5Wi1UixnoGr4qo"
+      "year" : "GNGUz1EZWvq",
+      "month" : "t1zZZITUlxiF8",
+      "day" : "BWoAsX5LFGjKM"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/f3enQTQBNnXKSJB",
-        "name" : "tXmkdnDtHzHDs",
+        "id" : "https://www.example.com/QIrBaAnElXW6",
+        "name" : "bnUMezJGKBYnkHFT7N",
         "nameType" : "Personal",
-        "orcId" : "WeW1oTIw7M"
+        "orcId" : "JXFHLBwYM4cfK1O"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/gCvCJ437OVqa1fD",
+        "id" : "https://www.example.com/bpL0qthpdbDnuVz",
         "labels" : {
-          "af" : "CWENDnhL2j"
+          "pl" : "OtMJcizd1JksAndGUM"
         }
       } ],
-      "role" : "ResearchGroup",
-      "sequence" : 7,
+      "role" : "Curator",
+      "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/KDWPF7slQXPdNTo1rg",
-        "name" : "VbMhuMTu0z6",
+        "id" : "https://www.example.com/EjRT6l6cba",
+        "name" : "S3zv3LFYONLmj",
         "nameType" : "Personal",
-        "orcId" : "jF3pfwjMpeIa8D"
+        "orcId" : "XjmGSZNTj2CxtM0"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/qKGuZj40blph",
+        "id" : "https://www.example.com/qriLJ0fYV3OSe",
         "labels" : {
-          "ca" : "Ab8YgS2pzcjCLR"
+          "sv" : "kDfHg53SRz"
         }
       } ],
-      "role" : "RelatedPerson",
-      "sequence" : 7,
+      "role" : "ArtisticDirector",
+      "sequence" : 6,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "EtQ2HkZ7ihXy7WerBr",
-    "tags" : [ "zeG7lopsdtK" ],
-    "description" : "OWSUPhzhNB",
+    "npiSubjectHeading" : "gcahdHSLQYECU6fSt",
+    "tags" : [ "CSc5k1BRasudSrLb8" ],
+    "description" : "2sAxVX1oE9SCBz4",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "UnconfirmedMediaContributionPeriodical",
-        "title" : "pYZpMmAdQflNYGMP",
-        "printIssn" : "4053-8656",
-        "onlineIssn" : "7974-6594"
+        "title" : "OmAitg0YB6QfOT",
+        "printIssn" : "7190-0551",
+        "onlineIssn" : "4911-3704"
       },
-      "doi" : "https://www.example.com/6XWHfUayu6AZFb",
+      "doi" : "https://www.example.com/uggMDFxQw8dRpe0",
       "publicationInstance" : {
         "type" : "MediaFeatureArticle",
-        "volume" : "LeYVo1gBNRrZpQW1u",
-        "issue" : "bnmUzvzpCG0zt3",
-        "articleNumber" : "JWMkOQ1Yi0",
+        "volume" : "SeOM5VfVvF",
+        "issue" : "GymU9Eq0eGOU",
+        "articleNumber" : "Kq5O8FRzyZ6xD",
         "pages" : {
           "type" : "Range",
-          "begin" : "Cvhm8oWg8QObe4",
-          "end" : "yQmu0TaOY4"
+          "begin" : "bRet0rwoc5zE07yzz",
+          "end" : "uu1LPNN81k"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/qPAzOqmOD0uO8KW",
-    "abstract" : "QJXgLWdXbdFk5lR9rf"
+    "metadataSource" : "https://www.example.com/PhwtvxlQiU10UW",
+    "abstract" : "WdrnwtTw77dJpDmIli"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/velitrem",
-    "name" : "mLhhvkNJ0Mkj8ZjZE9z",
+    "id" : "https://www.example.org/rerumet",
+    "name" : "ACcimPxkuwVUu",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "JH6AC1Z15dyY2f4mG",
-      "id" : "SYxYnbg2nz0k9hzsvVn"
+      "source" : "blesySWR27yrWPl",
+      "id" : "nkIBoZlVpyDJU2y6DsL"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
-      "approvalStatus" : "APPROVED",
-      "applicationCode" : "Cz5fD9hNpABpQfxG"
+      "approvedBy" : "REK",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "7cVzEWseQPuF9o"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -122,30 +122,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/hicin" ],
+  "subjects" : [ "https://www.example.org/distinctiout" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "9c1be97a-7b7c-45e7-a0ff-cc6a09c23b9c",
-    "name" : "5eDOBdISSDdzvAB3giq",
-    "mimeType" : "XmLqPxIKmjItg",
-    "size" : 2060547542,
+    "identifier" : "43b43e48-2184-45f7-a3e8-d775952fedb7",
+    "name" : "QNBPjQHyh3x06os",
+    "mimeType" : "jI0vXlmv2P",
+    "size" : 1199199236,
     "license" : {
       "type" : "License",
-      "identifier" : "yf1tEHTfOxq",
+      "identifier" : "2GVKDwp6vzyqjKTdF7",
       "labels" : {
-        "de" : "MBIzZaQ8Glq"
+        "de" : "fLFouXOnKJIlFZeXy"
       },
-      "link" : "https://www.example.com/JYJFauJdgfz"
+      "link" : "https://www.example.com/Pkujjx6cs7KJOuv"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/v0nAd2D5ohc0HLa",
-    "name" : "88hjrVt9Bs",
-    "description" : "BNUB3v1S7fv"
+    "id" : "https://www.example.com/8lfaZo23LQ9",
+    "name" : "tKrjWFTqdHfCouucUro",
+    "description" : "eRp4jogFAHp"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
