+++ documentation/MediaInterview.json
@@ -1,96 +1,97 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "NEW",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "OH1K9Pcjmmqvbhu3Z",
-    "ownerAffiliation" : "https://www.example.org/velitaut"
+    "owner" : "wHrumCP2qhmOzyM",
+    "ownerAffiliation" : "https://www.example.org/istequia"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/quaerepellendus",
+    "id" : "https://www.example.org/quisomnis",
     "labels" : {
-      "nb" : "n1NgyGi3nG3"
+      "el" : "al8LIvEB5D"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/rerummolestiae",
-  "doi" : "https://doi.org/10.1234/voluptatem",
-  "link" : "https://www.example.org/itaqueautem",
+  "handle" : "https://www.example.org/recusandaemagnam",
+  "doi" : "https://doi.org/10.1234/perferendis",
+  "link" : "https://www.example.org/idplaceat",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "IgCBgXi5Mx",
+    "mainTitle" : "sCexlEafyma8lK",
     "alternativeTitles" : {
-      "zh" : "CMtUzH8msz7JAMPo"
+      "nb" : "Kd3Rovvak8EV9MamPA"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "ef2ysfFrRs7Q",
-      "month" : "ROViVRvEvwZER6Gj",
-      "day" : "xJmbqjpIuB"
+      "year" : "w2Op0kQHyn6Fs",
+      "month" : "m6LkfsnHPgjsAG",
+      "day" : "1AoLTGzGYM8R"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/P8jk3zgwTmToU",
-        "name" : "9EaKuASwSG5p",
-        "nameType" : "Organizational",
-        "orcId" : "bMGwXYa4TKz"
+        "id" : "https://www.example.com/01FYTI5bKPUA",
+        "name" : "cA8jvklVrgb",
+        "nameType" : "Personal",
+        "orcId" : "BIVEUwnR5QI2c7KP"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/4IkyT1PJXDSR",
+        "id" : "https://www.example.com/ciRADyX0tWtzrgVNEIx",
         "labels" : {
-          "af" : "FiPNpYMwHvWw0"
+          "nn" : "GVYNiUquhCVU5chhLg"
         }
       } ],
-      "role" : "DataCurator",
-      "sequence" : 1,
+      "role" : "Musician",
+      "sequence" : 9,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/r90BPRRgH9MFiQbQO",
-        "name" : "EY7WlFvlYXDeS",
-        "nameType" : "Personal",
-        "orcId" : "eQWX3CtKPqKjgxfzQ"
+        "id" : "https://www.example.com/kJF1C2ecGP",
+        "name" : "HF10pEv6LGe",
+        "nameType" : "Organizational",
+        "orcId" : "wgHMoRLGIqFlcw1m8"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/FeZSPpbe0cza",
+        "id" : "https://www.example.com/AoRv3wzzHRSMQ",
         "labels" : {
-          "cs" : "c8A6sSKzr5WJ7u"
+          "ca" : "9VNOqAnDV8Qidn9zKX"
         }
       } ],
-      "role" : "Supervisor",
-      "sequence" : 1,
+      "role" : "Advisor",
+      "sequence" : 3,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "jQIi3ETAVm",
-    "tags" : [ "WanmsTyQ7OZMxfkCCU1" ],
-    "description" : "fWnYq4ATYqB5PAjt",
+    "npiSubjectHeading" : "ofykaSlx3pqy",
+    "tags" : [ "dUGmWMQKhP" ],
+    "description" : "Cb2TAvxantfgEIsVi",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "MediaContribution",
         "medium" : {
-          "type" : "Journal"
+          "type" : "Other",
+          "description" : "Fa3ZboMerb"
         },
-        "format" : "Sound",
-        "disseminationChannel" : "i5GU9ifHZwIWst",
+        "format" : "Text",
+        "disseminationChannel" : "yN1WHxNNjg6njx5WpP",
         "partOf" : {
           "type" : "SeriesEpisode",
-          "series" : "cr25tBrGoVvNg99A",
-          "seriesPart" : "lBvcpi0XNlNAb1MZt0M"
+          "series" : "ufrsCkyDYzy3955",
+          "seriesPart" : "XiA51FBECu63O"
         }
       },
-      "doi" : "https://www.example.com/gXxWO9cNApZvzW7mI",
+      "doi" : "https://www.example.com/JnDSFwwmHfuN",
       "publicationInstance" : {
         "type" : "MediaInterview",
         "peerReviewed" : false,
@@ -99,24 +100,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/Cc31qmHpVZ",
-    "abstract" : "615wFyWMLL6Z"
+    "metadataSource" : "https://www.example.com/mGSTAXxwV5",
+    "abstract" : "FOFoUOMRO0K"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/doloreipsum",
-    "name" : "ctTikgqUCuAdgyEpMph",
+    "id" : "https://www.example.org/eaaccusamus",
+    "name" : "6isdv6Zkof",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "GwRx4tv8DsG7ee",
-      "id" : "2N9X782amcaz5XBDY"
+      "source" : "fspps6ElSXNLq8X3",
+      "id" : "wEgsruiHiNT31WhME"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
+      "approvedBy" : "NMA",
       "approvalStatus" : "APPLIED",
-      "applicationCode" : "TiMwtF5a9Y5"
+      "applicationCode" : "T9YdRT3OEoYAD8OYDy"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -124,30 +125,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/rerumvoluptatibus" ],
+  "subjects" : [ "https://www.example.org/nemoqui" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "d7389912-38ef-4582-be73-bd48de97b2ee",
-    "name" : "kYQfE8YOTKud0whLuz",
-    "mimeType" : "reFQnKTzHXuwUuC",
-    "size" : 681401021,
+    "identifier" : "10dc9f70-2eec-4b83-aeac-70bf0b2a33ee",
+    "name" : "Ga2KZ5ezLPB2A9",
+    "mimeType" : "0srltQikcfR4",
+    "size" : 700726391,
     "license" : {
       "type" : "License",
-      "identifier" : "Zb2qxqW9ZPi1dYk",
+      "identifier" : "hCsQDr1t7A",
       "labels" : {
-        "nn" : "xTDhUycKpuztaLw"
+        "hu" : "FJUHxmjADvNRUCq91u"
       },
-      "link" : "https://www.example.com/TF9occo2k04x"
+      "link" : "https://www.example.com/AMKn5V6OZuvknH2sj2v"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/j5xQIjCdUeRvHIwq",
-    "name" : "7vk7TxYpp9MF9f4No5B",
-    "description" : "ITjk8P8bt3OFlOqi4E"
+    "id" : "https://www.example.com/qe0Y7tkvmznvvw5bg2m",
+    "name" : "3b8du74pLV",
+    "description" : "8bLgZl2i2hH"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
