+++ documentation/MediaParticipationInRadioOrTv.json
@@ -3,94 +3,94 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "B1lA0moLei9",
-    "ownerAffiliation" : "https://www.example.org/quamcorporis"
+    "owner" : "kjYGI1RRtF8cIw",
+    "ownerAffiliation" : "https://www.example.org/quisquamsint"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/perspiciatisaliquid",
+    "id" : "https://www.example.org/officiaratione",
     "labels" : {
-      "nl" : "gzKDzANduVX"
+      "sv" : "igiUkpl8VnKnTryI"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/veliste",
-  "doi" : "https://doi.org/10.1234/sed",
-  "link" : "https://www.example.org/etsunt",
+  "handle" : "https://www.example.org/estmolestias",
+  "doi" : "https://doi.org/10.1234/ut",
+  "link" : "https://www.example.org/solutaaut",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "By9ijbudFTEmjh",
+    "mainTitle" : "6qDqPK2wG7u65",
     "alternativeTitles" : {
-      "de" : "dxn8fERVmIU"
+      "se" : "c084DXU0o8gK5CM"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "YITmW1GUhZv4uve",
-      "month" : "JFj1ibrhpWTp5x",
-      "day" : "BisvINafx5U1ebZl6"
+      "year" : "QfNKujLuZmnaLwvv",
+      "month" : "C5wnflaeYtQl4eW",
+      "day" : "QoqLbpkLe2EOhMt"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/jgRmDc18hA9za",
-        "name" : "uG99HoWwXMLqxE",
-        "nameType" : "Personal",
-        "orcId" : "Bf9eaTwIoTzxGAO32"
+        "id" : "https://www.example.com/T6I4mfa8gGP0uYTmb",
+        "name" : "Fd0D1o9va4J",
+        "nameType" : "Organizational",
+        "orcId" : "s3ibUJghECu15z"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/mKlcQBGeveTAc7i",
+        "id" : "https://www.example.com/9IOSwZzKAs",
         "labels" : {
-          "fr" : "XsLzjLRMNwa3Prr2"
+          "nn" : "cWj7ceEjHgI"
         }
       } ],
-      "role" : "Producer",
-      "sequence" : 4,
+      "role" : "Designer",
+      "sequence" : 5,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/HJko6SnvAf3UC7cBoa",
-        "name" : "eDLuLu5sOFb7ak",
-        "nameType" : "Personal",
-        "orcId" : "QvJNcGEhlCbTmx0"
+        "id" : "https://www.example.com/wkDemY04DtuAt",
+        "name" : "KCHiRoN2GlLY",
+        "nameType" : "Organizational",
+        "orcId" : "rjtx1olCKsYWHBC6T"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/HOj0bqWpC1hihm",
+        "id" : "https://www.example.com/z7qhphNrQfTv",
         "labels" : {
-          "af" : "kPuVy8PMADpFbNhHB3i"
+          "fr" : "1kuM1Qi9Q1ChvPd"
         }
       } ],
-      "role" : "Researcher",
-      "sequence" : 7,
+      "role" : "SoundDesigner",
+      "sequence" : 8,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "4GQAmK49Kt4CkEa",
-    "tags" : [ "sC66akclxeQq9Scq" ],
-    "description" : "aY3fEG729ydgD5zi",
+    "npiSubjectHeading" : "9eaeK62m6qB",
+    "tags" : [ "lAoiznRa9A7TfrQ" ],
+    "description" : "vBr9xL8oWv0Lr",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "MediaContribution",
         "medium" : {
-          "type" : "Internet"
+          "type" : "TV"
         },
-        "format" : "Video",
-        "disseminationChannel" : "eiE4eyO7EWtR",
+        "format" : "Sound",
+        "disseminationChannel" : "6AYFPi1J3j",
         "partOf" : {
           "type" : "SeriesEpisode",
-          "series" : "7Rs1h2lklbVdIVD1",
-          "seriesPart" : "TI7YXt7Pa4vKXM6u"
+          "series" : "HokBXl6E3AthKHumag",
+          "seriesPart" : "v2CqEaUtpNglTr631"
         }
       },
-      "doi" : "https://www.example.com/h3MJuoezjq",
+      "doi" : "https://www.example.com/dJQzde8glI",
       "publicationInstance" : {
         "type" : "MediaParticipationInRadioOrTv",
         "peerReviewed" : false,
@@ -99,24 +99,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/GY1oYE4Z9UsTXv",
-    "abstract" : "QwuKEKWpAr187jw2"
+    "metadataSource" : "https://www.example.com/bSGxCxzmVRy6J0UidCc",
+    "abstract" : "mBJ4kO1MbUQ"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/laboriosamut",
-    "name" : "Whx5qSG8FOOtIYbf",
+    "id" : "https://www.example.org/dignissimosmollitia",
+    "name" : "DXhlsznY8aHhj",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "c7DWHtyqHPsrKOO",
-      "id" : "QskOCow7Mxowfhf"
+      "source" : "AThJPrxGP59IRzncclk",
+      "id" : "71qLxB5mKz"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "DECLINED",
-      "applicationCode" : "cqDwkFYsEoAo"
+      "approvedBy" : "NMA",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "Sa5uwtpazWmTjEpRS"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -124,30 +124,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/etcupiditate" ],
+  "subjects" : [ "https://www.example.org/officiaoptio" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "ce337f2b-2ae9-4beb-8857-93c27364fec9",
-    "name" : "LJpdqB7nYgfKoTN",
-    "mimeType" : "cTwhJYijhhE6Wo",
-    "size" : 307043008,
+    "identifier" : "5f7ca26b-0d97-46bc-942c-5e51547060a2",
+    "name" : "WHG5W0Qnkhkr5tT7Uv",
+    "mimeType" : "ZxN0z2Zogpn",
+    "size" : 851891564,
     "license" : {
       "type" : "License",
-      "identifier" : "bSe8uHNPc9feS3nNxr",
+      "identifier" : "0qOgPi7wtUshYzk",
       "labels" : {
-        "hu" : "j3zaKdmeGkie7rb"
+        "nb" : "UUtFZqCwuauGPeJViqc"
       },
-      "link" : "https://www.example.com/9wM9cFPY9qV3"
+      "link" : "https://www.example.com/jnrGgfSXYoJIxUSR"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/cTqFkmOCvsnp2",
-    "name" : "uB0VqxpmUMVtKs7vj",
-    "description" : "Uodk5d6RLm2Ur2zu2E"
+    "id" : "https://www.example.com/WS4mwqNeShv7v9JoQ7",
+    "name" : "b0TAM9z4IQDuSYt",
+    "description" : "RAJvVVBHQd6no"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
