+++ documentation/MediaPodcast.json
@@ -3,78 +3,78 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT",
   "resourceOwner" : {
-    "owner" : "PPO6UZy36RoZeJANfr",
-    "ownerAffiliation" : "https://www.example.org/expeditapossimus"
+    "owner" : "ztV4vnvlQsmk",
+    "ownerAffiliation" : "https://www.example.org/eteveniet"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/occaecatidolores",
+    "id" : "https://www.example.org/autemdoloribus",
     "labels" : {
-      "fi" : "qK3O8Y3jUvl1KHDZs"
+      "zh" : "UI4aPitckDvOa2a23Ot"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quiamolestiae",
-  "doi" : "https://doi.org/10.1234/odio",
-  "link" : "https://www.example.org/repellatneque",
+  "handle" : "https://www.example.org/noninventore",
+  "doi" : "https://doi.org/10.1234/inventore",
+  "link" : "https://www.example.org/inventoreid",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "pYgZSwXpDLr0a4Y",
+    "mainTitle" : "i4gtn4yW5FxGh8Pb0sF",
     "alternativeTitles" : {
-      "ca" : "Rs4JVHqPL2lRHIaRwM"
+      "se" : "ZueXWLEIU95"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "DBl4p2UBqhqI",
-      "month" : "x857C8kjcQXZ",
-      "day" : "zR8zIB8u5qvU1CS"
+      "year" : "BtKftRtkJaaeZH",
+      "month" : "UY26osqjUlJ4N",
+      "day" : "bkUrVcZfkLYn6DEEsSl"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/RIsBkcOcCSxwu",
-        "name" : "jjO8L1DNmOkjXn8Z",
+        "id" : "https://www.example.com/NalvA3m9ylF",
+        "name" : "mxK8FhoSc2",
         "nameType" : "Personal",
-        "orcId" : "auqK9EaDpgpLJ"
+        "orcId" : "qoFwg3MhsaaOz6FyII"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/utFPlUX2bH",
+        "id" : "https://www.example.com/HzjohqNyYN2AD",
         "labels" : {
-          "nn" : "APQ8KPBApVufSYXGl"
+          "ca" : "GWlqyq57boPYVfMpx"
         }
       } ],
-      "role" : "VideoEditor",
+      "role" : "Curator",
       "sequence" : 0,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/93G5miqz16qCDsYAFxg",
-        "name" : "5rviiZsa1st",
-        "nameType" : "Personal",
-        "orcId" : "1n5eTEKw9dy0A"
+        "id" : "https://www.example.com/ZLIqAHgcUiRKf3Xa86N",
+        "name" : "GndqIJJ23QzOo4Ssj",
+        "nameType" : "Organizational",
+        "orcId" : "ASCrq8Sca6qsvtpl"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/xbB5MUsFUqUSqstAx",
+        "id" : "https://www.example.com/wpsPoyY35XkC1O96lP",
         "labels" : {
-          "hu" : "pbXdcpWfBT"
+          "it" : "QYFkZg2gF2EadCAlLW"
         }
       } ],
-      "role" : "Sponsor",
-      "sequence" : 3,
+      "role" : "ArtisticDirector",
+      "sequence" : 6,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "9Tg9rgpeogpg7OySpaw",
-    "tags" : [ "lQpdIUjclTKZ" ],
-    "description" : "k1n4eTW5iQ9Og",
+    "npiSubjectHeading" : "lgo9Sf76P8RZ",
+    "tags" : [ "f3CWJs9BgIVDhOsFPQB" ],
+    "description" : "H1jEyMcWja2IX0qIS1M",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
@@ -82,15 +82,15 @@
         "medium" : {
           "type" : "TV"
         },
-        "format" : "Text",
-        "disseminationChannel" : "ej2K7RiR3ciEMSH",
+        "format" : "Sound",
+        "disseminationChannel" : "yU1IcBmTpP3",
         "partOf" : {
           "type" : "SeriesEpisode",
-          "series" : "ZHAUgPnh9xfGRp0YL",
-          "seriesPart" : "uxb2cRW13UqwkwJNX"
+          "series" : "lhDPRrj4IRP",
+          "seriesPart" : "ElMEdBRFiHZc"
         }
       },
-      "doi" : "https://www.example.com/RnN4lrqSJ9",
+      "doi" : "https://www.example.com/XhVU8pMyCzf2",
       "publicationInstance" : {
         "type" : "MediaPodcast",
         "peerReviewed" : false,
@@ -99,24 +99,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/sKaIUx3NEuyIU1tjE",
-    "abstract" : "5x543jgLeJisLB"
+    "metadataSource" : "https://www.example.com/eL8vkMudyo",
+    "abstract" : "PGLic3M3hJPUOGzJ8BM"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/exvoluptatibus",
-    "name" : "Hnw0t6vNW9HcpwAND",
+    "id" : "https://www.example.org/utab",
+    "name" : "NbBml6MXEEkStNG",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "Vdiduh9d9QG3l",
-      "id" : "MfZQlZnKpN0sBJYuMus"
+      "source" : "IQ4iFckn4Yj",
+      "id" : "5Qhi1hmu5zitjT"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "REK",
-      "approvalStatus" : "REJECTION",
-      "applicationCode" : "2fhdAQR1czdpUXmG"
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "9zURLQNV26aiq8TDqvO"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -124,30 +124,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/fugaharum" ],
+  "subjects" : [ "https://www.example.org/quoeos" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "4752dda8-af81-4069-a33e-5aca53a68416",
-    "name" : "4C8zKO8Tp8wJKfA8mUv",
-    "mimeType" : "Ri5fByerw7ufz",
-    "size" : 1411120275,
+    "identifier" : "3a86be6c-0b69-4a3b-a212-5d8e6972b848",
+    "name" : "LmTeJzJMdAJGJb",
+    "mimeType" : "6KNbkeYOm65Xz",
+    "size" : 1944034240,
     "license" : {
       "type" : "License",
-      "identifier" : "wsm2vU5gMvxw6UX",
+      "identifier" : "d53RPt22qXtMq",
       "labels" : {
-        "ca" : "W14fDLFA5TKZ0aCP"
+        "fi" : "PSKtvQy2MrkvXydFlP"
       },
-      "link" : "https://www.example.com/WSRvdAYSsHyaz"
+      "link" : "https://www.example.com/8dYr3undFw0n"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/dlGdNrhVmmntU",
-    "name" : "C1cu3WUZddM",
-    "description" : "M9rYsZulVvo0rzVPf"
+    "id" : "https://www.example.com/dNg2BNdiER0OXNtyY",
+    "name" : "0WtprovjUF4Zkytchn",
+    "description" : "TpmqBPkQQl0YgBcnill"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
