+++ documentation/MediaReaderOpinion.json
@@ -3,116 +3,118 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "zITWrV3azijc",
-    "ownerAffiliation" : "https://www.example.org/corruptiarchitecto"
+    "owner" : "BTkAT2VbGA",
+    "ownerAffiliation" : "https://www.example.org/dolorevoluptas"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/quasirepellat",
+    "id" : "https://www.example.org/quorerum",
     "labels" : {
-      "af" : "Wc6sRHiJX6DXGJYS5"
+      "fi" : "023hYH8T0jRQQxVqvHc"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quiofficia",
-  "doi" : "https://doi.org/10.1234/sint",
-  "link" : "https://www.example.org/commodimodi",
+  "handle" : "https://www.example.org/deseruntofficia",
+  "doi" : "https://doi.org/10.1234/repellat",
+  "link" : "https://www.example.org/quivoluptas",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "NFhTy0Iu2hV3n9",
+    "mainTitle" : "JFvpYkxbxAEpU",
     "alternativeTitles" : {
-      "hu" : "6uXtk0l0pUCCAdR4v"
+      "el" : "nfMxEtNwUzXk6JLdD"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "DodMuRcXlwh7i5wO",
-      "month" : "FeTU3XuwUXI8qdQj59",
-      "day" : "cozV6fBPXJnUt"
+      "year" : "mQszMjZL9Aa1",
+      "month" : "iK2RM167HRVDoCLG",
+      "day" : "DGzy2AqTpU"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/M7aRLadMTEiXlFGZ2WX",
-        "name" : "OKEoleJ3euH",
+        "id" : "https://www.example.com/llkgxFgQGU",
+        "name" : "TBTWksMTJejY6t70dv3",
         "nameType" : "Personal",
-        "orcId" : "uIrmwamDD4TMBQHrKt"
+        "orcId" : "Hv6nr56dhl88JSa"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/xpzeYuGEzOvkTmP",
+        "id" : "https://www.example.com/L7VFzMIggwDv6T",
         "labels" : {
-          "es" : "76jK6HlwjJzgsvJKoLu"
+          "is" : "jrwhADUAyMbAF1c9xB"
         }
       } ],
-      "role" : "Editor",
-      "sequence" : 4,
+      "role" : "Other",
+      "sequence" : 9,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/o2VsdDBve6hnUy",
-        "name" : "ehQWgPA7iS",
+        "id" : "https://www.example.com/qOy5ALwRDee9sDy",
+        "name" : "I7mRbPwbKSs5",
         "nameType" : "Organizational",
-        "orcId" : "cOGc2ccwXb08O"
+        "orcId" : "79FW5NeatTqcRVW12"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/ehiU0dguyFs",
+        "id" : "https://www.example.com/8MDJBEB0E8up",
         "labels" : {
-          "is" : "F4qZmTkIkT0Fnh3J"
+          "it" : "HU5cQOQeJhMentfMw"
         }
       } ],
-      "role" : "Artist",
-      "sequence" : 2,
+      "role" : "LightDesigner",
+      "sequence" : 4,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "P9cvw0iujdV6",
-    "tags" : [ "WXncPKZfHY" ],
-    "description" : "UHGqYM8lbK",
+    "npiSubjectHeading" : "5BELB1livDV",
+    "tags" : [ "W15srkz0ItqHC0q" ],
+    "description" : "VyVhHoKj4977W3cA5Nw",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
-        "type" : "MediaContributionPeriodical",
-        "id" : "https://api.dev.nva.aws.unit.no/publication-channels/veZp5mGQown"
+        "type" : "UnconfirmedMediaContributionPeriodical",
+        "title" : "R5mtBtWOVQo6IELPl1A",
+        "printIssn" : "8821-7868",
+        "onlineIssn" : "7389-6713"
       },
-      "doi" : "https://www.example.com/JbQBYQsIWxD837M",
+      "doi" : "https://www.example.com/MMr7AdhojFXWa5P",
       "publicationInstance" : {
         "type" : "MediaReaderOpinion",
-        "volume" : "UF932QHEcyy",
-        "issue" : "IRuR9f9xOpmCei7f",
-        "articleNumber" : "zMFvfHjGHHmBi84x9",
+        "volume" : "mx5wpN2nTZh1qQCs",
+        "issue" : "idVNixOJFDAMat",
+        "articleNumber" : "mGdnA79ypFeN",
         "pages" : {
           "type" : "Range",
-          "begin" : "4EDdgz91laTa7D7m",
-          "end" : "zFtTd3iJax"
+          "begin" : "qftvnjQZlNKPPBC",
+          "end" : "oxSt7JtWdJ8ENkbL"
         },
         "peerReviewed" : false
       }
     },
-    "metadataSource" : "https://www.example.com/5J3scXFdx9V",
-    "abstract" : "1iU4LlqQvDLU8vxNAkL"
+    "metadataSource" : "https://www.example.com/T7rRZStDfRv2Fg9k",
+    "abstract" : "4TFDnZMd9hND41r"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/laboriosamest",
-    "name" : "AUjd4RSXThAFt4",
+    "id" : "https://www.example.org/vitaesapiente",
+    "name" : "dPrEUGEYummzJqXyIzz",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "Z5aoX51NlvJWSSofkQ",
-      "id" : "6JV3maan49W44rX99a8"
+      "source" : "xhT9X9gnIaTTHQZ",
+      "id" : "BB9xhT2ThgfSOYWWN7"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "DIRHEALTH",
-      "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "UHhHD7isq7zOhKEe3Fi"
+      "approvedBy" : "REK",
+      "approvalStatus" : "APPROVED",
+      "applicationCode" : "h81I0YW4E5e"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -120,30 +122,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/estbeatae" ],
+  "subjects" : [ "https://www.example.org/assumendanobis" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "bbbf6eb0-f4c8-41e4-934b-866d3a93c76e",
-    "name" : "5qDLlZMCQTY0MWePVF",
-    "mimeType" : "Dk6Swkms5L2R76UTYk",
-    "size" : 189972456,
+    "identifier" : "401b21e2-0b03-41b0-93cb-d0defb6d858e",
+    "name" : "OHsVsP3vNymUpM",
+    "mimeType" : "5eOsAbMHxYcjeNOyp",
+    "size" : 1456007841,
     "license" : {
       "type" : "License",
-      "identifier" : "qnwG2byAFoyiBu",
+      "identifier" : "KAjDuaKLE0s",
       "labels" : {
-        "fi" : "1m1CggcGvYI1cIEiH"
+        "fr" : "pkREKB5ktdRkL"
       },
-      "link" : "https://www.example.com/Hsm7ZepW59RsLk"
+      "link" : "https://www.example.com/9QhlDt2z0s"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/GlcHrXhHfzoJwg3",
-    "name" : "vP4MbVq9rcjoN0",
-    "description" : "w9K0o5yrucgYpmq"
+    "id" : "https://www.example.com/yAuqeXPcTI5I1Frdq",
+    "name" : "3WNXW41cGxungCaY",
+    "description" : "4eIJkGAINvl8d"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
