+++ documentation/OtherPresentation.json
@@ -3,102 +3,103 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "hikp98fe1wmLg5y3O",
-    "ownerAffiliation" : "https://www.example.org/maioresreprehenderit"
+    "owner" : "rnuxpHXS0Em3qf",
+    "ownerAffiliation" : "https://www.example.org/autemnisi"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/autet",
+    "id" : "https://www.example.org/ullamincidunt",
     "labels" : {
-      "it" : "5ycDN5sBRw1jiZ3"
+      "de" : "8sCdJ31LTUqUPtkmj"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/dolorumarchitecto",
-  "doi" : "https://doi.org/10.1234/rerum",
-  "link" : "https://www.example.org/utdeleniti",
+  "handle" : "https://www.example.org/nondolore",
+  "doi" : "https://doi.org/10.1234/perferendis",
+  "link" : "https://www.example.org/accusantiumqui",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "kFmEzdp1zpCK7dYm94k",
+    "mainTitle" : "TJKU2HcujrbfW",
     "alternativeTitles" : {
-      "pt" : "enjnGhbSQRTuja53va6"
+      "ru" : "2zaDJ7n5o8TRb7gx"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "UFupi5gn507TWhS0",
-      "month" : "Mk8dUERMmJLAsF",
-      "day" : "rpxtuKTppD"
+      "year" : "0Jwu0JQnsW",
+      "month" : "EN7xKTS4OS",
+      "day" : "muY7yAGom3L"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/112LBPTBtYhnL",
-        "name" : "rTDD2gAWd1KsprCrNe",
-        "nameType" : "Organizational",
-        "orcId" : "bBOpZ9h2UOapf"
+        "id" : "https://www.example.com/FUWSYMp95dS1AQjP",
+        "name" : "tyJfq7qqvYe4",
+        "nameType" : "Personal",
+        "orcId" : "cMYGrB5rl2gz"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/Z2ieyy1Z9Wj0Pyr",
+        "id" : "https://www.example.com/2fXZoYh2ShvFd",
         "labels" : {
-          "nn" : "QR0EWvjpKh2w9jv"
+          "ru" : "TTix4UmItY81mpUVUmp"
         }
       } ],
-      "role" : "Producer",
-      "sequence" : 3,
+      "role" : "Dramatist",
+      "sequence" : 1,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/oEAAWg3A61e7TqLTLJ2",
-        "name" : "hnRaDVLQyG7vFYP",
-        "nameType" : "Organizational",
-        "orcId" : "MJOsNxZpS6wUwXR2bq"
+        "id" : "https://www.example.com/Fqb1v7yipeQD",
+        "name" : "7HGAiKdWrQ",
+        "nameType" : "Personal",
+        "orcId" : "7EVgTXLkCnM"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/VP5dQFDxePxrqV1m7F",
+        "id" : "https://www.example.com/WD4HNivCbTbyrP",
         "labels" : {
-          "se" : "h7w0VZSQl1k2kUKf"
+          "nn" : "Znb7N4jePEOCr"
         }
       } ],
-      "role" : "ProjectMember",
-      "sequence" : 9,
+      "role" : "Photographer",
+      "sequence" : 2,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "0PGdELzjDOtI",
-    "tags" : [ "0Q8t4jF21V8K50qgm" ],
-    "description" : "EVbDGHgwADIKyKSw",
+    "npiSubjectHeading" : "XxqY3irIeDVV6I32s",
+    "tags" : [ "Qvbns265AVEBr3S3" ],
+    "description" : "6ZV7QEFQbjccHNi7XBk",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Event",
-        "label" : "4KFyviB143",
+        "label" : "AwiWYqez3odxxzqaHU",
         "place" : {
           "type" : "UnconfirmedPlace",
-          "label" : "pBrQbprX2XVc4Sg",
-          "country" : "ljP7r1TOnet79"
+          "label" : "UyuCYkeaHI8Ks6FFyWG",
+          "country" : "g3NCYZZeJS4ziXTa7O"
         },
         "time" : {
-          "type" : "Instant",
-          "value" : "2020-09-23T09:51:23.044996Z"
+          "type" : "Period",
+          "from" : "2020-09-23T09:51:23.044996Z",
+          "to" : "2020-09-23T09:51:23.044996Z"
         },
         "agent" : {
           "type" : "Organization",
-          "id" : "https://www.example.com/vieSdDEYP0YIfi8w",
+          "id" : "https://www.example.com/frHgeaNQiL0GKSdz0Gn",
           "labels" : {
-            "is" : "AR75J3GG1uKfeh5v"
+            "el" : "YRiA87mpuwi"
           }
         },
-        "product" : "https://www.example.com/IIAg4ITLld68p3EATM"
+        "product" : "https://www.example.com/aVsEWuyCiWjLSlr4xEm"
       },
-      "doi" : "https://www.example.com/n8BjhU5d9uQ91o7",
+      "doi" : "https://www.example.com/cF0FC5PfNg3wxn4U",
       "publicationInstance" : {
         "type" : "OtherPresentation",
         "peerReviewed" : false,
@@ -107,24 +108,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/eCSeeqHECXbRkYlK5iS",
-    "abstract" : "e48HcDdLeNBpP"
+    "metadataSource" : "https://www.example.com/Vk0UeA14Y2",
+    "abstract" : "8GpmGR4uCbj5"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/nostrumquos",
-    "name" : "1ihbFV5SggtWHbLyvEC",
+    "id" : "https://www.example.org/inventorereprehenderit",
+    "name" : "BwXJe8E2XDKWvqQj",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "twMSlxoL3F",
-      "id" : "1XczsBkP3m2Czp3N"
+      "source" : "UyRovld1FKtJ9YB2bOM",
+      "id" : "Py8EyxtUHyTxzFM"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "DIRHEALTH",
       "approvalStatus" : "DECLINED",
-      "applicationCode" : "INQfo8P7tMj"
+      "applicationCode" : "kkcczTIXwXVq8"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -132,30 +133,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/rationeodit" ],
+  "subjects" : [ "https://www.example.org/quiavelit" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "f9e9eebb-7733-437c-b524-3e69d51301c4",
-    "name" : "5lk1ckUhlNBxSA",
-    "mimeType" : "wjdDx49irJZYrI",
-    "size" : 389668257,
+    "identifier" : "1e599cad-297d-4fe6-b5ed-b95bb8d8cb2b",
+    "name" : "0FoVYgaQw0itFxO95f",
+    "mimeType" : "9XiUN8DCDv1pQiI",
+    "size" : 239287947,
     "license" : {
       "type" : "License",
-      "identifier" : "DPIbheIChnjyqKsW",
+      "identifier" : "cpbOOFtY12U5OhtRdV",
       "labels" : {
-        "de" : "r7goaVQOtoR"
+        "zh" : "iSrWMa5dv9a"
       },
-      "link" : "https://www.example.com/RFGbeqFaRm2zEcJ"
+      "link" : "https://www.example.com/nvQmwvjEGrVhqAey"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/AN3REwJdPSHk",
-    "name" : "aK77K3J32O3PCV",
-    "description" : "I9ZnfW2iBb3sFBYdVOE"
+    "id" : "https://www.example.com/8MDfpj3Kq154AWv",
+    "name" : "UocIRNzhVUNZro",
+    "description" : "vyWQVePY3VQes7IHSNF"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
