+++ documentation/OtherStudentWork.json
@@ -1,135 +1,135 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "NEW",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "j2oBJskROeQ",
-    "ownerAffiliation" : "https://www.example.org/vitaeducimus"
+    "owner" : "1PaQ3enVYeN",
+    "ownerAffiliation" : "https://www.example.org/inciduntcumque"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/accusantiumrepellat",
+    "id" : "https://www.example.org/inet",
     "labels" : {
-      "pt" : "uLOpnFOwvc"
+      "zh" : "AXacywB9R49v5TwxRI"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/quiquo",
-  "doi" : "https://doi.org/10.1234/impedit",
-  "link" : "https://www.example.org/voluptatemerror",
+  "handle" : "https://www.example.org/doloredolor",
+  "doi" : "https://doi.org/10.1234/quo",
+  "link" : "https://www.example.org/impeditomnis",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "PokynHvqmpld",
+    "mainTitle" : "bOHRKCfu5bGiWBdJty",
     "alternativeTitles" : {
-      "pt" : "coMvqvnaykerR"
+      "se" : "HuQmKInQUPhvn"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "UnFaqp7CBKlD4BBX",
-      "month" : "w2lXMJcYesLe",
-      "day" : "VCalKiUfdOcRy"
+      "year" : "XA4rRQr94FT",
+      "month" : "Fd1UUOYXMpxLVF",
+      "day" : "EMUwN2l6g9MS4lc1D7"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/HmpRyZsmMZaHBzdoi",
-        "name" : "PnvLMhIUdCef1IFIB",
-        "nameType" : "Organizational",
-        "orcId" : "yZFOmG9S4mam"
+        "id" : "https://www.example.com/KFr4oKGSVTEd",
+        "name" : "mU3LhqrYIuRq0ewN3je",
+        "nameType" : "Personal",
+        "orcId" : "lMrepVzHOD"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/F3HdosboqR",
+        "id" : "https://www.example.com/wbqAnmIznKsSX3",
         "labels" : {
-          "sv" : "u6nuHT99cdeNvLJ"
+          "pl" : "I8L4o0pwBFJaACF"
         }
       } ],
-      "role" : "HostingInstitution",
-      "sequence" : 1,
+      "role" : "ContactPerson",
+      "sequence" : 3,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/mqhp1dQxG1",
-        "name" : "36xPBip4PY2mGc",
-        "nameType" : "Personal",
-        "orcId" : "eXFlWoTwEZWg6P"
+        "id" : "https://www.example.com/4B5c8Y4otuQTSQ1",
+        "name" : "L710IgdcQbXaH",
+        "nameType" : "Organizational",
+        "orcId" : "FfVYwcI3iJv8r"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/YqoeDE1w5uex5V6",
+        "id" : "https://www.example.com/crN11IWES6",
         "labels" : {
-          "is" : "f7a30niPWX"
+          "bg" : "PLAbD2dwMqptGB8UDtI"
         }
       } ],
-      "role" : "Artist",
+      "role" : "WorkPackageLeader",
       "sequence" : 3,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "HRGuYodqLx06oNLL8",
-    "tags" : [ "Ya2vhMtHx1" ],
-    "description" : "KpYpgt3MZ6dN7n2",
+    "npiSubjectHeading" : "2UELPBqhs5uT59cCA",
+    "tags" : [ "HJS3i3XL89ixQ3" ],
+    "description" : "gWKyN1VHmxtq",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Book",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/kDAhszNoF9d9rRTkAJ"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/FJ7axdQHeRjLqd0i"
         },
-        "seriesNumber" : "aGP9yekpE2IwMTO9Rop",
+        "seriesNumber" : "pCKHV7d4sZra",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/SX33AYITbSbQzPiHAai"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/3DaLvZsymglsu"
         },
-        "isbnList" : [ "9791648748683" ]
+        "isbnList" : [ "9781204683565" ]
       },
-      "doi" : "https://www.example.com/Y2RS3KRfCvnfZ9yucp",
+      "doi" : "https://www.example.com/ezLBoC6Le5N4TZ0",
       "publicationInstance" : {
         "type" : "OtherStudentWork",
         "submittedDate" : {
           "type" : "PublicationDate",
-          "year" : "KOnoFfIONvK",
-          "month" : "GhpviCr6uwf",
-          "day" : "Yl8zH24rZiVj"
+          "year" : "fPfm9F7kIba",
+          "month" : "sDOGg97iFQV",
+          "day" : "GlTZmF1EyQywbJhBPIn"
         },
         "peerReviewed" : false,
         "pages" : {
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "2dtII5cnZk0kkyRl5S",
-            "end" : "5B3Y6L0wgXLfZ5M"
+            "begin" : "9mVhNkao9eFcNfd84",
+            "end" : "DC3zGGmMF0"
           },
-          "pages" : "C7NdsJeFxWc9h",
-          "illustrated" : false
+          "pages" : "8zT6dk8AyidUwdNFo",
+          "illustrated" : true
         }
       }
     },
-    "metadataSource" : "https://www.example.com/tttVkYM1vhiBq2Yw3VI",
-    "abstract" : "IyNKaFFn9RKlfZJSW"
+    "metadataSource" : "https://www.example.com/Vq376wzgpcmfsHvb3",
+    "abstract" : "EsAhbQj1zNGz"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/illorerum",
-    "name" : "K7uVUzUgAXzl0N",
+    "id" : "https://www.example.org/ullamvel",
+    "name" : "gavkpyP4Flya5b1R7",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "aVAmuP0eNfqHGxHjpVg",
-      "id" : "DS5exybHJVFXmp"
+      "source" : "5VG6c7Enzx8OI9",
+      "id" : "04B9PYHm82xnB8"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
+      "approvedBy" : "NMA",
       "approvalStatus" : "APPROVED",
-      "applicationCode" : "LDiV9Y129BW6zQ"
+      "applicationCode" : "V1Mz9VUqww5a5TYzK"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -137,30 +137,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/quaeratet" ],
+  "subjects" : [ "https://www.example.org/ametrepellendus" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "1e23cfd9-a778-43d5-aaec-3e04e03712f1",
-    "name" : "lAfoosdmQS3P",
-    "mimeType" : "gQ1KZVlalEUhhtOKu",
-    "size" : 913897587,
+    "identifier" : "8d62e817-3e89-4d9d-8d07-486816ca7c4d",
+    "name" : "FGgfbHZu7nPzcg",
+    "mimeType" : "PgSIAlH2mySq",
+    "size" : 1116272754,
     "license" : {
       "type" : "License",
-      "identifier" : "i7Ywls2Cw6AX3",
+      "identifier" : "d6fD9cwUZdl6dUbDh",
       "labels" : {
-        "fr" : "ozCrBeUCZGPe"
+        "nl" : "0UB44IcrkvV"
       },
-      "link" : "https://www.example.com/MC9uME2lsG"
+      "link" : "https://www.example.com/fzdrOdW5KeD"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/NLQ8S6OlUWl2ukryrs",
-    "name" : "QA1HRFkh0lF5d",
-    "description" : "97UIVkjL2h"
+    "id" : "https://www.example.com/vyvMb4QnD3in6m6PnaU",
+    "name" : "xmm6uWj78P",
+    "description" : "UX1gIGivmDQwyuZ"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
