+++ documentation/PerformingArts.json
@@ -1,52 +1,52 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "mZunsQ0jHx",
-    "ownerAffiliation" : "https://www.example.org/expeditaet"
+    "owner" : "D4CjmLzJva",
+    "ownerAffiliation" : "https://www.example.org/laborumassumenda"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/ideos",
+    "id" : "https://www.example.org/suscipitquos",
     "labels" : {
-      "fi" : "ManhX4pZqhJDoKeZ"
+      "ca" : "QMwHE4qRugaTz3mJeI"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/abqui",
-  "doi" : "https://doi.org/10.1234/et",
-  "link" : "https://www.example.org/sitsint",
+  "handle" : "https://www.example.org/estnam",
+  "doi" : "https://doi.org/10.1234/odit",
+  "link" : "https://www.example.org/etrerum",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "VEZLBdDtl9h",
+    "mainTitle" : "SBv5i4oFayzwzmS0hMu",
     "alternativeTitles" : {
-      "fi" : "oiGW7qOTEnTw80"
+      "en" : "27zrlzChDIE"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "f8ktjRsyZDqYn",
-      "month" : "lSPszMWae6POQOGP",
-      "day" : "k5lEMLlLIGPfjIz"
+      "year" : "7ukax6yR2LLfEo5m",
+      "month" : "OzAOuWy4MXFNh5OV",
+      "day" : "l3dbhLLdyR"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Kh1uGBe54e",
-        "name" : "O4Wy5gC0noV",
+        "id" : "https://www.example.com/mRQhw87XKafOoEl",
+        "name" : "3C7wpk3BRfB0Fl6q",
         "nameType" : "Personal",
-        "orcId" : "38qlk9ooltmB7poFVk"
+        "orcId" : "ZoblifjQSPMWHfP1J"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/JDHSz79UAXL4O3ibkVA",
+        "id" : "https://www.example.com/HdHOUDB1FzqUQs",
         "labels" : {
-          "se" : "CPdqj5qoBtL2MlW"
+          "el" : "5d4aUMVLfIHYA29Uhug"
         }
       } ],
       "role" : "Designer",
@@ -56,50 +56,51 @@
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/5KTm1pzaIweBdIgGyd",
-        "name" : "aI3aDWWL58hAAuPZ",
-        "nameType" : "Personal",
-        "orcId" : "AB1E07zqPYF8eI9"
+        "id" : "https://www.example.com/35uSXgz0J0ircPBKz",
+        "name" : "yv5qUMIlC1",
+        "nameType" : "Organizational",
+        "orcId" : "Z4y8fsr3MyFC2bDVvnT"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/B0Kiz6z1F9R79",
+        "id" : "https://www.example.com/0ZdWYgsLlk",
         "labels" : {
-          "nb" : "pEYjqAbRtPtICjfub"
+          "se" : "AAqzVVc0y8KF3n"
         }
       } ],
-      "role" : "Musician",
-      "sequence" : 6,
+      "role" : "ProjectManager",
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "kjvJ9Ddihk",
-    "tags" : [ "Vh1frCVgLH52Fvyfb" ],
-    "description" : "VsE7wpmQYWXb8V690A",
+    "npiSubjectHeading" : "eY3aG6NWAfjSL",
+    "tags" : [ "zbG3fP9MngveJgbN6tK" ],
+    "description" : "T8DWPUZuvlDvpVhuE",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Artistic"
       },
-      "doi" : "https://www.example.com/roIWasnwKe",
+      "doi" : "https://www.example.com/90NWdYw8i1DuBF5iXA4",
       "publicationInstance" : {
         "type" : "PerformingArts",
         "subtype" : {
-          "type" : "Broadcast"
+          "type" : "Other",
+          "description" : "M1KZatIQS3C13mE6rBr"
         },
-        "description" : "VNF5OXzrlSX",
+        "description" : "y8dZQWoR2OdBV",
         "outputs" : [ {
           "type" : "PerformingArtsVenue",
           "place" : {
             "type" : "UnconfirmedPlace",
-            "label" : "JGpGvChcCMOT",
-            "country" : "gN9MuutPZmPhHLdL"
+            "label" : "j380nSoE6HVvmO4MiX",
+            "country" : "xivt01s4nF1uc3AVHv"
           },
           "date" : {
             "type" : "Period",
             "from" : "2020-09-23T09:51:23.044996Z",
             "to" : "2020-09-23T09:51:23.044996Z"
           },
-          "sequence" : 1156216273
+          "sequence" : 896730560
         } ],
         "peerReviewed" : false,
         "pages" : {
@@ -107,24 +108,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/xvJsjkXfUCPNmp",
-    "abstract" : "dPGZ8HcOK5gauXBexf"
+    "metadataSource" : "https://www.example.com/wUoDJ2kUDrLm53RxC",
+    "abstract" : "SJIVaktRfFJYSqg"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/dignissimosmolestiae",
-    "name" : "YewEbExHQugo4v",
+    "id" : "https://www.example.org/undequibusdam",
+    "name" : "GYjYOD2QffkkcnUUW",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "Z0M8t66aoWi2O",
-      "id" : "aBPuYyr0WtFVrA5KF"
+      "source" : "fTefk7zHjqbp4MD",
+      "id" : "NLpCHhbAyIM6XiOKL"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "REK",
-      "approvalStatus" : "APPROVED",
-      "applicationCode" : "KmaFH0VpQ33wAdlj"
+      "approvedBy" : "NARA",
+      "approvalStatus" : "NOTAPPLIED",
+      "applicationCode" : "Clskf9bPoizE0r3LSr0"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -132,30 +133,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/velitut" ],
+  "subjects" : [ "https://www.example.org/evenietest" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "119b825d-4829-4d0c-bc5d-555e48ede214",
-    "name" : "lwk9v22QDw",
-    "mimeType" : "1UbK4cuNRu",
-    "size" : 936511384,
+    "identifier" : "892bbeaf-3f2f-4978-b544-30910cc9ade7",
+    "name" : "vUp4BurUyYso",
+    "mimeType" : "A59VbIeNk8",
+    "size" : 1810914703,
     "license" : {
       "type" : "License",
-      "identifier" : "AjB2pY767lJ",
+      "identifier" : "yTHXoqvhplb1OaYd",
       "labels" : {
-        "cs" : "6OMot68I3903q"
+        "bg" : "Dn2BUYJFbvuvw"
       },
-      "link" : "https://www.example.com/7qH4Er8Sc4MXDHhs"
+      "link" : "https://www.example.com/TSrG0iDodEJ7ryo"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/fuPyE9iaxTh2WC8",
-    "name" : "zsk1N335DQB",
-    "description" : "K0HGmjFWQzq4u"
+    "id" : "https://www.example.com/5pIUvdw5Zo1rx2",
+    "name" : "Ip3ebwCpmNw93",
+    "description" : "6Z6F0qKJudgz6UysTx"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
