+++ documentation/ReportBasic.json
@@ -1,96 +1,96 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "DRAFT_FOR_DELETION",
+  "status" : "NEW",
   "resourceOwner" : {
-    "owner" : "Du5xGByGMszObPdIoi",
-    "ownerAffiliation" : "https://www.example.org/utquae"
+    "owner" : "yhhCKw5R0tf4NypY",
+    "ownerAffiliation" : "https://www.example.org/veroducimus"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/quibusdamquidem",
+    "id" : "https://www.example.org/nihilaccusamus",
     "labels" : {
-      "hu" : "6eNpgyxfDMapxDsj6Rt"
+      "da" : "uHthD04gPGlZZPMG"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/delectusdolor",
-  "doi" : "https://doi.org/10.1234/ea",
-  "link" : "https://www.example.org/autemut",
+  "handle" : "https://www.example.org/utvoluptatibus",
+  "doi" : "https://doi.org/10.1234/nostrum",
+  "link" : "https://www.example.org/quisdolorem",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "ABlECcnjx2H3GQbZmP",
+    "mainTitle" : "7zyKcM15Idyunw",
     "alternativeTitles" : {
-      "ca" : "Q8YoU4ZcMFMsrdgV"
+      "ru" : "Yo9CBydBt5UxpFj"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "ZtfFfCBwuhhVhAreom",
-      "month" : "2Fuwn8zDyDzpmnSrGl",
-      "day" : "LY856nJhLyu6RUO9jtI"
+      "year" : "nL2w0ZxuGIisB",
+      "month" : "qXxHUhPe045ptI",
+      "day" : "zyUeJcF9JdsjOQo"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/yVb8gqy46uIKkZ",
-        "name" : "6ZYuplcF5tCal",
+        "id" : "https://www.example.com/pjKu2tv2OO",
+        "name" : "hhbVGYWHRzp8SOaH2C",
         "nameType" : "Organizational",
-        "orcId" : "g2WAn2zw1QQ3jtZpnj"
+        "orcId" : "DOqv4Ba7zzv"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/fVxoJJnq8E",
+        "id" : "https://www.example.com/RNOC41kadcXf",
         "labels" : {
-          "en" : "mZjc2VZJKc4Sy"
+          "sv" : "YxU1c1kry7"
         }
       } ],
-      "role" : "Soloist",
-      "sequence" : 9,
+      "role" : "ContactPerson",
+      "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/DAltD5ancA",
-        "name" : "D7JQUwuGYhEetUZjj2",
+        "id" : "https://www.example.com/1bizm6EmS51cgoJ6i",
+        "name" : "9pMW15MOtRnLueDwXTR",
         "nameType" : "Personal",
-        "orcId" : "D6wgKD0JYG1w"
+        "orcId" : "FsmiJXuFEA"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/CT2bxiXbNGQRGwm",
+        "id" : "https://www.example.com/GpV8PFepuX5f",
         "labels" : {
-          "bg" : "gMnaomAjMKCoGwp"
+          "fr" : "Ouj3eiehEVw"
         }
       } ],
-      "role" : "ProductionDesigner",
-      "sequence" : 3,
+      "role" : "ResearchGroup",
+      "sequence" : 0,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "KNk3kUyuCjOpNIUVEKL",
-    "tags" : [ "mxx0Ve6udk9SLtZOM" ],
-    "description" : "VLlUg6tBcGIoLb",
+    "npiSubjectHeading" : "uqrDnOLVHQ15Iiyo827",
+    "tags" : [ "eWNzBSs02LpVO" ],
+    "description" : "uNIRWuo2uh",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Report",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/tVBwHesOcDwEgIXj0"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/0X8MhKmwfKATxhXU9"
         },
-        "seriesNumber" : "Lo807ZVjbRp6h3",
+        "seriesNumber" : "weTbfFpxnKD",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/ZIzBnckg01FGtK5FYKy"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/RbrgDICxfpqNBdi"
         },
-        "isbnList" : [ "9781039153585" ]
+        "isbnList" : [ "9790872373357" ]
       },
-      "doi" : "https://www.example.com/SKQvBf7V2PlSh",
+      "doi" : "https://www.example.com/8zG4UB9jga",
       "publicationInstance" : {
         "type" : "ReportBasic",
         "peerReviewed" : false,
@@ -98,32 +98,32 @@
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "WngIFndNJMMSgQdUpHP",
-            "end" : "dIBdRS6SenfQW6th"
+            "begin" : "WkzXraUvc8IRbmxVcpL",
+            "end" : "8uo5NEzdn3gw8d"
           },
-          "pages" : "MhQVb3gmojnWsEQEB7I",
-          "illustrated" : true
+          "pages" : "qFxrcFoxLgaAZM5ku",
+          "illustrated" : false
         }
       }
     },
-    "metadataSource" : "https://www.example.com/KMIZwDT5A1dOsCo6",
-    "abstract" : "gqqkarunKlnJc4nb1"
+    "metadataSource" : "https://www.example.com/xSaYWTwxIjFyVHF7pE",
+    "abstract" : "pR4MIHpRm3r3E99"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/rationealias",
-    "name" : "u56xW4uNTYrCL5P",
+    "id" : "https://www.example.org/providentdoloremque",
+    "name" : "uLOF1xtLPY",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "GgwpFTDW7PF",
-      "id" : "fu5aT7GG3JzCQ"
+      "source" : "xnD2HLmdjPv",
+      "id" : "N0M17LRJ28z"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NARA",
+      "approvedBy" : "NMA",
       "approvalStatus" : "APPROVED",
-      "applicationCode" : "bDUO76VxtA5"
+      "applicationCode" : "zXQDoO4jTbAkfVVKe4"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -131,30 +131,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/omnisquia" ],
+  "subjects" : [ "https://www.example.org/voluptatibusquaerat" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "124c5ac0-5664-42a4-809a-dde250996ce2",
-    "name" : "JCsKoM0C1GqryICc",
-    "mimeType" : "Gas1LJqnVaZa",
-    "size" : 10217655,
+    "identifier" : "c6d9406d-bac7-4db6-8e8f-b289ba3cc673",
+    "name" : "ySNK6bXUPEim",
+    "mimeType" : "mX9l9vjbHtyN",
+    "size" : 1502656181,
     "license" : {
       "type" : "License",
-      "identifier" : "OLs24nRhdxvv3Va6",
+      "identifier" : "siNrDCiVuBO",
       "labels" : {
-        "fi" : "m8B8phxbjPPDoEf"
+        "se" : "uD4xzmJlggKmBXLagW"
       },
-      "link" : "https://www.example.com/jQml4E5APfLL32M"
+      "link" : "https://www.example.com/Rb9fQCIcSfzhiDa"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/ngOMb0cWogo8",
-    "name" : "G0smlDXGHykaIpC5A4",
-    "description" : "j4LlMsgTqF3xqiyzhBo"
+    "id" : "https://www.example.com/9eunLn4pnS",
+    "name" : "6ZnHSp4qezv",
+    "description" : "COTpkJ4TAc"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
