+++ documentation/ReportBookOfAbstract.json
@@ -1,96 +1,96 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "NEW",
+  "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "9FMupyXJYbUPj",
-    "ownerAffiliation" : "https://www.example.org/fugitaut"
+    "owner" : "vziIb2pxE2C9Gjebh",
+    "ownerAffiliation" : "https://www.example.org/beataeimpedit"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/dictalaudantium",
+    "id" : "https://www.example.org/molestiaevoluptatem",
     "labels" : {
-      "sv" : "yyTrMYsTgDLbraqZI3"
+      "is" : "BpHlcsBDssje9"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/cumquesint",
-  "doi" : "https://doi.org/10.1234/rerum",
-  "link" : "https://www.example.org/voluptateincidunt",
+  "handle" : "https://www.example.org/utexercitationem",
+  "doi" : "https://doi.org/10.1234/aut",
+  "link" : "https://www.example.org/laboriosamrerum",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "nIKhcHrnLSLnw8Ktag",
+    "mainTitle" : "pGwL1gLNUATPr2FB8e",
     "alternativeTitles" : {
-      "nl" : "TsiD5m7sFpY"
+      "sv" : "TR9KCCPXGlR2krPtru"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "NPfSZDYc7OXFV",
-      "month" : "KhR2dSUVK5aq",
-      "day" : "qNMzts1uCcFswYjMbE"
+      "year" : "WzJHYFkxUOc6gBHHl1",
+      "month" : "PZ6eFeoOWXMIdNdSWgW",
+      "day" : "nYstkNvsvQuK0k1Yv"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/TDipxxTh494c",
-        "name" : "zRHIzE8eGh5z2K",
+        "id" : "https://www.example.com/QVJcLb0laIlxU",
+        "name" : "LXmTC9AOSqWZW96IR",
         "nameType" : "Organizational",
-        "orcId" : "GuDv5tt6X7Rc1"
+        "orcId" : "544HuGGoAa"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/0Mp6EGDcfjH",
+        "id" : "https://www.example.com/GZifOfZahMfHZEL",
         "labels" : {
-          "pl" : "CWaCeP00XTg2"
+          "pl" : "VBnO0ANCdn4"
         }
       } ],
-      "role" : "Distributor",
-      "sequence" : 1,
+      "role" : "TranslatorAdapter",
+      "sequence" : 5,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/yJsDKA9ktixi5",
-        "name" : "SKkh84vTm9FYWDcN",
+        "id" : "https://www.example.com/9RLoPBtvVGdzZm3Ji",
+        "name" : "wXZginTP3nLYzD5P",
         "nameType" : "Organizational",
-        "orcId" : "fFm7JqBlLuxfmWj0"
+        "orcId" : "PZ34CC6fh99"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/HYDh0KQDehV49",
+        "id" : "https://www.example.com/BGyIEtQMeQdm5iGb3",
         "labels" : {
-          "ru" : "iK86GFhj4tqu9EAc"
+          "cs" : "Q9i1FoodQ8UW1lS"
         }
       } ],
-      "role" : "SoundDesigner",
-      "sequence" : 9,
+      "role" : "Funder",
+      "sequence" : 3,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "JUXZ5Ocf9mhO",
-    "tags" : [ "Lk7HSGQl7Ql" ],
-    "description" : "7rbKSJu2sYKIsQGh8X",
+    "npiSubjectHeading" : "8BpkIEnHiXEEpuuqe2A",
+    "tags" : [ "Wm1NQIs2dzNIebu3i" ],
+    "description" : "BSjSKfCFIthBS1B07N",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Report",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/VGnorW8iQYb"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/Xcxsq73M1As"
         },
-        "seriesNumber" : "uT2f7VqgVc",
+        "seriesNumber" : "brbH3apoykjcNgXw",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/qYhGg3Egzo"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/y3y5EzWXNy9vh1d"
         },
-        "isbnList" : [ "9791943614546" ]
+        "isbnList" : [ "9780882658674" ]
       },
-      "doi" : "https://www.example.com/cNC6X1BL7qK",
+      "doi" : "https://www.example.com/TOhBB8uPZK",
       "publicationInstance" : {
         "type" : "ReportBookOfAbstract",
         "peerReviewed" : false,
@@ -98,32 +98,32 @@
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "um4Qf2PtsnxFxZQUF0",
-            "end" : "pK8vkowfBMG"
+            "begin" : "HCotDnDlx2dy0C",
+            "end" : "b682IFlaDMaG3h3KE"
           },
-          "pages" : "AXMESogVAQMIvosLXo",
-          "illustrated" : true
+          "pages" : "QSHoQ1w6ty6",
+          "illustrated" : false
         }
       }
     },
-    "metadataSource" : "https://www.example.com/Iuo670jZS5xcqnAF",
-    "abstract" : "0dafJ3Py3VipPel1V"
+    "metadataSource" : "https://www.example.com/CTqHgFL2bbTCVAj3",
+    "abstract" : "fveBdwlOIAi1"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/autembeatae",
-    "name" : "Bp8hbkER6TV9gRYTB",
+    "id" : "https://www.example.org/beataenobis",
+    "name" : "qiYNqUEvKXODwhg0B0N",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "y6xW4v9RWt9aErB2b",
-      "id" : "cbaUVU4dc4d"
+      "source" : "10jNGOekFeVyjdE",
+      "id" : "XqtCvMR0yC"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
-      "approvedBy" : "NMA",
+      "approvedBy" : "DIRHEALTH",
       "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "EvMMmx2CATxeSR"
+      "applicationCode" : "Hji5ABc8q56Ff1fEd"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -131,30 +131,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/utprovident" ],
+  "subjects" : [ "https://www.example.org/perspiciatisnecessitatibus" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "cd719740-239b-4417-b830-4451c45df0de",
-    "name" : "YXfZ8EMXuXR0nWEdsTZ",
-    "mimeType" : "RQDCWnPYdL7",
-    "size" : 1228819383,
+    "identifier" : "560aa889-0e46-4345-9a40-fc4774daf388",
+    "name" : "I3SkohJSb1w6vc",
+    "mimeType" : "pceQuWf8mFD",
+    "size" : 1782477835,
     "license" : {
       "type" : "License",
-      "identifier" : "fH73qaNxuEj",
+      "identifier" : "t0iZvQkLmJ",
       "labels" : {
-        "nn" : "VPiEH0B87dW8MC"
+        "de" : "VRoBt4M4XgzczpL9o"
       },
-      "link" : "https://www.example.com/syWgp5pPuVMZwiMG"
+      "link" : "https://www.example.com/rLJc645xLPdM87A08X"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/aRVz47C9cQaZUGICZz0",
-    "name" : "8xs8aSmCda2XPel9",
-    "description" : "gVk7I9oDhkIkHhMO"
+    "id" : "https://www.example.com/go4mhVBfm1AErZ",
+    "name" : "nsEYaDG69yeKZiHY6",
+    "description" : "R7aijyq9pB"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
