+++ documentation/ReportPolicy.json
@@ -3,94 +3,94 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "LxJSKjSHwVK63PIQup",
-    "ownerAffiliation" : "https://www.example.org/eossapiente"
+    "owner" : "9wHZXrvcfjypjyY2",
+    "ownerAffiliation" : "https://www.example.org/architectosunt"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/sedunde",
+    "id" : "https://www.example.org/doloribusut",
     "labels" : {
-      "hu" : "oUHVO6AOqonpHbBFF"
+      "is" : "AcAd7FEI5a"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/eumat",
-  "doi" : "https://doi.org/10.1234/deleniti",
-  "link" : "https://www.example.org/delectusaccusamus",
+  "handle" : "https://www.example.org/numquammagnam",
+  "doi" : "https://doi.org/10.1234/maxime",
+  "link" : "https://www.example.org/sunthic",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "pXzE0PmcWCa6KTJW",
+    "mainTitle" : "oUKWX72hct6aAC45a",
     "alternativeTitles" : {
-      "es" : "pEdwNb0KQgbpAZO35"
+      "cs" : "po67fEWbhTjyNN8hLT"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "bfkD6dNVmOwZjrqN2Xb",
-      "month" : "lUtwuKKqVFdQ",
-      "day" : "T7OXovZBsDxbgbZFL"
+      "year" : "MGT7nlMo02w3VttzS",
+      "month" : "GmkKXI9Di3388W",
+      "day" : "ytkWH3ZoGw1mmQA"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/1y7E0ti0IgKRF9",
-        "name" : "qBFlUOgssiAFL",
+        "id" : "https://www.example.com/qRpEQ1eDDw1n",
+        "name" : "qZ6e3rZhQr",
         "nameType" : "Personal",
-        "orcId" : "cnTzNdbY6YB"
+        "orcId" : "FePe4bdRmW1c"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/cOzJI1Njwvzfz",
+        "id" : "https://www.example.com/SvaMTlmyH4U2lTXP",
         "labels" : {
-          "ru" : "rAWlaEtDZ7g"
+          "se" : "XJ9clR1MkgbDrZtIGJR"
         }
       } ],
-      "role" : "Dramaturge",
+      "role" : "Consultant",
       "sequence" : 6,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/P9zw8YbCVoTB",
-        "name" : "b4jA92oNUVLxU",
-        "nameType" : "Personal",
-        "orcId" : "eaYWQEf8CgZsp"
+        "id" : "https://www.example.com/YRoKhc4IqVXRZHV",
+        "name" : "bndiTbDmjGdAxE",
+        "nameType" : "Organizational",
+        "orcId" : "mRXL0rSHMt6gytMA"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/d4S3zThx54JYyO8hPb",
+        "id" : "https://www.example.com/6KZ59bEFOWZtIkVh1",
         "labels" : {
-          "el" : "gFWchnaPCrZbrqICZ"
+          "pt" : "NGyqqKVDn3FMjb5eDL"
         }
       } ],
-      "role" : "Soloist",
-      "sequence" : 5,
+      "role" : "Architect",
+      "sequence" : 6,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "1o7FBFF7SbCnyl69n",
-    "tags" : [ "6R1VDMrV4QncuUHq" ],
-    "description" : "pLn4YuRyCLZx",
+    "npiSubjectHeading" : "FZTv5N6veXa3djQh1",
+    "tags" : [ "UeXFN57GQQ" ],
+    "description" : "mjtaNOw70fxvfXy8f",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Report",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/1cAUfvMF1dRjIV"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/8SmQvZqrB19UppM"
         },
-        "seriesNumber" : "7mhZwoRTmlwiqmq",
+        "seriesNumber" : "se3npyGtaclmJ0bO",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/wFkD4GXXJ1"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/IDxQrFc7dqJEz1YWk6T"
         },
-        "isbnList" : [ "9781894670807" ]
+        "isbnList" : [ "9780859017343" ]
       },
-      "doi" : "https://www.example.com/dgMrwFss71TUGC",
+      "doi" : "https://www.example.com/DmOG9RnSETxEXkj9Drn",
       "publicationInstance" : {
         "type" : "ReportPolicy",
         "peerReviewed" : false,
@@ -98,32 +98,32 @@
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "85qOGnmgm4daVqVHZ",
-            "end" : "MP60Wt0BfpbBbGpM"
+            "begin" : "gPjoQej6Lk",
+            "end" : "Ogj0LhE4g1NxXVg"
           },
-          "pages" : "IncvJuxMJCIBTkprKvZ",
+          "pages" : "nQKI4gPFqI9Hlx",
           "illustrated" : true
         }
       }
     },
-    "metadataSource" : "https://www.example.com/TdM5mUma3o2jClQE",
-    "abstract" : "kjumAtxGy7JK"
+    "metadataSource" : "https://www.example.com/azfaszAFgs4U3",
+    "abstract" : "ukmyevyCXsDF"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/estsimilique",
-    "name" : "lzS2i48SDo5",
+    "id" : "https://www.example.org/commodiet",
+    "name" : "Fau9rS5gAitKZJQyC",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "lecCZ6UB6g",
-      "id" : "UZvOuroM4sU"
+      "source" : "gk7intFrc8HeK0Va",
+      "id" : "Pp7rDDK2pKxX"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "NARA",
       "approvalStatus" : "NOTAPPLIED",
-      "applicationCode" : "SgbPxnRhyYQu50Av"
+      "applicationCode" : "XM32BUypjGwkur"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -131,30 +131,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/utharum" ],
+  "subjects" : [ "https://www.example.org/eosat" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "44ae196a-5638-4f64-99fa-22064adeca78",
-    "name" : "2EhxbKH347P531I",
-    "mimeType" : "fpvf74pVSF5Ul7",
-    "size" : 690778695,
+    "identifier" : "8627bc4f-65a8-4b1e-991c-8064254f61d7",
+    "name" : "q8ppxGSHAtMK7SY",
+    "mimeType" : "Pvw21BZFSdS8r",
+    "size" : 1450762290,
     "license" : {
       "type" : "License",
-      "identifier" : "2xhAWqACIatX",
+      "identifier" : "n19k039Zzp5upuMVwC",
       "labels" : {
-        "es" : "632DO4QznH"
+        "nb" : "SRIai7VCHEYp9xMIcM"
       },
-      "link" : "https://www.example.com/LeCRlfbNynktLFc"
+      "link" : "https://www.example.com/DbqOJXMUZjdP"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/2onkNTTlq8xWcc",
-    "name" : "eLi3JE4ZZr",
-    "description" : "XUYjsjDTVX2VPQsO"
+    "id" : "https://www.example.com/IRC15dEDkLLpuD17aw",
+    "name" : "ua4FaWFrr2",
+    "description" : "zRiZs0AQru5rY"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
