+++ documentation/ReportResearch.json
@@ -3,94 +3,94 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "PUBLISHED",
   "resourceOwner" : {
-    "owner" : "plcI9y9FB1BXKaQ",
-    "ownerAffiliation" : "https://www.example.org/itaquenulla"
+    "owner" : "mDfYtcaVRO2dv",
+    "ownerAffiliation" : "https://www.example.org/quieum"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/maximeaccusantium",
+    "id" : "https://www.example.org/sintconsequatur",
     "labels" : {
-      "ru" : "G1TU4xEGEDRT"
+      "ca" : "3X3Cwi54cTds"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/occaecativoluptatem",
-  "doi" : "https://doi.org/10.1234/officiis",
-  "link" : "https://www.example.org/ipsamincidunt",
+  "handle" : "https://www.example.org/liberopraesentium",
+  "doi" : "https://doi.org/10.1234/magnam",
+  "link" : "https://www.example.org/molestiasnihil",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "djPbNP4CCTiBdLedMp",
+    "mainTitle" : "rW6kktK0LG5w1mgAhWW",
     "alternativeTitles" : {
-      "hu" : "GXaEIckvLtH"
+      "el" : "DoCXNBhk5vp5LZtFZKW"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "2annKdKoSdMQOtl1",
-      "month" : "VTVhE6S66KmHGg",
-      "day" : "ZRo9iqrgcNVd"
+      "year" : "PUJbVKrKdmOfXxH",
+      "month" : "AJoY8VaMcZ",
+      "day" : "JvXtf393DOM"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/SNpRIHXJHLiU4fUa",
-        "name" : "Og8uPksg3cSHt",
+        "id" : "https://www.example.com/AyH0owWZarzeJsGT7i",
+        "name" : "5FMhywcqY0ADbx",
         "nameType" : "Organizational",
-        "orcId" : "Ol7GPSOpqgW7"
+        "orcId" : "XgM8FhN1Ah"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/FqAdTYWnjYZdQX",
+        "id" : "https://www.example.com/r1iUqrN58QiNaB",
         "labels" : {
-          "nl" : "Sgk096sYfwj9"
+          "hu" : "HR2NQXFzCNO5Pv"
         }
       } ],
-      "role" : "VideoEditor",
-      "sequence" : 9,
+      "role" : "DataCollector",
+      "sequence" : 2,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/8D4jwFEaMuPb4",
-        "name" : "H5k7Bt0cAWkmfmV",
+        "id" : "https://www.example.com/zaZL4McEav",
+        "name" : "8TFySgmjyUVBloT",
         "nameType" : "Organizational",
-        "orcId" : "a0BZgUdfRL"
+        "orcId" : "YpsesJ0L0viFhh"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/hZTSFNb4a3cOwMrgeq",
+        "id" : "https://www.example.com/h4HQLY4rNw4j4",
         "labels" : {
-          "fi" : "2DYV1bctlZoRIJyF5l"
+          "ru" : "2oU32MUPeyks4tI9"
         }
       } ],
-      "role" : "ProjectManager",
-      "sequence" : 9,
+      "role" : "DataCurator",
+      "sequence" : 3,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "kkyGn1SN0Ui",
-    "tags" : [ "LKi4DnIK2ROx9B42i" ],
-    "description" : "JHmf3oa1duV1CINwrT",
+    "npiSubjectHeading" : "rcV1L3iBsqnXK5X2",
+    "tags" : [ "tMpEbSsTjOg0V4H7M" ],
+    "description" : "S4hV9ZkpYZuDrv7yvct",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Report",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/M01O9PWIDu5ift"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/orIc0OHk5c3Z8ixT1a"
         },
-        "seriesNumber" : "UWXd0PyDz8Qs",
+        "seriesNumber" : "ZqTM8Hi3tZ",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/dZgJaM88EvStJgU"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/JYoyzvwiwS455UDyAO"
         },
-        "isbnList" : [ "9790739891345" ]
+        "isbnList" : [ "9790308847506" ]
       },
-      "doi" : "https://www.example.com/5XU7RJYtstSskfY",
+      "doi" : "https://www.example.com/uQNRujcTWX",
       "publicationInstance" : {
         "type" : "ReportResearch",
         "peerReviewed" : false,
@@ -98,32 +98,32 @@
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "a6Xpg0cQ52ko",
-            "end" : "MfSuQWZbhow4ryvj0n"
+            "begin" : "uP08wcOuUk",
+            "end" : "moytoeAj37aREJ1f"
           },
-          "pages" : "FbpL2qJLIzS",
+          "pages" : "hxIYNgrwf9JxbE",
           "illustrated" : true
         }
       }
     },
-    "metadataSource" : "https://www.example.com/4gVH6Nh7eZ",
-    "abstract" : "0hOdL1bJAbEYafPQ"
+    "metadataSource" : "https://www.example.com/u5nspOIs8ChYfZm8",
+    "abstract" : "X13JpYrNKzVm3"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/ametfugiat",
-    "name" : "IHa8wPeLem46BpOXi",
+    "id" : "https://www.example.org/quiiste",
+    "name" : "utVQGHq63eQAwi9Rp",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "rr8pgYSH8hyQNBFg",
-      "id" : "zjhzKUxC4AR"
+      "source" : "bHfmnw8SXrNMF8DR4sp",
+      "id" : "TZ2uh2DCAqcIBE"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "NARA",
       "approvalStatus" : "REJECTION",
-      "applicationCode" : "KcWxdLOTUDHC"
+      "applicationCode" : "EvO8QoZFjSQ"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -131,30 +131,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/nesciuntillo" ],
+  "subjects" : [ "https://www.example.org/nullavoluptatem" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "bb89e3e8-1b97-4bb6-bd9e-efe9818358ab",
-    "name" : "AuE2QEPSBc6gp",
-    "mimeType" : "yRUvAAtiEMyiDh",
-    "size" : 443834960,
+    "identifier" : "9d739e67-f465-46ac-8478-61a31d7c5c5d",
+    "name" : "Gfolx0JbsYWv5unD",
+    "mimeType" : "fGZdYs8yDbzqCD",
+    "size" : 962992104,
     "license" : {
       "type" : "License",
-      "identifier" : "wvjXSiTI5swOnUqz",
+      "identifier" : "wa2hyBuLqwtHTv0r",
       "labels" : {
-        "de" : "1HGs9BDO6jCr"
+        "af" : "Qd7hTbHbV6y7z"
       },
-      "link" : "https://www.example.com/wAgmvdfK2kvzem9"
+      "link" : "https://www.example.com/I9AGuP9YR63k"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/kRVxvqoULb57E7U",
-    "name" : "a5DhKLGlyeRYtkB7NcI",
-    "description" : "eaCGJ3fewRJmA"
+    "id" : "https://www.example.com/eg6aRooFGuLk",
+    "name" : "I1CGCPi2hk9mjI",
+    "description" : "3G0zwK43DhGhkgBt2"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
