+++ documentation/ReportWorkingPaper.json
@@ -1,96 +1,96 @@
 {
   "type" : "Publication",
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
-  "status" : "PUBLISHED",
+  "status" : "DRAFT_FOR_DELETION",
   "resourceOwner" : {
-    "owner" : "AwT9nfktwWzY1t",
-    "ownerAffiliation" : "https://www.example.org/doloreset"
+    "owner" : "1gr215xtOjfwOZi2",
+    "ownerAffiliation" : "https://www.example.org/etquasi"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/asperioresin",
+    "id" : "https://www.example.org/rerumlaudantium",
     "labels" : {
-      "cs" : "RJb5WQsIY1Khdy"
+      "sv" : "YAxrjrro26ggY4ChH"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/doloremin",
-  "doi" : "https://doi.org/10.1234/illo",
-  "link" : "https://www.example.org/optioid",
+  "handle" : "https://www.example.org/minusplaceat",
+  "doi" : "https://doi.org/10.1234/nostrum",
+  "link" : "https://www.example.org/omniscupiditate",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "dPbDFdd7S7Vz",
+    "mainTitle" : "P5vz3C4ZHR4syeaZbX",
     "alternativeTitles" : {
-      "es" : "RW6GHjdzJNyd"
+      "fr" : "DZMlWVqKxp9eLkcCV"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "g1G6hQnvsRXSTeRv",
-      "month" : "CRm74GJ8OqsPHL02",
-      "day" : "B9XRRSiUTCO1WBs"
+      "year" : "w9xV4slSrCiA6",
+      "month" : "gRpU1klwuSwB9ylT4",
+      "day" : "4ba01XUxzJIsJiufx"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/YNu83FN0WQFy2cu1vvL",
-        "name" : "wjflwKYWJYXbR21HuU",
+        "id" : "https://www.example.com/XI0dKmkPJo",
+        "name" : "3YvKRESp1bb5c",
         "nameType" : "Personal",
-        "orcId" : "lDpZ6T9hukfb881Ou4"
+        "orcId" : "YTim1Vy3Qgf5xXF3FJ"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/BN56F27Be5el4bBd",
+        "id" : "https://www.example.com/BmEWP8kHpCQ2YU3",
         "labels" : {
-          "ca" : "VZMFqY81HPvvNfOIScG"
+          "ca" : "B85wiV0Ep75NpYzvfI"
         }
       } ],
-      "role" : "Supervisor",
-      "sequence" : 1,
+      "role" : "RegistrationAuthority",
+      "sequence" : 8,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/IQIgKeUyV8u1y",
-        "name" : "VbkjhpOdeUn",
-        "nameType" : "Organizational",
-        "orcId" : "MUENAFr9ur"
+        "id" : "https://www.example.com/Ore1BkdLs7",
+        "name" : "w6Qfao1rjcf",
+        "nameType" : "Personal",
+        "orcId" : "WULnFagkLfFej"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/0VPx9SYs5vEmcbE1Xr",
+        "id" : "https://www.example.com/8uptXmNwLYm5MUY",
         "labels" : {
-          "hu" : "fDRrGGuaGDIGbHD34k"
+          "el" : "l4Cq7YRotb94Tb8GZ"
         }
       } ],
-      "role" : "CostumeDesigner",
-      "sequence" : 4,
+      "role" : "Musician",
+      "sequence" : 7,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "TXomKqQIgt9cFree3a",
-    "tags" : [ "Gq35DBSI3WKT" ],
-    "description" : "CRHFtzbKXEvt",
+    "npiSubjectHeading" : "u4EKnBEHSht",
+    "tags" : [ "PiyNtzYlHhI" ],
+    "description" : "p4thfC2zBXeBQAZ680F",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Report",
         "series" : {
           "type" : "Series",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/CrESwrtvUBH"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/QdSp6dz6iOvHq89bR"
         },
-        "seriesNumber" : "70GPYmFfQDxs",
+        "seriesNumber" : "S4FK9Kp2LeHiQolJ",
         "publisher" : {
           "type" : "Publisher",
-          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/NoDEW1jNpoPRGprw1"
+          "id" : "https://api.dev.nva.aws.unit.no/publication-channels/5M8vTJV0zF"
         },
-        "isbnList" : [ "9791545035589" ]
+        "isbnList" : [ "9780347218191" ]
       },
-      "doi" : "https://www.example.com/DhsKc5dVxo7Q",
+      "doi" : "https://www.example.com/Jes033HfxIR0pVyhorY",
       "publicationInstance" : {
         "type" : "ReportWorkingPaper",
         "peerReviewed" : false,
@@ -98,32 +98,32 @@
           "type" : "MonographPages",
           "introduction" : {
             "type" : "Range",
-            "begin" : "IEK13DazBPhyroWX",
-            "end" : "Ixq7SZTuwiF"
+            "begin" : "tpDycH15wSoBUDH",
+            "end" : "1yOuD6x214"
           },
-          "pages" : "G2fzM9EUIs4x",
+          "pages" : "349gDQBwXC",
           "illustrated" : false
         }
       }
     },
-    "metadataSource" : "https://www.example.com/l9U2Aw6uIYpGYm",
-    "abstract" : "g0veNVBQgi8QP"
+    "metadataSource" : "https://www.example.com/mVr53Hss2zyJ8yKFR",
+    "abstract" : "Lzr5bmBpLRYkg"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/namquibusdam",
-    "name" : "4WVPOFsGfVk91OG9",
+    "id" : "https://www.example.org/utet",
+    "name" : "Y106fKiKwt",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "vL7wZybktvwwCxxxe",
-      "id" : "DjSfiAK9TQ"
+      "source" : "Vj7dV26HAAq",
+      "id" : "ASkG3WC02FwO"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "NMA",
-      "approvalStatus" : "APPLIED",
-      "applicationCode" : "7q0agLOKAFv71"
+      "approvalStatus" : "APPROVED",
+      "applicationCode" : "XjcX7Ek5UGLV8PthY"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -131,30 +131,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/autimpedit" ],
+  "subjects" : [ "https://www.example.org/quisearum" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "7a9c3eb3-d6c3-47db-9474-e92ca7df6812",
-    "name" : "SnDoijVnSOqo",
-    "mimeType" : "DRYRWcpN20w0JBfdJ",
-    "size" : 1740927162,
+    "identifier" : "5e127c33-85b0-4fa5-90ac-8868bc92813b",
+    "name" : "0Qb0ilIdEpM",
+    "mimeType" : "7qY0bA7F1xBrGB",
+    "size" : 1301423986,
     "license" : {
       "type" : "License",
-      "identifier" : "kIdG8moO75",
+      "identifier" : "pKNkM947qXyl",
       "labels" : {
-        "el" : "Thsgvn0Go2sc"
+        "cs" : "de93SR4CiT"
       },
-      "link" : "https://www.example.com/g8Sk9EQwvTF"
+      "link" : "https://www.example.com/p1vAws61mdmOFA"
     },
     "administrativeAgreement" : false,
     "publisherAuthority" : true,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/TivdjgytjjSxZq",
-    "name" : "9E5kM8l21Vmgkdofr",
-    "description" : "Xg6IYxkLlxdfnfVM"
+    "id" : "https://www.example.com/l0p71Ge1wDyFJxaQQD",
+    "name" : "iEL8Mq8JHartKy",
+    "description" : "MMub0b2Sr3alfxt"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
