+++ documentation/VisualArts.json
@@ -3,103 +3,103 @@
   "identifier" : "c443030e-9d56-43d8-afd1-8c89105af555",
   "status" : "DRAFT",
   "resourceOwner" : {
-    "owner" : "wU5G1N8S7VI",
-    "ownerAffiliation" : "https://www.example.org/solutamolestias"
+    "owner" : "uX7WP6wgFS4r4se7SJV",
+    "ownerAffiliation" : "https://www.example.org/oditmollitia"
   },
   "publisher" : {
     "type" : "Organization",
-    "id" : "https://www.example.org/ipsaharum",
+    "id" : "https://www.example.org/rerumsunt",
     "labels" : {
-      "sv" : "OTDxoEq5Thk"
+      "bg" : "NCOdWiWsoNTOOj3"
     }
   },
   "createdDate" : "2020-09-23T09:51:23.044996ZZ",
   "modifiedDate" : "2020-09-23T09:51:23.044996ZZ",
   "publishedDate" : "2020-09-23T09:51:23.044996ZZ",
   "indexedDate" : "2020-09-23T09:51:23.044996ZZ",
-  "handle" : "https://www.example.org/temporererum",
-  "doi" : "https://doi.org/10.1234/et",
-  "link" : "https://www.example.org/fugarerum",
+  "handle" : "https://www.example.org/estesse",
+  "doi" : "https://doi.org/10.1234/quos",
+  "link" : "https://www.example.org/corporisvoluptates",
   "entityDescription" : {
     "type" : "EntityDescription",
-    "mainTitle" : "KlANdHaUdeG",
+    "mainTitle" : "UWcRRnXRtUcM0N",
     "alternativeTitles" : {
-      "ru" : "xoPA5NkbRdEJbbA"
+      "pt" : "QMqGU8f0uYgAN"
     },
     "language" : "http://lexvo.org/id/iso639-3/und",
     "date" : {
       "type" : "PublicationDate",
-      "year" : "xFXMigQJp0",
-      "month" : "gLAvk9mf1Fb",
-      "day" : "mhsheppQ7hbGaq0I"
+      "year" : "NsB4xP6CAZsfdN",
+      "month" : "RIHqegLo7xpjebTAn1",
+      "day" : "qeeAyECGgWL7bFnvypd"
     },
     "contributors" : [ {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/U9BrWR0AKJx",
-        "name" : "01wl7OrOnan31UhE",
-        "nameType" : "Organizational",
-        "orcId" : "UQrRvDGabBsoh2rBs"
+        "id" : "https://www.example.com/a0H7kLJt7uHxYiMzn",
+        "name" : "fIqXjNu3sKYCV",
+        "nameType" : "Personal",
+        "orcId" : "a3x91UXo6WXV5OV"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/QRLkStoIhjH",
+        "id" : "https://www.example.com/IYg8iUddVt",
         "labels" : {
-          "es" : "NSVX1U38YeM3o74sZg"
+          "cs" : "b1I5rD8PzUxmZbpp"
         }
       } ],
-      "role" : "Other",
-      "sequence" : 8,
+      "role" : "Screenwriter",
+      "sequence" : 1,
       "correspondingAuthor" : false
     }, {
       "type" : "Contributor",
       "identity" : {
         "type" : "Identity",
-        "id" : "https://www.example.com/Tz7wXTD7xlMJEio",
-        "name" : "FPhsobeIKmju",
-        "nameType" : "Personal",
-        "orcId" : "iSapUuu2lxUTOKOY"
+        "id" : "https://www.example.com/jg2HSS2HZod",
+        "name" : "YKgHgDiAccSP8IW6Qph",
+        "nameType" : "Organizational",
+        "orcId" : "pGraxWdSEOuwrc"
       },
       "affiliations" : [ {
         "type" : "Organization",
-        "id" : "https://www.example.com/kgWyQOwFcZ37r",
+        "id" : "https://www.example.com/sQJOOaw8TMwDakNhH",
         "labels" : {
-          "el" : "ZqIpTYeYY3B"
+          "nn" : "376i51bpBhEwlvc0u3G"
         }
       } ],
-      "role" : "Dramatist",
-      "sequence" : 1,
+      "role" : "HostingInstitution",
+      "sequence" : 5,
       "correspondingAuthor" : false
     } ],
-    "npiSubjectHeading" : "5fvO0bwDFDUclTNPw3",
-    "tags" : [ "o1bM7GJvTeUFl" ],
-    "description" : "zyDUhVUyLolTn0O",
+    "npiSubjectHeading" : "xD7qlcbJa1IcXy3",
+    "tags" : [ "MMBGyFjz6QvFWptwg" ],
+    "description" : "syDwH7z8Qc",
     "reference" : {
       "type" : "Reference",
       "publicationContext" : {
         "type" : "Artistic"
       },
-      "doi" : "https://www.example.com/NUI9fHqKQvX",
+      "doi" : "https://www.example.com/KlPYDBmREy",
       "publicationInstance" : {
         "type" : "VisualArts",
         "subtype" : {
-          "type" : "Performance"
+          "type" : "ArtistBook"
         },
-        "description" : "SHmiMeFSR5lUHIE",
+        "description" : "HVgSGWBfBtFh8CjJ",
         "venues" : [ {
           "type" : "Venue",
           "place" : {
             "type" : "UnconfirmedPlace",
-            "label" : "CwVjuae8KBhTB",
-            "country" : "BUQuXRpfTUXtuPYB"
+            "label" : "xGwNNXXfU1WeAgb",
+            "country" : "bw5F6JmJ4uS"
           },
           "date" : {
             "type" : "Period",
             "from" : "2020-09-23T09:51:23.044996Z",
             "to" : "2020-09-23T09:51:23.044996Z"
           },
-          "sequence" : 383224988
+          "sequence" : 1585536817
         } ],
         "peerReviewed" : false,
         "pages" : {
@@ -107,24 +107,24 @@
         }
       }
     },
-    "metadataSource" : "https://www.example.com/P4n6Og9U7r6BGvdc6ul",
-    "abstract" : "K3EWv6shE8WV"
+    "metadataSource" : "https://www.example.com/G7JJdIiddhY0XCj",
+    "abstract" : "QD6iriF5hFbX6v"
   },
   "projects" : [ {
     "type" : "ResearchProject",
-    "id" : "https://www.example.org/atqueamet",
-    "name" : "YcrbYje9v4Y",
+    "id" : "https://www.example.org/nisisunt",
+    "name" : "NJmNLcb457KGaQRr",
     "grants" : [ {
       "type" : "Grant",
-      "source" : "VToRGV3CB5UbSa",
-      "id" : "k7XgzhF6XQiHt"
+      "source" : "QZMdLy9qoCbSsA",
+      "id" : "WR8FJX67PaHN3mBUih"
     } ],
     "approvals" : [ {
       "type" : "Approval",
       "date" : "2020-09-23T09:51:23.044996ZZ",
       "approvedBy" : "REK",
-      "approvalStatus" : "APPROVED",
-      "applicationCode" : "3u33oTypW8xr74qd"
+      "approvalStatus" : "DECLINED",
+      "applicationCode" : "J6POxy2NbfnQspihpJ"
     } ]
   } ],
   "additionalIdentifiers" : [ {
@@ -132,30 +132,29 @@
     "source" : "fakesource",
     "value" : "1234"
   } ],
-  "subjects" : [ "https://www.example.org/saepeest" ],
+  "subjects" : [ "https://www.example.org/etid" ],
   "associatedArtifacts" : [ {
     "type" : "PublishedFile",
-    "identifier" : "5629110f-62b8-4922-b122-ddb3ef87dd17",
-    "name" : "u6OXqU6FZa5Sk9l1",
-    "mimeType" : "ZZlE7frvUi",
-    "size" : 1231890018,
+    "identifier" : "40ede384-3180-43a1-903c-0ac97f15e2e9",
+    "name" : "GeQrknyeTPL3iZ",
+    "mimeType" : "9OP3oMZ5QQZ6jXC",
+    "size" : 1448263130,
     "license" : {
       "type" : "License",
-      "identifier" : "91fqsfBrDNTP",
+      "identifier" : "OY1NxylbOTSgHdVv9A",
       "labels" : {
-        "nl" : "cCtYDDakmc4zH4xS2"
+        "it" : "K1ZyqLFytsnZK"
       },
-      "link" : "https://www.example.com/q85SzWIKwEIwbYaEewQ"
+      "link" : "https://www.example.com/VTltpfrdo23npju"
     },
     "administrativeAgreement" : false,
-    "publisherAuthority" : false,
-    "embargoDate" : "2020-09-23T09:51:23.044996ZZ",
+    "publisherAuthority" : true,
     "visibleForNonOwner" : true
   }, {
     "type" : "AssociatedLink",
-    "id" : "https://www.example.com/rZgC6SIvfSXxX",
-    "name" : "jVRZqeCUsypXc2guO2",
-    "description" : "NMmSWOohNjgx7XjYyG"
+    "id" : "https://www.example.com/cTtshnRU8AdJdOzfe",
+    "name" : "H75U9W34ISbnA",
+    "description" : "VF5vw49Yi5l1oVDdP0"
   } ],
-  "modelVersion" : "0.19.9"
+  "modelVersion" : "0.19.14"
 }
