+++ documentation/schema.yaml
@@ -56,6 +56,8 @@ schema:
     associatedArtifacts:
       type: array
       properties:
+        publishable:
+          type: boolean
         empty:
           type: boolean
       items:
@@ -244,6 +246,8 @@ referencedSchemas:
   AssociatedArtifactList:
     type: array
     properties:
+      publishable:
+        type: boolean
       empty:
         type: boolean
     items:
@@ -2007,6 +2011,8 @@ referencedSchemas:
       associatedArtifacts:
         type: array
         properties:
+          publishable:
+            type: boolean
           empty:
             type: boolean
         items:
