+++ nva-datamodel-testutils/src/main/java/no/unit/nva/model/testing/associatedartifacts/AssociatedLinkGenerator.java
@@ -1,11 +1,13 @@
-package no.unit.nva.model.associatedartifacts;
+package no.unit.nva.model.testing.associatedartifacts;
+
+import no.unit.nva.model.associatedartifacts.AssociatedLink;
 
 import static no.unit.nva.model.testing.PublicationGenerator.randomUri;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 
-public class AssociatedLinkGenerator {
+public final class AssociatedLinkGenerator {
     private AssociatedLinkGenerator() {
-        // NO-OP
+        super();
     }
