+++ nva-datamodel-testutils/src/main/java/no/unit/nva/model/testing/associatedartifacts/PublishedFileGenerator.java
@@ -1,4 +1,4 @@
-package no.unit.nva.model.associatedartifacts;
+package no.unit.nva.model.testing.associatedartifacts;
 
 import no.unit.nva.model.associatedartifacts.file.PublishedFile;
 
@@ -7,13 +7,13 @@ import java.util.UUID;
 import static no.unit.nva.testutils.RandomDataGenerator.randomInteger;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 
-public final class PublishedFileGenerator extends AssociatedFileGenerator {
+public final class PublishedFileGenerator {
     private PublishedFileGenerator() {
         // NO-OP
     }
 
     public static PublishedFile random() {
         return new PublishedFile(UUID.randomUUID(), randomString(), randomString(), randomInteger().longValue(),
-                randomLicense(), false, true, null);
+                LicenseGenerator.random(), false, true, null);
     }
 }
