+++ nva-datamodel-testutils/src/main/java/no/unit/nva/model/testing/associatedartifacts/UnpublishedFileGenerator.java
@@ -1,4 +1,4 @@
-package no.unit.nva.model.associatedartifacts;
+package no.unit.nva.model.testing.associatedartifacts;
 
 import no.unit.nva.model.associatedartifacts.file.UnpublishedFile;
 
@@ -7,14 +7,14 @@ import java.util.UUID;
 import static no.unit.nva.testutils.RandomDataGenerator.randomInteger;
 import static no.unit.nva.testutils.RandomDataGenerator.randomString;
 
-public final class UnpublishedFileGenerator extends AssociatedFileGenerator {
+public final class UnpublishedFileGenerator {
 
     private UnpublishedFileGenerator() {
-        //NO-OP
+        // NO-OP
     }
 
     public static UnpublishedFile random() {
         return new UnpublishedFile(UUID.randomUUID(), randomString(), randomString(), randomInteger().longValue(),
-        randomLicense(), false, true, null);
+        LicenseGenerator.random(), false, true, null);
     }
 }
